refactor: 💡 refactor after pr review and discussion
