--- src/components/iconSet/ts/icons/IconEnvelope.tsx
@@ -1,21 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconEnvelope = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			width="72"
-			height="72"
-			viewBox="0 0 72 72"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<path
-				d="M65.9963 28V55.4658C65.9963 56.9883 65.4719 58.2913 64.4232 59.3748C63.3745 60.4583 62.1134 61 60.6397 61H11.3565C9.88288 61 8.62171 60.4583 7.57303 59.3748C6.52434 58.2913 6 56.9883 6 55.4658V28C6.98244 29.13 8.10987 30.1334 9.38229 31.0102C17.4618 36.6833 23.0083 40.6614 26.0219 42.9445C27.2943 43.9131 28.3267 44.6685 29.1192 45.2109C29.9116 45.7533 30.9666 46.3066 32.284 46.871C33.6014 47.4353 34.8288 47.7181 35.9663 47.7194H36.0337C37.1724 47.7194 38.3999 47.4366 39.716 46.871C41.0322 46.3054 42.0871 45.752 42.8808 45.2109C43.6745 44.6698 44.707 43.9144 45.9781 42.9445C49.7729 40.1087 55.3307 36.1305 62.6515 31.0102C63.9239 30.1114 65.0401 29.108 66 28H65.9963ZM66 18.36C66 20.1235 65.4532 21.8083 64.3595 23.4143C63.2658 25.0204 61.904 26.3927 60.2741 27.5313C51.8821 33.3568 46.6593 36.9839 44.6057 38.4124C44.382 38.5687 43.9076 38.9093 43.1827 39.4342C42.4577 39.9591 41.8553 40.3835 41.3753 40.7072C40.8953 41.0309 40.3147 41.3933 39.6335 41.7945C38.9523 42.1957 38.3105 42.4969 37.708 42.6982C37.1056 42.8994 36.5475 43 36.0337 43H35.9663C35.4525 43 34.8944 42.8994 34.292 42.6982C33.6895 42.4969 33.0477 42.1957 32.3665 41.7945C31.6853 41.3933 31.1047 41.0309 30.6247 40.7072C30.1447 40.3835 29.5423 39.9591 28.8173 39.4342C28.0924 38.9093 27.618 38.5687 27.3943 38.4124C25.3632 36.9839 22.439 34.9473 18.6217 32.3026C14.8045 29.6579 12.5165 28.0675 11.7578 27.5313C10.3741 26.5939 9.06856 25.3047 7.84114 23.6637C6.61371 22.0226 6 20.4991 6 19.093C6 17.352 6.4631 15.9015 7.38929 14.7417C8.31548 13.5818 9.6379 13.0012 11.3565 13H60.6397C62.0909 13 63.3464 13.5243 64.4063 14.5729C65.4663 15.6215 65.9963 16.8826 65.9963 18.3562L66 18.36Z"
-				fill="#3F373F"
-			/>
-		</svg>
-	);
-};
