--- src/components/iconSet/ts/icons/IconPin.tsx
@@ -1,24 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconPin = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			xmlns="http://www.w3.org/2000/svg"
-			xmlnsXlink="http://www.w3.org/1999/xlink"
-			width="72"
-			height="72"
-			viewBox="0 0 72 72"
-		>
-			<defs>
-				<path
-					id="pin-a"
-					d="M46.6382979,25.4366197 C46.6382979,22.6368502 45.6410767,20.246767 43.6466344,18.26637 C41.6521921,16.2859731 39.245158,15.2957746 36.4255319,15.2957746 C33.6059059,15.2957746 31.1988717,16.2859731 29.2044294,18.26637 C27.2099871,20.246767 26.212766,22.6368502 26.212766,25.4366197 C26.212766,28.2363892 27.2099871,30.6264725 29.2044294,32.6068694 C31.1988717,34.5872663 33.6059059,35.5774648 36.4255319,35.5774648 C39.245158,35.5774648 41.6521921,34.5872663 43.6466344,32.6068694 C45.6410767,30.6264725 46.6382979,28.2363892 46.6382979,25.4366197 Z M55.9956251,26.0007276 C55.9956251,28.838703 55.5661526,31.169168 54.7072077,32.9921224 L40.4886799,63.2261861 C40.071603,64.0851623 39.4532794,64.7625721 38.6337088,65.2584157 C37.8141383,65.7542593 36.9355062,66.0014519 35.9978125,65.9999936 C35.0601189,65.9985352 34.1814867,65.7513426 33.3619162,65.2584157 C32.5423457,64.7654888 31.9371468,64.088079 31.5463196,63.2261861 L17.2884174,32.9921224 C16.4294725,31.169168 16,28.838703 16,26.0007276 C16,20.4793632 17.953407,15.7659323 21.8602209,11.8604347 C25.7670349,7.95493714 30.4802946,6.00145918 36,6.00000082 C41.5197054,5.99854245 46.2329651,7.95202041 50.1397791,11.8604347 C54.046593,15.768849 56,20.48228 56,26.0007276 L55.9956251,26.0007276 Z"
-				/>
-			</defs>
-			<use xlinkHref="#pin-a" />
-		</svg>
-	);
-};
