--- src/components/iconSet/ts/index.ts
@@ -1,11 +1,0 @@
-export * from './icons/IconEnvelope';
-export * from './icons/IconLock';
-export * from './icons/IconPerson';
-export * from './icons/IconPin';
-
-export * from './logos/IconCaritasWhite';
-export * from './logos/IconInVia';
-export * from './logos/IconKreuzbund';
-export * from './logos/IconRaphaelswerk';
-export * from './logos/IconSKF';
-export * from './logos/IconSKM';
