--- src/components/iconSet/ts/logos/IconCaritasWhite.tsx
@@ -1,166 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconCaritasWhite = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			width="60"
-			height="60"
-			viewBox="0 0 60 60"
-			fill="none"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<path
-				d="M52.293 56.3469V3.64317C52.293 2.02574 50.923 0.715149 49.2335 0.715149H10.7662C9.07644 0.715149 7.70679 2.02574 7.70679 3.64317V56.3469C7.70679 57.9638 9.07644 59.2748 10.7662 59.2748H49.2335C50.923 59.2748 52.293 57.964 52.293 56.3469Z"
-				fill="#CC1E1C"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M16.8363 52.1455C16.5304 51.0363 15.7585 50.3737 14.5939 50.3737C12.803 50.3737 12.0603 51.9867 12.0603 53.5427C12.0603 55.1133 12.7886 56.6828 14.5939 56.6828C15.8897 56.6828 16.705 55.9192 16.9089 54.6807H17.8408C17.5349 56.337 16.4867 57.4608 14.5939 57.4608C12.1914 57.4608 11.1285 55.7469 11.1285 53.5427C11.1285 51.4687 12.337 49.5955 14.5939 49.5955C16.3266 49.5955 17.4623 50.417 17.7244 52.1455H16.8363Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M23.4464 53.3411C22.7035 53.6438 21.9467 53.7015 21.1894 53.7587C19.8641 53.8596 19.0635 54.3355 19.0635 55.3437C19.0635 56.2509 19.8641 56.6828 20.6945 56.6828C22.0774 56.6828 23.4464 56.0205 23.4464 54.5079V53.3411ZM21.9756 52.9089C22.296 52.8801 22.8347 52.7937 23.1257 52.6639C23.4171 52.5341 23.4464 52.1455 23.4464 51.886C23.4464 51.0076 22.9223 50.3737 21.568 50.3737C20.3594 50.3737 19.5295 50.7625 19.3693 52.0591H18.4813C18.6125 50.302 19.937 49.5955 21.5827 49.5955C23.1113 49.5955 24.3198 50.1864 24.3198 51.9149V55.9771C24.3198 56.6544 24.4946 56.8127 25.2371 56.5531V57.2305C25.106 57.2739 24.8002 57.3747 24.5529 57.3747C24.4071 57.3747 24.2908 57.3601 24.1595 57.3311C23.5772 57.245 23.4752 56.7404 23.4608 56.2225C22.7329 57.0286 21.7428 57.4609 20.6506 57.4609C19.3259 57.4609 18.1317 56.7551 18.1317 55.3144C18.1317 54.061 19.0053 53.2979 20.7672 53.0674L21.9756 52.9089Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M26.6709 57.2596H25.7975V49.7976H26.6709V51.195H26.7001C27.1804 49.9994 27.8939 49.5955 29.1606 49.5955V50.5326C27.1953 50.4887 26.6709 51.9296 26.6709 53.6004V57.2596Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M29.834 57.2596H30.7078V49.7976H29.834V57.2596Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M34.8532 57.2878C34.6494 57.3311 34.2705 57.4609 33.8483 57.4609C32.9895 57.4609 32.5082 57.0865 32.5082 55.9771V50.5182H31.4455V49.7976H32.5082V47.723H33.382V49.7976H34.7946V50.5182H33.382V55.6317C33.382 56.3664 33.4113 56.6828 34.1393 56.6828C34.373 56.6828 34.6196 56.6398 34.8532 56.5822V57.2878Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M40.1689 53.3411C39.426 53.6438 38.6686 53.7015 37.9119 53.7587C36.5867 53.8596 35.7856 54.3355 35.7856 55.3437C35.7856 56.2509 36.5867 56.6828 37.417 56.6828C38.8 56.6828 40.1689 56.0205 40.1689 54.5079V53.3411ZM38.6981 52.9091C39.0187 52.8802 39.5573 52.7937 39.8486 52.6639C40.1396 52.5341 40.1689 52.1455 40.1689 51.886C40.1689 51.0077 39.6443 50.3737 38.2898 50.3737C37.0819 50.3737 36.252 50.7625 36.0919 52.0591H35.2033C35.335 50.302 36.6597 49.5955 38.3043 49.5955C39.8338 49.5955 41.0422 50.1864 41.0422 51.9149V55.9771C41.0422 56.6544 41.2172 56.8127 41.9596 56.5533V57.2305C41.8287 57.2739 41.5229 57.3747 41.2759 57.3747C41.1291 57.3747 41.0133 57.3601 40.8821 57.3311C40.2992 57.245 40.1978 56.7404 40.1833 56.2225C39.4549 57.0289 38.4657 57.4609 37.3728 57.4609C36.0481 57.4609 34.8542 56.7552 34.8542 55.3144C34.8542 54.061 35.7284 53.2979 37.4891 53.0674L38.6981 52.9091Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M42.8174 54.9255C43.0943 56.2079 43.7643 56.6828 45.0745 56.6828C46.3416 56.6828 46.9377 56.0205 46.9377 55.3002C46.9377 53.1397 42.1043 54.6669 42.1043 51.7283C42.1043 50.4889 43.0794 49.5957 44.7537 49.5957C46.3848 49.5957 47.4479 50.2582 47.6663 51.9442H46.778C46.618 50.7914 45.8318 50.3737 44.7393 50.3737C43.633 50.3737 43.0357 50.8919 43.0357 51.6561C43.0357 53.6871 47.8697 52.4196 47.8697 55.2858C47.8697 56.712 46.5743 57.4609 44.7979 57.4609C43.0218 57.4609 42.0025 56.337 41.9293 54.9255H42.8174Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M50.0154 20.0608H33.6822V3.63367H26.5762V20.0608H10.2425V27.207H26.5762V43.6346H33.6822V27.207H50.0154V20.0608Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M23.7587 19.1935H25.7179V16.8122C25.2007 16.135 24.6469 15.4464 24.0694 14.8196C22.8073 13.4208 21.1587 12.5398 19.6137 11.5036C18.5305 10.6879 17.5135 9.44111 16.7441 8.38745C16.008 9.04175 15.3194 9.749 14.6846 10.5036C14.8161 10.7302 15.0387 11.0893 15.415 11.6592C16.4346 13.203 17.965 14.0425 19.4848 15.0009C19.9226 15.234 20.3347 15.5968 20.7467 15.9076C21.9255 16.846 22.899 18.0841 23.7587 19.1935Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M22.1377 19.1935C21.7018 18.4049 20.9499 17.5039 20.3604 16.8662C18.5571 15.234 16.0846 14.6902 14.7454 12.4881C14.3691 11.8697 14.1524 11.5169 14.0277 11.3264C13.4123 12.1395 12.8556 13 12.366 13.9022C13.289 15.4614 14.4785 16.8517 16.1362 17.695C17.3984 18.3373 18.0675 18.9322 18.3267 19.1935H22.1377Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M43.2838 8.18558C42.6517 8.64494 42.0438 9.0465 41.482 9.58659C40.1424 10.7008 39.3679 12.2632 38.3648 13.6796C37.405 15.0364 35.9256 16.1719 34.5074 17.0232V19.1935H37.0149C38.1688 18.4079 39.4317 17.2781 40.297 16.2703C41.7134 14.3016 42.769 11.8924 44.8562 10.623C45.0684 10.4938 45.2444 10.3838 45.3918 10.2893C44.7412 9.53734 44.0364 8.83436 43.2838 8.18558Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M48.0498 33.0315C47.1482 31.8391 46.247 30.7255 44.9591 29.9225C44.0495 29.3928 43.0525 28.7018 42.1509 28.066H38.2744L38.6228 28.6271C39.8846 30.3623 41.7394 31.3471 43.4137 32.5384C44.6119 33.3337 45.6586 34.5186 46.4085 35.7047C47.0229 34.8685 47.5746 33.983 48.0572 33.0557L48.0498 33.0315Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M45.3198 36.0362C44.0505 33.998 41.7394 32.8239 39.9101 31.5024C38.5947 30.5542 37.6632 29.3354 36.8134 28.066H34.5074V30.1127C35.5155 31.5994 37.0205 33.359 38.4681 34.2224C40.3644 35.3825 42.2245 36.9273 43.7144 38.7036C44.4206 38.059 45.0801 37.3651 45.6903 36.6277C45.5906 36.4653 45.4669 36.2724 45.3198 36.0362Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M23.2462 28.066C21.6052 29.2323 20.1595 30.6653 19.0468 32.3312C17.965 34.2224 16.7542 35.7769 14.977 37.0461L14.9558 37.0821C15.5995 37.8162 16.2949 38.5028 17.0367 39.1371C17.1975 39.0308 17.3846 38.8975 17.6043 38.7298C19.2893 37.4461 20.618 35.9065 21.6227 34.197C22.6388 32.6009 24.18 31.1987 25.7179 30.0795V28.066H23.2462Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M16.9896 28.066C15.9761 29.0502 15.3812 30.0039 14.6938 31.1396C14.185 32.0202 13.328 32.8957 12.4345 33.4911C12.931 34.3906 13.4928 35.2478 14.1142 36.0567L15.1829 35.3108C15.6209 34.9998 16.0846 34.6113 16.3936 34.197C17.218 33.3156 17.8103 32.15 18.5316 31.1917C19.3301 30.0777 20.2831 29.197 21.2875 28.3682C21.3522 28.3155 21.5152 28.1755 21.6422 28.066H16.9896Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M25.7179 31.5534C24.2873 32.5778 22.9008 33.9483 22.0605 35.4405C21.2102 37.0203 19.9226 38.3933 18.3514 39.4295C18.1284 39.5765 17.9576 39.6901 17.8263 39.7788C18.6567 40.4199 19.5374 40.9984 20.4622 41.5055C20.6633 41.3541 20.9375 41.1494 21.3136 40.8806C23.2881 39.4652 24.3417 37.3372 25.7179 35.4074V31.5534Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M38.7771 35.9322C37.2067 35.08 35.5836 33.4338 34.5074 31.9782V36.2586C34.6285 36.4007 34.8244 36.6382 35.1456 37.0203C36.4558 38.5816 38.4936 39.1448 39.8072 40.6475C40.0219 40.8924 40.1853 41.0759 40.3099 41.2126C41.1631 40.7125 41.9759 40.1505 42.7424 39.5339C41.7394 37.9933 40.3392 36.7822 38.7771 35.9322Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M39.8751 5.80878C39.4842 6.04668 38.8088 6.43683 37.7215 7.46253C36.5387 8.57705 35.6888 10.6009 34.5074 11.8647V15.474C36.0085 14.4203 37.5356 12.9555 38.4681 11.426C39.3697 10.0012 40.5032 8.65398 41.9452 7.85125C42.178 7.72165 42.3515 7.61648 42.4814 7.52938C41.6603 6.89144 40.7894 6.31569 39.8751 5.80878Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M25.7179 10.8123C25.5297 10.5812 25.2296 10.2247 24.9969 10.0012C23.5674 8.6299 21.3839 7.61095 20.0108 6.01938C19.1492 6.52125 18.3281 7.08502 17.5536 7.7048C18.5837 9.10353 20.0773 10.6749 21.5966 11.4521C23.166 12.3644 24.5246 13.7221 25.7179 14.9903V10.8123Z"
-				fill="#FFFFFE"
-			/>
-			<path
-				fillRule="evenodd"
-				clipRule="evenodd"
-				d="M46.1483 11.2185C45.9915 11.3046 45.8081 11.4136 45.5898 11.5516C43.7871 12.6914 42.7046 14.6602 41.5462 16.3182C40.8732 17.333 39.8427 18.3362 38.882 19.1935H43.0539C44.0402 18.3095 44.9311 17.0365 45.4608 16.1886C46.0277 15.2823 46.9876 14.4421 47.7603 13.6651C47.2808 12.8093 46.7419 11.9918 46.1483 11.2185Z"
-				fill="#FFFFFE"
-			/>
-			<mask
-				id="mask0"
-				mask-type="alpha"
-				maskUnits="userSpaceOnUse"
-				x="7"
-				y="0"
-				width="46"
-				height="60"
-			>
-				<path d="M7 0H53V59.9593H7V0Z" fill="white" />
-			</mask>
-			<g mask="url(#mask0)">
-				<path
-					fillRule="evenodd"
-					clipRule="evenodd"
-					d="M29.835 48.2351H30.8162V46.9305H29.835V48.2351Z"
-					fill="#FFFFFE"
-				/>
-				<path
-					fillRule="evenodd"
-					clipRule="evenodd"
-					d="M49.0207 0H10.9788C8.78489 0 7 1.72764 7 3.85112V56.1139C7 58.2375 8.78489 59.965 10.9788 59.965H49.0207C51.2149 59.965 53 58.2376 53 56.1139V3.85132C53 1.72764 51.2149 0 49.0207 0ZM49.0207 0.947639C50.6917 0.947639 52.0467 2.24729 52.0467 3.85132V56.1139C52.0467 57.7178 50.6917 59.0175 49.0207 59.0175H10.9788C9.30782 59.0175 7.95328 57.7175 7.95328 56.1139V3.85112C7.95328 2.24729 9.30782 0.947639 10.9788 0.947639H49.0207Z"
-					fill="#FFFFFE"
-				/>
-			</g>
-		</svg>
-	);
-};
