--- src/components/iconSet/ts/logos/IconInVia.tsx
@@ -1,46 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconInVia = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			viewBox="0 0 54 60"
-			xmlns="http://www.w3.org/2000/svg"
-			xmlnsXlink="http://www.w3.org/1999/xlink"
-		>
-			<defs>
-				<path id="a-2" d="M.045.178h14.353v15.409H.045z" />
-			</defs>
-			<g fill="none" fillRule="evenodd">
-				<path d="M-4 0h60v60H-4z" />
-				<g fill="#fff">
-					<path d="M6.872 57.967c.116-.36.105-.866.105-1.32.088-.044.011-.098.022-.163.055-.165-.022-.449.22-.47.24-.536.065-1.147.131-1.76-.164-.404.077-.82 0-1.224l.044-.109c-.021-.6.14-1.037.163-1.613v-.001l.001-.069c.143-.218.143-.546.23-.808 0-.267.096-.51.16-.75.034-.125.06-.249.06-.375.208-.558.154-1.236.307-1.815-.022-.458.175-.961.033-1.442-.241-.23-.625-.295-.9-.448-.23.142-.547-.033-.679.295-.187.077-.252.24-.428.339.055.164.022.371.033.546-.263.667-.132 1.388-.318 2.076A48.773 48.773 0 006 51.306v.001c.001.816.018 1.639.034 2.51.099.71.12 1.65.176 2.382.079.588.109 1.176.208 1.769.025.146.053.292.088.44.098.108.022.316.153.392v-.11c.022-.087.11-.119.187-.13 0 0 .024-.589.026-.592v-.001M15.278 49.071c-.146.498-.176 1.037-.289 1.548a1.088 1.088 0 00-.067.148h-.002c-.19.49-.091 1.088-.272 1.574-.06.045-.095.205-.225.171-.259-.388-.507-.844-.707-1.266-.07-.156-.138-.315-.206-.475-.247-.58-.488-1.175-.777-1.7l-.066-.119c.012-.433-.33-.661-.366-1.072-.212-.216-.224-.628-.448-.821a2.976 2.976 0 00-1.356-1.37c-.306.08-.46-.25-.742-.136-.118.194-.449.057-.531.33-.024.389-.177.73-.2 1.118-.107.206-.048.48-.142.685l-.059.856-.059.056c-.011.157-.024.313-.038.47-.049.566-.108 1.133-.156 1.701a22.663 22.663 0 00-.089 1.686c-.13.08.024.182-.023.297-.13.57 0 1.13 0 1.643.047.695.212 1.3.59 1.813.33.08.435.4.789.331.472-.148.13-.73.236-1.072-.177-.377-.271-.89-.142-1.289.012-.411.083-.913.048-1.335.2-.49.259-1.038.342-1.563.07-.16.108-.333.134-.51.035-.24.049-.487.09-.722.106-.103.035-.297.176-.353l.165.102c.17.326.33.653.517.974.035.06.071.119.108.178.047.183.342.274.307.491.719 1.14 1.497 2.042 2.381 3.046.177.4.802.707.767 1.198.082.296-.06.57.106.821.849.263 1.568 1.016 2.5 1.095.117-.182.471-.296.424-.604-.354-.48-.59-.81-.743-1.37-.177-.17-.047-.364-.141-.546.094-.526-.177-1.005-.083-1.53-.106-.034-.07-.182-.118-.273.189-.16 0-.4.059-.57-.33-.206.2-.469-.141-.64-.038-.43.028-.87.092-1.296v-.002c.017-.115.034-.23.049-.344.07-.195-.023-.446.083-.628.02-.238.053-.481.095-.726.076-.434.185-.869.329-1.27a.585.585 0 01.094-.628c-.13-.046-.188-.24-.342-.217.012-.08-.047-.183.06-.228.105-.662.483-1.233.613-1.883h-.083c-.106-.09-.024-.239-.024-.353-.118.069-.188-.034-.283-.092-.448.354-.483.8-.766 1.21-.471 1.015-1.096 1.973-1.332 3.092-.055.12-.1.243-.136.369M30.544 51.621c.074-.16.264-.317.032-.457.091-.407.263-.79.39-1.184-.024-.096-.001-.152.038-.193.099-.098.3-.1.179-.362.172-.333.355-.568.47-.925.343-.407.526-.888.905-1.357.263-.247.515-.592.664-.925.114-.21.355-.297.378-.555-.035-.099-.126-.049-.195-.062-.94.494-1.695 1.37-2.474 2.048-.15.258-.332.567-.573.74-.335.463-.678.923-1.015 1.397a21.633 21.633 0 00-1.024 1.563c-.075.09-.144.183-.21.276v.002c-.463.669-.754 1.456-1.325 1.942-.24-.197-.218-.604-.298-.924-.037-.34-.078-.68-.123-1.02a52.991 52.991 0 00-.278-1.838 80.55 80.55 0 00-.653-3.384c-.172-.297-.573-.654-1.008-.592-.034.086-.126.049-.195.061-.068.136-.32.21-.194.445.08.395.034.702.114 1.085.115.755.23 1.579.278 2.385.01.156.016.31.02.464.033.453.13.911.206 1.374.058.357.102.717.092 1.081.114.58.229 1.246.332 1.862.137.74.584 1.727.057 2.369-.022.136-.091.382.058.443.01.47.503.309.71.593.458.234.848.74 1.375.826.206-.037.378-.049.583 0 .07-.16.298-.185.298-.382.207-.222-.045-.53.115-.765.035-.715.527-1.16.71-1.813.539-.444.71-1.184 1.043-1.74.023-.086-.08-.086-.103-.148.32-.505.676-1.023.825-1.664-.115 0-.195-.087-.275-.148-.07-.136-.012-.284.046-.408a.206.206 0 01.025-.106v-.004" />
-					<g transform="translate(33.6 44.4)">
-						<mask id="b-2" fill="#fff">
-							<use xlinkHref="#a" />
-						</mask>
-						<path
-							d="M2.239 6.87a.86.86 0 01.037-.094c.062-.553-.04-1.095.1-1.635V5.14c.01-.04.021-.08.034-.12-.179-.267.011-.615-.056-.894.135-.57-.134-1.14-.078-1.697-.101-.592-.123-1.138-.19-1.708 0-.174.268-.337.056-.5h-.828C1.1.141.787.151.742.442.608.407.485.49.43.617c.112.348-.123.824 0 1.22l-.056.058C.384 2.592.205 3.266.24 3.963a7.93 7.93 0 00-.15 1.177l-.018.357a53.9 53.9 0 00-.026 1.376v.003c-.006 1.402.049 2.89.216 4.28.112.221.078.512.168.756-.045.255.09.51.134.778-.045.326.168.534.135.86.036.135.073.271.115.405.123.395.291.764.635.99.224.081.447 0 .503-.245-.012-.24-.025-.49-.04-.745v-.001c-.04-.658-.1-1.343-.228-1.938-.057-.522-.123-1.208-.023-1.731-.034-.535.168-1.07.135-1.615.056-.128.145-.303.111-.5.323-.378.191-.877.333-1.297V6.87M9.126 7.077c-.627.024-1.254.326-1.847.175.021-.127.063-.254.11-.383.102-.285.227-.58.204-.907.044-.232.48-.384.302-.698.011-.041.026-.082.042-.123.1-.245.279-.48.327-.76.134-.22.246-.44.403-.65.355.39.418.942.624 1.41.015.034.03.068.048.1.248.504.423 1.093.625 1.631.008.024.015.045.024.066-.235.22-.594.058-.862.14zm4.947 7.926c-.045-.325-.257-.535-.134-.895-.035-.05-.066-.1-.094-.153-.134-.253-.198-.535-.253-.823-.068-.489-.515-.895-.358-1.476l-.045-.081c.01-.5-.045-.825.045-1.29-.134-.558-.045-1.116-.18-1.674-.26-.559-.337-1.195-.624-1.739-.047-.09-.1-.176-.159-.26-.19-.395-.325-.86-.649-1.15a.977.977 0 01-.119-.321v-.002c-.033-.157-.048-.324-.094-.479-.403-.522-.671-1.173-1.13-1.673-.27-.371-.481-.836-.94-.976-.426-.244.022-.709-.023-.976A2.789 2.789 0 008.107.78c-.112.034-.212.162-.324.115-.806.849-1.187 2.115-1.769 3.15.1.243-.168.325-.168.523-.118.155-.174.396-.257.571h-.001l-.01.021c-.079.14-.18.279-.213.442-.078.36-.28.65-.426.976.065.111-.008.195-.048.294a.223.223 0 00-.007.182c-.359.419-.247 1.035-.537 1.476.022.279-.269.337-.157.64-.392.255.1.708-.325.929-.09.453-.268.755-.268 1.254-.19.477-.101 1.116-.057 1.639.128.363.315.862.687.963h.002c.02.006.04.01.061.013a.787.787 0 00.192-.013h.003a.496.496 0 00.376-.324c.023-.476.124-1.069.347-1.533.168-.222.023-.593.291-.745-.056-.266.101-.522.19-.731-.1-.337.28-.407.292-.698.022-.394.234-.743.515-.999.582.07 1.31-.116 1.902-.174.638 0 1.276-.302 1.992-.279.359.162.303.638.46.953.201.29.033.65.235.918.045.848.57 1.499.805 2.266.012.418.504.592.404 1.022.106.104.204.212.295.322v.002c.358.43.617.905.947 1.35.392-.047.48.383.862.278-.101-.197-.168-.406-.325-.58z"
-							mask="url(#b)"
-						/>
-					</g>
-					<path
-						d="M25.905 3.774c.553-.093.979.151 1.52.104.219.14.553-.023.748.14.254.047.623.012.887.14.242.046.461.162.737.116.104.104.323.023.45.14.207-.094.264.173.472.08.138.035.242.163.414.116.012.105.162.012.22.082.379.093.713.302 1.105.36-.012.046.046.07.08.058h.08c.093.046.208.14.335.14.127-.012.219.139.334.08.184.082.356.268.552.221.208.128.404.221.657.28l.864.452c.23-.012.345.232.575.22-.011.024.012.035.035.059.15.058.31.174.495.22.703.36 1.405.837 2.13 1.232 1.417.894 2.776 2.067 3.813 3.39-.012.175.23.233.218.396.288.418.542.72.691 1.196.116.22.173.418.3.616.034.127.115.243.172.36-.023.174.138.394.138.557.035.082.07.21.139.314-.058.22.126.383.104.557.08.256.206.534.115.836.115.175-.012.453.115.64l-.057 1.312c-.082.197-.082.441-.116.662.034.267-.173.36-.104.639-.103.046-.046.15-.057.232-.116.07-.023.279-.116.383-.057.093-.08.244-.103.36-.08.244-.184.523-.276.79-.047-.011-.07.035-.058.081-.092.28-.265.64-.357.976-.035.174-.173.314-.196.5-.07.057-.092.185-.138.278-.23.406-.484.778-.691 1.232-.161.011-.058.22-.196.22-.253.453-.725.778-1.025 1.08-1.174.709-2.383.825-3.754.894-.599.21-1.197-.08-1.739.256-.207-.07-.333.116-.506.058a13.25 13.25 0 00-3.145.918c-.725.279-1.335.545-1.969.94-.23.116-.449.28-.69.372-.415.314-.876.709-1.325.941-.161-.29-.046-.535-.057-.894l-.046-.047.046-.058.092-2.346c.08-.209.046-.464.138-.662-.092-.349.15-.546.104-.871l1.105-1.08c.657-.953 1.347-1.765 1.969-2.706.092-.268.346-.465.415-.779.253-.406.576-.871.795-1.254.333-.558.564-1.162.863-1.707-.046-.128.07-.21.104-.279.115-.314.218-.627.334-.941.046 0 .07-.046.057-.093V16.2c.104-.022.012-.174.104-.22.173-.43.334-.999.45-1.417-.093-.325.15-.51.103-.813l.057-.08.058-1.755c-.115-.244-.011-.557-.08-.813.011-.534-.208-1.069-.357-1.533-.173-.406-.438-.732-.61-1.115-.163-.14-.243-.325-.335-.5-.633-.638-1.267-1.23-2.073-1.707-.138 0-.253-.162-.414-.162-.127-.14-.38-.174-.53-.28-.138-.01-.3-.127-.495-.116-.104-.046-.242-.115-.38-.139l-.058-.046c-.08-.035-.276-.058-.415-.093-.195-.104-.437-.128-.667-.162-.07-.106-.22-.036-.334-.059-.093-.093-.323-.023-.438-.139-.288.011-.518-.14-.829-.082-.161-.127-.438.035-.576-.14-.564-.092-1.186-.022-1.716-.162-.219-.022-.472.059-.633-.092-.898.023-1.716-.128-2.58-.186-.184-.047-.426-.14-.518-.232-.115-.21.08-.28.219-.36.771-.128 1.624-.094 2.43-.14l.057-.058c.922.058 1.889-.012 2.764.174"
-						fillOpacity=".804"
-					/>
-					<path
-						d="M20.355 9.668c.189.058.377.115.566.16.213-.011.354.138.566.115.035.115.2.011.283.08.436.138.92.275 1.332.412.732.206 1.475.412 2.288.447.142.114.377-.012.507.103.401-.012.767.023 1.132.057.047.115.224-.012.236.137.047.86.224 1.752.283 2.61 0 .115.118.184 0 .298-.154.024-.342-.045-.46.058-.637-.058-1.238.126-1.804.057h-4.245v-.023c-.212-.034-.377-.068-.566.023-1.108-.034-2.063.057-3.172.023l-.082.057c-1.262-.114-2.382.103-3.656.034l-.059.046c-1.344-.114-2.735.126-4.068 0-.27-.091-.613-.034-.907-.046-.036-.08-.166-.171-.118-.309.2-.023.259-.24.424-.274.165-.23.425-.298.649-.47 1.202-.778 2.546-1.397 3.773-1.946.141-.057.365-.08.507-.218.342-.114.672-.217 1.014-.332.483-.263 1.108-.366 1.615-.55.236.035.436-.182.708-.136.165-.046.318-.138.507-.138.33-.126.731-.08 1.05-.252.223.069.377-.114.565-.046v-.034c.413-.057.72.126 1.132.057"
-						opacity=".8"
-					/>
-					<path
-						d="M.943 21.878c.069.22.149.438.16.646.104.184.184.45.207.669.138.16.092.415.253.565.126.461.39.865.632 1.257.3.496.529.992.931 1.418l.023.08 1.38 1.443c.16.069.286.3.459.369.437.45.965.818 1.494 1.142.184.172.437.253.644.437.149-.034.218.15.344.185l.023-.012c.173.139.368.231.552.346h.046c.345.174.666.392 1.011.531.127 0 .253.115.368.173.15.173.46.092.598.288.137.035.253.115.38.184.137 0 .217.14.367.116.16.093.356.104.494.242.253.047.506.104.724.243.126 0 .218.07.333.127.253.011.448.184.69.242.184-.035.275.173.46.126.333.173.666.185 1.01.335.185-.023.311.08.46.127.092.011.207-.046.265.057.253-.034.368.116.62.093.046.08.161-.023.184.092.104.046.265-.034.357.069.08.023.207-.046.253.058.333.023.631.138.953.196.138 0 .276 0 .368.08.242-.057.414.104.632.07.07.127.253-.011.322.115.16.023.356-.069.46.07.367-.035.666.057.977.126.483.035 1 .012 1.494.024.092.057.195.207.126.345-.563.98-1.862.969-2.827 1.372l-3.643 1.085c-1.782.012-3.483-.012-5.24-.058-1.197-.496-2.323-1.188-3.483-1.788-.977-.83-1.977-1.672-2.977-2.398-.23-.22-.506-.438-.747-.658-.15-.057-.253-.276-.414-.322-.333-.462-.827-.739-1.207-1.12-.24-.507-.413-1.037-.678-1.533-.15-.588-.494-1.096-.597-1.695l-.046-.058L0 22.65c.184-.46.15-1.06.517-1.395.07-.035.15-.011.207.012.16.173.127.415.218.611"
-						fillOpacity=".8"
-					/>
-					<path d="M36.551 8.16v.058c-.16 0-.264-.115-.39-.173-.483-.162-1.058-.335-1.46-.496.644.127 1.23.403 1.85.61zm-10.481.634v.392c-.299-.011-.678.024-.943-.138-.368-.058-.77.15-1.08 0l.448-.093c-.184 0-.207-.207-.413-.138.643-.115 1.321-.115 1.93-.104l.058.081zm.023.554c.046.322.207.75.115 1.084-.942-.092-1.908-.277-2.874-.392-.344-.115-.712-.173-1.045-.3v-.058a18.83 18.83 0 013.804-.334zm-9.677 1.96c-.31.116-.632.381-1 .312v-.058c.357-.011.632-.322 1-.254zm29.617 8.42h-.162c-.022-.197.012-.462.138-.612v-.75c.088.044.09.13.086.226-.004.077-.014.16.018.224.046.312.103.646-.08.911zm-19.47-2.157c-1.183.057-2.286-.012-3.504.057-.702.092-1.414-.034-2.127.139-1.609.115-3.24.334-4.827.599-.581.015-1.11.12-1.648.226-.323.06-.65.123-.995.166-1.574.208-3.149.577-4.723.854-.138.138-.357.08-.518.173-.459-.343-.996-.718-1.337-1.194-.256-.358-.401-.773-.317-1.274.05-.148.1-.297.154-.443.174-.458.398-.878.834-1.08 1.597-.057 3.264-.126 4.907-.219.334.012.701.104 1 .024 1.804-.047 3.85.126 5.677-.047 1.253.035 2.472-.092 3.701 0 .357.012.713-.046 1.07.047.183-.14.424.046.585-.047.678.093 1.391.035 2.011.104.067.384.039.8.031 1.216 0 0-.002.47.027.699zm1.851-1.88c.575-.092 1.173-.058 1.793-.139 1.804-.08 3.712 0 5.563.138.747-.01 1.528.116 2.263.105.885.115 1.701.138 2.564.276.93-.07 1.804.093 2.735.196.264.162.551.035.85.116.322.256 1.074-.113 1.082.49l-.001.063c.056.538.117 1.103.092 1.657-.016.346-.067.69-.173 1.019-.276.161-.552-.035-.804-.104-.104-.196.034-.6-.242-.669.1-.124.044-.188-.032-.25H44.1c-.064-.054-.144-.105-.16-.188-.005.067-.005.13-.002.19h.042l-.042.002c.019.286.125.524.107.858-1.518-.254-2.909-.577-4.437-.808-.07-.017-.141-.035-.212-.051-2.315-.544-4.918-.466-7.293-.78-1.218 0-2.516-.092-3.723-.137a20.008 20.008 0 01-.03-.803s.01-.792.064-1.181zm16.182 5.858l-.046-.046c.127-.184.334-.242.414-.473.057-.023.092.035.138.058l-.506.461zm-1.402.67v-.082c.299-.472.713-.807.943-1.326l.24.046c-.275.635-.689.97-1.183 1.361zm-.551 0c-1.046.703-2.057 1.326-3.31 1.499-.265.045-.495.138-.747.219-.805.277-1.69.138-2.54.334-1.598-.115-3.092.104-4.69-.093-.46-.08-.999.07-1.424-.08-.644-.115-1.242-.115-1.932-.104-.115-1.58.277-3.125.3-4.682l.057-.058c1.47-.058 2.954-.15 4.413.081 1.172-.057 2.275.196 3.448.23 1.31.047 2.62.255 3.919.416 1.207.196 2.38.473 3.55.738-.137.623-.597 1.072-1.045 1.5zm-16.331-.947c.103.404-.08.82-.058 1.246-.057.485-.08.98-.184 1.442-.701-.058-1.368.127-2.046-.058-.092.023-.23.081-.299-.023-.586-.058-1.195-.116-1.747-.161l-.045-.058c-1.23-.023-2.506-.127-3.644-.45-1.643-.162-3.114-.484-4.631-.934-.207-.15-.471-.173-.724-.196-1.035-.3-1.897-.68-2.862-1.165v-.058c1.206-.173 2.402-.472 3.586-.634 2.045-.357 4.01-.519 6.09-.715 1.138-.161 2.449-.15 3.61-.22 1.023-.126 2.115-.068 3.172-.092.046.715-.173 1.373-.219 2.076zm2.01-9.04c1.828.183 3.713.137 5.552.38 2.436.104 4.47-.047 6.815-.197l.908-.138c.793.577 1.597 1.35 2.23 2.053.138.161.333.38.39.577-1.276-.139-2.574-.37-3.919-.439-.23.162-.425-.08-.69 0-3.125-.242-6.297-.196-9.32-.357-.288-.15-.61.046-.908-.058l-1-.08c-.15-.173.16-.404-.023-.612.115-.357-.103-.76-.035-1.13zm17.164-.52c1.702-.534 3.428-.996 5.04-1.73.54-.139.954-.473 1.414-.75-.046-.057-.046-.08-.035-.161-.183.08-.379.15-.551.277l-.08-.116c-.196-.034-.346.058-.495.139-.15.058-.356.253-.471.058-.16.08-.356.207-.517.195.103-.184.402-.195.551-.357-.126-.057-.298 0-.413-.057.057-.185.356-.082.356-.3-.069-.058-.15-.023-.218 0-.092-.07-.242-.046-.196-.197-.344.139-.666.3-1.022.415-1.184.485-2.448.808-3.701 1.131-.678.115-1.333.311-2.011.473-.587-.335-1.115-.634-1.678-1.027-.449-.403-1.115-.403-1.575-.853-.598-.173-1.115-.52-1.713-.692-.183.07-.401-.115-.608-.173-.219-.012-.54-.323-.656-.046l-1.08-.415c-.47.046-.943-.289-1.436-.37.034.058-.035.15-.104.174l-1.793-.473v.08c-.24.105-.517-.057-.77-.08-.413.3-.988-.265-1.333.162l-.023-.024c-.172 0-.494-.138-.552.139-.068-.012-.16.023-.218-.024-.483.081-.92-.057-1.413.081v.081c.988-.081 2.022.357 3.01.334 1.863.381 3.644.935 5.322 1.742a42.616 42.616 0 013.23 1.684l-.058.058c-1.655.115-3.367.345-5.069.334-1.735.034-3.47-.139-5.137-.277-.448-.138-1.069 0-1.379-.358l-.161-1.695c-.058-.127-.15-.346-.034-.438l1.138.057c.103-.045.114-.103.103-.218.15-.081.38.023.47-.058-.39-.219-.895-.15-1.32-.277-.219-.057-.472 0-.633-.173-.31-1.718-.804-3.264-1.138-4.982-.023-.254-.172-.53-.08-.773.184.265.39.312.69.358.264.058.597.058.746-.185-.023-.288.185-.472.115-.784.334-.057-.068-.288.024-.438.195-.254-.219-.289-.254-.473-.206-.208-.55-.254-.827-.38-.333.057-.574-.277-.965-.197-.264.023-.39.312-.598.358.127.265-.23.058-.195.277-.276-.023-.08.253-.334.276-.011.116.058.15.08.254.012.058-.149.092-.057.162.472 1.995.391 4.186.943 6.17-.012.092.172.173.023.253-1.023-.035-2.092.138-3.115.277-.241.058-.54-.023-.747.139-.839.103-1.632.334-2.482.437-.575-.346-1.241-.495-1.816-.853v-.035c-.31-.103-.563-.3-.908-.38-.724-.519-1.563-.865-2.206-1.465-.07-.23-.207-.45-.277-.645-.333-.196-.321-.692-.746-.854-.242.023-.414-.161-.69-.057-.207-.047.046-.15-.023-.277-.069-.046-.092-.208-.195-.22-.093.289-.46.324-.667.416-.23-.023-.471.23-.609.415-.092.208-.264.23-.276.473-.126.23.253.288.092.496.138.126.16.404.299.553-.115.323.23.45.39.577 1.713.68 3.403 1.73 5.012 2.526v.057c-1.173.392-2.391.75-3.529 1.188-.832.318-1.552.623-2.35 1.013l-.097.048a24.735 24.735 0 00-4.747 3.01c-1.207.173-2.414.092-3.62.277-.207.196-.538.103-.807.152a.73.73 0 00-.044.01l.08.138c-.057.046-.149.023-.218.023v.092c.494.023.943-.104 1.437-.058.011.023.069.046.057.081-.23.104-.517.035-.781.058-.092.104.057.173.115.242.643.035 1.195-.058 1.77-.023-.115.08-.357.069-.472.115-.184-.011-.402-.011-.494.081.345.116.644.011.989-.035.356.046.632.035.965.116-.035.184-.15.334-.218.496-.066.122-.12.251-.163.385a2.717 2.717 0 00-.01 1.552c.013.056.028.112.046.169.25.76.95 1.291 1.39 1.884-.804.357-1.666.611-2.482.968-.045.104.104.197-.023.278-.138.01-.115.184-.115.241-.069.035-.172.047-.195.14-.138.264-.494.322-.77.449v.08c.242-.057.61-.242.804 0 .53-.23 1.207-.357 1.679-.691.149-.173.459-.104.609-.277.241-.012.436-.208.666-.162-.103.07-.241.185-.39.185-.265.219-.61.265-.885.392-.138.115-.437.07-.334.358h.138c.563-.231 1.126-.427 1.69-.635.505-.138 1.195-.588 1.655-.023 1.218.646 2.517 1.072 3.861 1.43 1.552.6 3.276.773 4.965 1.026.126.07.322-.022.437.081.862-.011 1.609.265 2.459.277 1.908.323 3.93.507 5.93.692.012.507-.114 1.026-.195 1.523a34.583 34.583 0 00-.494 2.906c-.104.184.08.392-.058.553.024.219-.206.404-.114.588-.058.024.023.15-.081.139.012.173.012.346-.103.472.057.15-.115.231-.115.323-.012.024-.035.047-.058.035v.796c.127.103-.011.311.115.415-.08.38.265.508.242.865.172.346.655.438.862.692l.023-.035c.172.058.413.15.551.035.253 0 .402-.104.61-.22.355-.172.678-.426.988-.691-.047-.288.356-.265.195-.553.322-.185-.057-.485.115-.669.069-.127 0-.208-.08-.278.01-.218.103-.449-.115-.576-.22 0-.185-.242-.277-.37-.08-.322-.378-.068-.528-.299.035-.22-.035-.484-.184-.634-.146-.283-.017-.488-.163-.757a.144.144 0 00-.01-.016c.058-.346.093-.761.253-1.003.207-.508.127-1.13.357-1.627.057-.472.115-.945.16-1.418.081-.184.207.012.334-.057 2.31.08 4.77.057 7.034.23.551-.07 1.069-.127 1.597-.058 2.08-.265 4.195-.426 6.125-1.026 1.736-.657 3.31-1.407 4.586-2.964.701.3 1.483.508 2.173.888.482.265 1.068.462 1.413.888.15.323.506.461.85.554.138-.266.471-.035.667-.312-.034-.023-.08-.057-.058-.103a.29.29 0 01.277-.08c.379-.186-.22-.44.218-.59l-.058-.057c.012-.3-.356-.323-.47-.553-.472-.139-.851-.508-1.345-.6-1.012-.554-2.184-.807-3.253-1.28.066-.6.148-1.175.13-1.772a5.378 5.378 0 00-.026-.385c.01-.45-.23-.842-.196-1.269.908.104 1.759.335 2.655.438.575.139 1.16.289 1.793.3-.07-.138-.288-.276-.471-.323.092-.369.46 0 .689-.115.046-.092-.08-.115-.058-.196.185-.115.369.104.61.035.046-.104.103-.023.16-.035.139-.104.368-.023.587-.08v-.081l.003-.007c.014-.028.046-.028.077-.028.092-.092-.045-.115-.115-.138-.172-.15-.47-.139-.712-.22-1.07-.23-2.172-.541-3.31-.691-.804-.138-1.632-.277-2.482-.358l-.151-.285c-.657-1.185-1.542-2.236-2.309-3.198.25-.083.5-.163.752-.242zm-21.38 0c.735.152 1.481.28 2.24.38.091.15.022.37.08.554.15.427.057.9.08 1.35l-.16.057c-1.23-.092-2.345-.046-3.54-.138-1.782.265-3.414-.046-5.207.115-.333.058-.758-.011-1.16.046-1.288.138-2.552.023-3.862.173-1.115-.035-2.15.058-3.23.104-.022-.15.161-.023.196-.162.16.046.333-.16.494-.253 1.184-.588 2.39-.946 3.586-1.488.483-.242 1.023-.404 1.494-.611v-.115c-.506.069-1.011.288-1.494.507-1.069.254-2.057.773-3.034 1.188.597-.346 1.333-.773 2.068-1.05.573-.136 1.033-.518 1.593-.657l.021-.005c.058-.014.096-.021.156-.03.966-.438 1.988-.726 3.034-.968.46-.139.897-.346 1.38-.415.448-.047.758.276 1.183.322 1.332.43 2.688.806 4.082 1.095z" />
-				</g>
-			</g>
-		</svg>
-	);
-};
