--- src/components/iconSet/ts/logos/IconKreuzbund.tsx
@@ -1,207 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconKreuzbund = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			viewBox="0 0 62 60"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<g fill="none" fillRule="evenodd">
-				<path d="M1 1h60v60H1z" />
-				<path
-					d="M1.985 32.605L1 16.144s9.6-2.315 10.832-2.304c1.232.01 7.221-.659 7.221-.659l.985-11.852L40.556 1l1.474 11.524 18.709 3.621-.985 16.132L42.03 30.63l-2.298 15.803s-6.771-1.52-9.847-1.646c-3.075-.125-11.817.988-11.817.988V32.277l-16.083.329z"
-					stroke="#FFFFFE"
-					strokeWidth=".558"
-				/>
-				<path
-					d="M4.367 56.946l1.647 1.783c.331.367.48.541.48.673 0 .148-.193.218-.593.218v.227h2.962v-.227c-.505-.026-.906-.35-1.603-1.119l-1.9-2.106 1.142-1.205c.549-.568 1.028-1.014 1.855-1.058v-.227H5.831v.227c.41 0 .592.07.592.271 0 .157-.104.323-.374.62l-1.682 1.8V54.85c0-.62.166-.69.837-.717v-.227H2.38v.227c.636.026.801.096.801.717v4.054c0 .62-.165.69-.801.717v.227h2.823v-.227c-.671-.026-.837-.096-.837-.717v-1.957M11.107 57.226h.383l2.03 2.622h1.664v-.228c-.33.027-.487-.104-.696-.358l-1.69-2.158c.853-.219 1.463-.69 1.463-1.643 0-.979-.697-1.556-2.605-1.556H9.103v.227c.654.027.82.097.82.717v4.055c0 .62-.166.69-.82.716v.228h2.84v-.228c-.67-.026-.836-.096-.836-.716v-1.678zm0-2.657c0-.227.096-.314.505-.314.715 0 1.43.384 1.43 1.267 0 .97-.767 1.337-1.935 1.354V54.57zM20.531 59.847l.4-1.59h-.226c-.505 1.048-1.132 1.24-2.195 1.24-1.046 0-1.15-.122-1.15-.48v-2.054h1.01c.767 0 1.115.175 1.238.874h.226V55.74h-.226c-.114.62-.314.786-1.238.786h-1.01v-1.878c0-.245.06-.306.33-.306h.898c1.159 0 1.385.21 1.595.874h.243l-.034-1.311h-4.95v.227c.567.026.733.096.733.717v4.054c0 .62-.166.69-.732.717v.227h5.088M23.275 54.849c0-.62.165-.69.767-.717v-.227h-2.719v.227c.602.027.767.097.767.717v2.971c0 1.503.802 2.167 2.405 2.167 1.69 0 2.282-.909 2.282-2.324v-2.648c0-.577.157-.813.837-.883v-.227h-2.056v.227c.627.07.784.306.784.883v2.656c0 1.093-.506 1.792-1.525 1.792-1.09 0-1.542-.524-1.542-1.538v-3.076M32.437 59.847l.375-1.608h-.262c-.418.918-.827 1.171-1.838 1.171H28.96l3.78-5.226v-.28h-4.73l-.21 1.521h.262c.21-.682.54-1.083 1.577-1.083h1.63l-3.782 5.225v.28h4.949M36.174 59.847c1.568 0 2.317-.682 2.317-1.643 0-.909-.862-1.302-1.629-1.45v-.018c.723-.122 1.316-.612 1.316-1.363 0-.961-.906-1.468-2.335-1.468h-2.614v.227c.636.026.802.096.802.717v4.054c0 .62-.166.69-.802.717v.227h2.945zm-.958-2.849c1.35-.026 2.064.376 2.056 1.206-.009.656-.514 1.294-1.438 1.294-.496 0-.618-.14-.618-.376v-2.124zm0-2.342c0-.288.165-.402.47-.402.871 0 1.307.49 1.307 1.136 0 .857-.505 1.259-1.777 1.259v-1.993zM40.887 54.849c0-.62.166-.69.767-.717v-.227h-2.719v.227c.602.027.767.097.767.717v2.971c0 1.503.802 2.167 2.405 2.167 1.69 0 2.282-.909 2.282-2.324v-2.648c0-.577.157-.813.837-.883v-.227H43.17v.227c.627.07.784.306.784.883v2.656c0 1.093-.506 1.792-1.525 1.792-1.089 0-1.542-.524-1.542-1.538v-3.076M50.781 57.995l-3.529-4.09h-1.611v.227c.366.088.6.254.801.603v4.003c0 .577-.157.813-.801.882v.228h2.09v-.228c-.644-.07-.8-.305-.8-.882v-3.434l4.042 4.683h.296v-4.972c0-.577.157-.813.801-.882v-.228h-2.09v.227c.644.07.801.306.801.883v2.98M55.377 59.847c1.865 0 3.38-.996 3.38-2.971s-1.515-2.971-3.38-2.971h-2.962v.227c.583.026.749.096.749.717v4.054c0 .62-.166.69-.75.717v.227h2.963zm-1.028-5.05c0-.403.165-.543.697-.543 1.62 0 2.492 1.101 2.492 2.622 0 1.52-.871 2.622-2.492 2.622-.532 0-.697-.14-.697-.542v-4.16zM21.103 44.121c-.048.14-.076.295-.076.478 0 .146.125.25.166.375h-.499v-.853h.409zm4.159 0c.175.06.333.106.547.106.193 0 .341-.042.482-.106h-1.03zm8.3 0c-.153.033-.26.122-.438.17-.005-.047-.037-.065-.037-.11a.09.09 0 01.02-.06h.455zm3.49 0c-.1.045-.207-.004-.304.06-.146.09-.1.27-.146.44-.44-.147-.636-.44-1.099-.44-.251 0-.585.022-.585.274 0 .556.69.68 1.244.68h.476c.357 0 .559.11.916.11-.046.138-.11.22-.11.367 0 .046.036.064.036.11.165-.004.257-.036.421-.036.147 0 .22.073.366.073.055 0 .087-.041.128-.073.12.087.165.179.257.293h-.66c-.09-.041-.15-.11-.255-.11-.078 0-.106.074-.183.074-.586 0-.88-.33-1.465-.33h-.403c-.292 0-.439-.184-.732-.184h-.33c-.315 0-.489-.147-.805-.147h-.402c-.124 0-.17-.11-.293-.11h-.874a.388.388 0 00-.005-.404c.435 0 .668-.161 1.099-.22.224-.028.549.008.549-.221a.224.224 0 00-.132-.207l3.29.001zm-10.537.133c.325 0 .463.293.787.293a.596.596 0 00.476-.22c.083.096.183.16.183.294 0 .078-.06.11-.073.183-.06-.013-.082-.073-.146-.073h-.732c-.083.1-.16.183-.293.183-.106 0-.151-.073-.256-.073-.106 0-.152.073-.257.073h-1.537c-.485 0-.76.106-1.245.147-.014-.068-.028-.114-.037-.183.5-.046.71-.44 1.208-.44h.403c.124 0 .174.1.293.11.366.027.581.036.952.036.133 0 .183-.16.183-.294.037 0 .055-.036.091-.036zm3.863.33v.367h-.623a.188.188 0 00-.073-.073v-.11c.28 0 .458-.042.696-.184zm3.917 0c.16 0 .33.096.33.257 0 .152-.197.184-.348.184-.129 0-.202-.055-.312-.11.064-.166.151-.33.33-.33zm-12.996.22c.055.143.146.211.146.367 0 .092-.055.134-.073.22-.027 0-.041.038-.073.038-.115 0-.147-.138-.256-.147-.426-.046-.678.096-1.099.009v-.124c.12-.028.188.005.312.005.178 0 .251-.138.42-.184.243-.064.408-.055.623-.184z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.446 43.192a.996.996 0 01-.079.055v-.055h.079zm1.176 0c-.198.362-.41.617-.519.948h-.408v-.278c.332-.043.54-.064.872-.064l.077-.607h-.022zm5.455 0c.04.16.072.33.108.494-.208.085-.73-.176-.959-.176-.81 0-1.154.287-1.965.287l.222-.606h2.593zm9.899 0c.122.018.246.038.358.071.247.075.343.3.604.3.114 0 .183-.047.275-.113a.19.19 0 00.073.075v.225c-.256.014-.412.052-.66 0-.53-.112-.754-.524-1.299-.524-.137 0-.21.075-.348.075-.091 0-.128-.075-.22-.075h-.695c-.096 0-.142.075-.238.075-.123 0-.187-.075-.311-.075-.146 0-.224.042-.366.075a.812.812 0 00-.044-.11h2.87zm-10.685.947c.162-.073.315-.176.515-.278-.333-.255-.54-.425-.956-.425-.395 0-.478.425-.79.637l.202.067h1.029zm6.816 0c.043-.054.146-.065.2-.127.082-.093.073-.262.201-.262.316 0 .453.262.769.262.197 0 .302-.084.494-.112.54-.084.86-.075 1.41-.075.356 0 .54.187.896.187v.113a.215.215 0 01-.026.014h-3.29a.169.169 0 00-.07-.014.584.584 0 00-.129.014h-.455zm.639-.8a.257.257 0 01-.146 0 .743.743 0 01.146 0zm-8.768 0c.105 0 .151.074.256.074.225 0 .343-.075.568-.075.174 0 .274.019.44.075-.047.075-.092.15-.184.15h-1.281c-.041-.052-.078-.09-.11-.15.119-.028.187-.075.311-.075z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.895 42.262c-.073.405-.141.712-.453.949l-.076-.001v-.754a7.63 7.63 0 00.086-.195l.443.001zm1.403 0c.048.06.097.124.145.194h-.582c-.106-.05-.19-.126-.276-.194h.713zm7.516 0c.096.122.16.3.33.3.165 0 .202-.225.367-.225.169 0 .233.188.402.188.064 0 .092-.043.147-.075v.3c-.11-.076-.179-.15-.312-.15-.37 0-.453.486-.823.486h-.476c-.22 0-.339.056-.55.113v-.188h.294c.123-.089.292-.121.292-.28 0-.249-.279-.295-.512-.356v-.113h.84zm4.109 0c.19.199.3.532.61.532.32 0 .462-.26.634-.531l.659-.001-.067.046c.05.01.096.004.146 0a.117.117 0 01-.03-.047l.909.001c-.013.269-.513.218-.513.495 0 .35.346.396.704.452h-2.87c-.157-.322-.496-.452-.872-.452h.183c.114 0 .22-.089.22-.206 0-.22-.277-.185-.467-.289h.754zm-12.278.95l.007-.054-.029.054h.022zm2.838-.002l.084-.229c.166-.297.415-.51.768-.51h1.08c.402 0 .559.337.661.738h-2.593zm.86-.753c.115 0 .18.057.275.113.01.028.019.047.037.075-.041.01-.064.037-.11.037-.165 0-.256-.084-.403-.15.07-.042.12-.075.202-.075zm9.794.225c.082 0 .123.061.183.112-.05.07-.105.108-.146.188h-.184c0-.061-.036-.09-.036-.15 0-.094.091-.15.183-.15z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.752 40.404c.02.083.038.169.038.278 0 .166.27.166.333.291.166-.125.374-.25.374-.479a.705.705 0 00-.006-.09h-.739zm4.55 0a.822.822 0 00.382.097.822.822 0 00.381-.097h.189c.165.37.4.555.635.889h-.582c-.229-.104-.353-.334-.623-.334-.125 0-.208.084-.333.084-.083.458-.145.792-.54 1.042v-.792c.156-.332.238-.554.316-.888l.175-.001zm4.982 0c.082.164.134.355.167.57l-.79.082a7.592 7.592 0 01-.039-.652h.662zm6.663 0c-.287.1-.524.144-.785.36-.183.152-.256.313-.476.405-.352.151-.879.073-.879.459 0 .142.096.275.238.275.311 0 .476-.147.787-.147.385 0 .531.367.916.367 0 .051.036.079.036.129v.01h-.909a.184.184 0 01-.006-.028l-.043.028h-.659l.043-.065c-.256-.239-.416-.477-.769-.477h-.366c-.238 0-.366-.11-.604-.11-.165 0-.27.068-.42.004v-.262c.182-.06.246-.183.402-.293.27-.189.659-.216.659-.551a.341.341 0 00-.016-.104h2.85zm1.044 0c-.01.023-.022.046-.034.07a.379.379 0 00.003-.069h.031zm-17.539 1.858c.112-.257.179-.465.246-.768.208.083.312.167.54.167.229 0 .333-.084.54-.167.134.336.322.523.52.768h-.713c-.1-.077-.202-.143-.347-.143-.124 0-.207.083-.332.083l-.01.06h-.444zm8.521.004v-.113c.252.019.389 0 .64 0 .09 0 .15.049.2.114h-.84zm4.196-.004a.359.359 0 01-.046-.028c.197-.019.298-.11.494-.11.13 0 .225.057.306.139h-.754zm1.986-1.68c.201 0 .348.11.421.293-.362.18-.673.028-1.025.22-.105.055-.133.184-.256.184-.078 0-.11-.055-.183-.073.073-.12.21-.11.293-.22.096-.13.054-.368.22-.368.068 0 .095.073.164.073.146 0 .22-.11.366-.11zm1.647.293c.357.018.307.51.604.698.252.16.733-.028.733.275 0 .11-.096.147-.147.239-.201-.262-.42-.326-.732-.44-.28-.102-.732.045-.732-.258 0-.202.087-.33.004-.514h.27zm-11 .477h.22c.008.042.036.065.036.11 0 .102-.11.138-.11.24 0 .233.32.21.55.238v.11a.858.858 0 01-.44.11c-.284 0-.453-.027-.732-.073-.129-.019-.202-.078-.33-.11a.275.275 0 00-.037-.074c.312-.225.476-.431.842-.55z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.698 39.475c-.012.193.01.398.01.636 0 .123.023.216.044.311h.74c-.026-.201-.13-.338-.203-.524l.207-.423h-.798zm5.126 0c.092.077.173.177.307.23.01.028.019.047.037.075-.243.098-.398.15-.659.15h-.366v-.225c.086-.066.142-.159.207-.23h.474zm10.863 0c.067.279.395.417.702.417.274 0 .42-.094.677-.187v.412c0 .122-.03.207-.077.306h-.03c-.008-.154-.108-.306-.26-.306-.228 0-.342.13-.548.224a2.71 2.71 0 01-.204.082h-2.85c-.053-.171-.236-.267-.424-.306-.233-.052-.398.089-.623.004v-.229c.284.052.458-.112.751-.112.27 0 .275.365.53.45.385.13.637.111 1.044.111.293 0 .714.038.714-.262 0-.22-.306-.178-.512-.224-.376-.09-.641-.07-.952-.3.082-.033.184-.047.26-.08h1.802zm-11.559.95l.016-.073c.06.025.11.05.159.072h-.175zm.938 0c.048-.024.099-.048.159-.073l.03.073h-.19zm3.557-.002c0-.17.01-.345.04-.567.301.115.496.309.622.566h-.662zm6.833-.943c-.042.09-.101.131-.183.187a.272.272 0 00-.074-.037v-.15c.101-.005.156-.005.257 0zm.164.487a.62.62 0 01.385.112v.08c-.275.07-.462.07-.732-.005.05-.14.196-.187.347-.187z"
-					fill="#fff"
-				/>
-				<path
-					d="M25.626 38.546c.112.174.249.343.392.533l.54-.085a5.357 5.357 0 01-.01-.447l-.781-.001a.25.25 0 01-.039.016.04.04 0 00-.003-.016h-.099zm4.867 0c.174.074.429.078.639.102.183.024.288-.004.476-.004-.05.15-.096.234-.147.379h-.805c-.247 0-.375-.112-.623-.112-.2 0-.329 0-.512.074a.19.19 0 00-.073-.074v-.365h1.046zm5.954 0a.178.178 0 01.008.054c0 .206-.257.262-.458.262-.325 0-.494-.164-.787-.3a.715.715 0 00-.037-.015l-.706-.001.084.016c-.293.455-.737.422-1.208.675.032.023.05.042.073.075.21 0 .32-.075.53-.075.147 0 .22.075.367.075.178 0 .242-.188.421-.188.128 0 .256.094.256.225 0 .076-.044.117-.106.145h1.803a.46.46 0 01-.013-.108c0-.309.366-.383.366-.693a.47.47 0 00-.022-.147h-.571zm-14.951.948l.084-.172c-.27-.149-.54-.638-.707-.383a1.08 1.08 0 00-.176.554h.799zm3.854.002c.056-.063.118-.11.213-.11a.38.38 0 01.26.109h-.473z"
-					fill="#fff"
-				/>
-				<path
-					d="M2.936 31.113c-.23.127-.438.189-.718.318v.588c0 .137.064.293.202.293.215 0 .315-.147.53-.147.344 0 .54.106.88.074.1-.01.15-.074.255-.074.069 0 .096.074.165.074.471 0 .956-.409 1.208-.01.055-.215-.018-.353-.018-.578a.77.77 0 01-.403.11c-.22 0-.512-.018-.512-.238 0-.115.11-.16.11-.275a.293.293 0 00-.035-.135H2.936zm42.584.002c-.03.009-.062.02-.095.033a.212.212 0 00-.024-.034h.12zm1.87-.001a1.757 1.757 0 00-.026.071h-.805c-.144 0-.247-.038-.358-.07l1.19-.001zm11.658-.001c.156.07.314.126.507.216v.587c0 .138-.064.294-.201.294-.215 0-.316-.147-.53-.147-.344 0-.541.106-.88.074-.1-.01-.15-.074-.256-.074-.068 0-.096.074-.165.074-.47 0-.956-.409-1.208-.01-.055-.215.019-.353.019-.578a.77.77 0 00.402.11c.22 0 .513-.018.513-.238 0-.115-.11-.161-.11-.276 0-.01 0-.021.002-.031l1.907-.001zm-33.324 7.433c-.009-.02-.033-.031-.033-.057 0-.05.046-.074.073-.11.014.055.028.091.037.147a.332.332 0 01-.035.02h.781c.002-.322.006-.598-.155-.936-.083-.166-.374-.166-.374-.375 0-.292.208-.458.208-.75 0-.27-.416-.438-.54-.667 0 .521.02.813.02 1.334 0 .25-.353.292-.353.542 0 .346.11.604.273.852h.098zm3.723 0v-.157h.476c.055-.068.11-.11.11-.202 0-.362-.384-.66-.732-.569 0-.105-.005-.179-.073-.257.279-.036.43-.146.714-.146.356 0 .32.624.677.624h.879c.064 0 .086-.06.146-.074.01 0-.11.024-.146 0l.183.478c-.197-.156-.353-.22-.605-.22-.32 0-.787-.083-.787.238 0 .15.083.234.204.285h-1.046zm5.02 0a4.697 4.697 0 00-.978-.094c-.092 0-.133.046-.22.074a.518.518 0 00.147-.368c.366 0 .563-.133.915-.22.188-.046.44-.06.44-.257 0-.29-.445-.12-.696-.257-.17-.092-.243-.257-.44-.257v-.698h.11c.288.097.421.23.696.367.22.11.412.083.586.258.169.17.123.514.366.514h.33c.26 0 .274.344.512.44.317.131.69.218.783.499h-.572c-.038-.125-.212-.168-.358-.168-.16 0-.242.073-.402.073-.312 0-.403-.367-.714-.367-.183 0-.302.087-.421.22v.147c.238.061.407.01.622.095h-.706zm24.136-7.217h.11l.073.073c-.078.115-.151.22-.293.22-.105 0-.165-.064-.256-.11.082-.133.224-.114.366-.183zM25.436 33.64c0 .229.083.354.083.583 0 .167.27.167.333.292.166-.125.374-.25.374-.48 0-.25-.125-.395-.208-.604l.29-.583c-.27-.146-.54-.625-.706-.375-.249.375-.166.708-.166 1.167zm8.837-.706c.229.041.353.062.582.083.041.209.083.313.083.521 0 .271-.166.417-.208.688l-.457.041v-1.333zm.387-1.716h.403v.11a.964.964 0 00-.256.514c-.032 0-.042.037-.073.037-.27 0-.32-.335-.586-.368v-.183c.196-.037.343-.01.512-.11zm.405 3.746v.076-.076zm-23.947 1.834v.074-.074zm9.58-.243c.208.083.311.167.54.167.229 0 .332-.084.54-.167.166.417.416.604.665.959h-.582c-.228-.105-.353-.334-.623-.334-.125 0-.208.083-.332.083-.084.459-.146.792-.54 1.042v-.791a4.04 4.04 0 00.332-.96zM3.17 31.43h-.11l-.073.074c.078.114.15.22.293.22.105 0 .164-.064.256-.11-.082-.133-.224-.115-.366-.184z"
-					fill="#fff"
-				/>
-				<path
-					d="M2.279 30.184a.336.336 0 00.25.111c.11 0 .165-.075.275-.075.229 0 .362.117.586.075 0 .033.036.042.036.075 0 .08-.073.108-.073.187 0 .099.11.127.11.225 0 .16-.201.136-.33.225a2.332 2.332 0 01-.197.124H4.6c-.043-.087-.112-.159-.112-.274 0-.23.17-.314.256-.524-.26 0-.398-.099-.659-.113-.131-.006-.299.007-.406-.036h-1.4zm13.299 0c.006.037.01.076.01.12 0 .149-.115.21-.183.336h-.44c-.324 0-.526-.06-.805-.224v-.113h.769c.126 0 .239-.044.327-.12l.322.001zm10.348 0c.163.208.253.478.301.798l-.79.085c-.03-.359-.05-.591-.033-.883h.522zm8.702 0c0 .003-.002.006-.003.008h-.146a.4.4 0 01-.003-.007h.152zm11.9 0c-.186.012-.334.066-.334.25 0 .37.485.351.842.394.183.023.288-.005.476-.005-.041.124-.08.203-.12.308h-1.19a.931.931 0 00-.264-.041c-.166 0-.282 0-.42.042h-.115a.186.186 0 00-.05-.042v-.525h.475c.055-.07.11-.112.11-.206 0-.06-.01-.119-.03-.175h.62zm.97 0l.087.232c-.197-.16-.352-.225-.604-.225-.065 0-.136-.004-.208-.006l.725-.001zm11.12 0a.558.558 0 01-.233.007c0 .033-.036.042-.036.075 0 .08.073.107.073.187 0 .098-.11.126-.11.225 0 .159.201.136.33.224.146.104.276.171.408.23h-1.908c.017-.146.145-.222.145-.38 0-.229-.17-.313-.256-.524.118 0 .21-.02.301-.043l1.285-.001zm.7 0a.394.394 0 01-.073.008.233.233 0 01-.056-.007h.13zm.202.419c.014.056.028.093.037.15-.087.032-.138.065-.22.112-.06-.042-.092-.07-.146-.112.114-.075.2-.113.33-.15zm-57.264.104c-.014.056-.027.094-.037.15.087.033.138.066.22.112.06-.042.092-.07.147-.112a1.109 1.109 0 00-.33-.15z"
-					fill="#fff"
-				/>
-				<path
-					d="M4.167 29.255l.028.01c.133.052.33.038.33.188 0 .182-.261.15-.44.15-.43 0-.672-.094-1.098-.15-.032.098-.11.154-.11.262 0 .065.046.093.073.15l-.073.074c-.174-.177-.174-.486-.42-.486-.166 0-.349.074-.349.243 0 .185.056.386.17.507H3.68c-.064-.025-.106-.07-.106-.151 0-.08.041-.117.073-.187h.55c.123 0 .168-.113.292-.113h.403c-.052-.19-.08-.34-.095-.496l-.629-.001zm21.654 0c.04.071.09.137.114.21h-.499v-.21h.385zm21.911 0a.314.314 0 000 .001zm9.558 0a.113.113 0 00-.042.093c0 .183.261.15.44.15.43 0 .672-.094 1.098-.15.032.099.11.155.11.262 0 .066-.046.094-.074.15l.074.075c.174-.178.174-.487.42-.487.165 0 .349.075.349.244 0 .264-.114.56-.349.61h-.13c-.075-.018-.128-.067-.217-.067-.14 0-.243.043-.354.068H57.33c.11-.028.215-.06.358-.068.21-.01.512.028.512-.188 0-.08-.041-.117-.073-.187h-.55c-.123 0-.169-.112-.292-.112h-.403c.04-.148.066-.271.083-.393h.325zm-42.035.948a.572.572 0 00.185-.33c.063.121.116.213.137.33h-.322zm10.148 0c.006-.106.017-.22.033-.35.21.08.37.199.489.35h-.522zm9.072 0c-.036-.094-.07-.15-.07-.256 0-.294.352-.374.64-.374.137 0 .311.047.311.187 0 .34-.595.139-.73.443h-.15zm11.432 0c-.096-.284-.412-.484-.703-.405 0-.108-.005-.183-.073-.262.279-.038.43-.15.714-.15.357 0 .32.636.677.636h.879c.064 0 .087-.06.146-.075.01 0-.11.025-.146 0l.096.256h-.726a2.682 2.682 0 00-.245 0h-.619z"
-					fill="#fff"
-				/>
-				<path
-					d="M2.147 28.325l-.075.06v.261c.068.01.11.038.183.038.137 0 .146-.197.146-.337a.05.05 0 00-.005-.022h-.249zm.504 0c-.019.087-.03.18-.03.284 0 .164.192.183.33.262.43.252.76.226 1.216.401h.629a4.205 4.205 0 01-.015-.401c-.06.112-.128.225-.256.225h-.513c-.293 0-.293-.412-.293-.712 0-.022.004-.041.011-.06l-1.079.001zm12.79 0v.18c0 .07.073.098.073.168 0 .094-.064.126-.11.206-.032-.145.005-.248-.073-.375-.059-.098-.189-.12-.267-.18l.377.001zm10.53 0a1.558 1.558 0 00-.202.775c0 .064.023.12.053.173h-.385v-.938l.076-.01h.458zm9.056 0c.011.01.023.02.037.03v.13h-.842v-.16h.805zm11.71 0a.561.561 0 00.371.142c.184 0 .31-.058.444-.142h.179v.48c0 .05.073.06.073.111 0 .144-.067.219-.073.357-.015-.33-.436-.394-.77-.394-.123 0-.187.075-.31.075-.096 0-.142-.075-.238-.075h-.44c-.096 0-.141.075-.238.075-.15 0-.206-.14-.347-.187-.16-.052-.33-.033-.403-.188.226-.007.485-.071.562-.254h1.19zm12.404 0c.007.056.011.115.011.18 0 .163-.192.182-.329.262-.44.257-.773.224-1.245.412-.095.036-.222.04-.287.093h-.326c.022-.155.027-.308.027-.505.06.112.129.224.257.224h.512c.279 0 .292-.372.293-.666h1.087zm.56 0v.217c-.068.01-.11.037-.182.037-.112 0-.139-.13-.145-.253l.327-.001zm-23.393.137v.08c-.082.084-.132.145-.22.225-.086-.164-.228-.178-.365-.3.22-.051.375-.094.585-.005z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.972 27.396c.01.138.018.285.027.445l.11.112h.475c.083-.145.115-.257.183-.412a.272.272 0 00.074-.037l.073.075c-.083.302-.207.505-.263.765h1.08c.034-.087.142-.128.244-.128h.367c.178 0 .265-.127.439-.15v-.187c-.247.103-.39.224-.66.224-.393 0-.95.066-.95-.337 0-.098.054-.15.109-.225.082.043.142.075.238.075.2 0 .327-.116.463-.22H1.972zm13.462 0c.004.048.006.098.006.154v.794h-.376a.141.141 0 01-.063-.12c0-.3.403-.356.403-.656 0-.064-.01-.12-.03-.172h.06zm10.43 0c-.144.182-.318.318-.513.487a3.086 3.086 0 01-.04-.488l.553.001zm10.59 0v.154c-.11-.075-.178-.15-.311-.15-.37 0-.453.487-.824.487h-.476c-.22 0-.338.056-.549.112v-.187h.293c.124-.09.293-.122.293-.281a.25.25 0 00-.036-.134l1.61-.001zm9.955 0c.174.147.442.234.442.49 0 .164-.155.216-.256.338.05.043.096.083.142.12h-1.19a.305.305 0 00.023-.12c0-.305-.375-.281-.622-.45-.14-.094-.172-.227-.2-.379l1.662.001zm13.385 0l-.02.34-.11.113h-.475c-.082-.145-.115-.257-.183-.412a.274.274 0 01-.073-.037l-.074.075c.094.34.24.554.281.868h-1.086V28.3c0-.127-.133-.188-.256-.188h-.366c-.179 0-.266-.126-.44-.15v-.187c.247.103.39.225.66.225.393 0 .951.065.951-.337 0-.098-.055-.15-.11-.225-.082.042-.142.075-.238.075a.551.551 0 01-.326-.116h1.865zm-57.399.95c-.009-.02-.028-.034-.031-.055h-.146c-.028.02-.05.037-.071.054h.248zm23.117-.002c.29-.036.49-.054.796-.054l.084-.66c-.146.278-.302.49-.422.714h-.458zm8.709.001v-.084h.732c.023.034.045.06.073.085h-.805zm13.33 0l.178-.121v.12h-.178zm11.822-.001a1.679 1.679 0 01-.002-.083c0-.033.032-.047.037-.075h.146c.06.042.092.07.146.113v.045h-.327zM46.21 28.11c.123 0 .201.056.274.15-.105.019-.164.075-.274.075-.083 0-.124-.052-.202-.075v-.112c.074-.014.124-.038.202-.038z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.91 26.467c.016.313.04.605.062.948h2.01c.107-.081.22-.154.378-.154.165 0 .274.009.421.074 0-.191.064-.294.073-.486-.068.032-.105.075-.183.075h-.476c-.279-.225-.805-.085-.805-.45v-.007h-.58a.184.184 0 01-.152.082.544.544 0 01-.308-.082h-.44zm33.918 0c-.11.079-.223.165-.362.24h-.146a.63.63 0 00-.096-.24h.604zm.927 0l-.005.016a.343.343 0 00-.004-.016h.01zm10.502 0c-.13.025-.258.056-.258.184 0 .21.251.28.457.28.092 0 .124-.07.202-.112h.146v.3c-.17-.047-.265-.122-.44-.15-.265-.042-.425-.037-.695-.037-.174 0-.403.028-.403.206 0 .122.062.206.144.276h-1.66c-.015-.082-.029-.169-.058-.258.343 0 .842.07.842-.28 0-.136-.142-.207-.274-.207-.092 0-.128.075-.22.075-.165 0-.256-.075-.421-.075a.747.747 0 00-.029-.203l2.667.001zm12.601 0c-.017.306-.04.599-.062.947H57.93c-.157-.11-.291-.258-.515-.258a.911.911 0 00-.421.075c0-.192-.064-.295-.074-.487.069.033.105.075.183.075h.476c.251-.202.701-.109.79-.353l1.49.001zm-44.483.948c-.045-.122-.14-.214-.263-.296a.86.86 0 01.22-.187c.05.19.089.317.103.483h-.06zm9.936 0v-.193c.145-.021.228-.128.374-.128.124 0 .207.085.332.085a1.492 1.492 0 01-.152.235h-.553zm9.535 0c-.082-.134-.294-.173-.477-.221v-.225c.252.019.389 0 .64 0 .261 0 .27.412.532.412.164 0 .2-.225.366-.225.169 0 .233.188.402.188.064 0 .092-.043.147-.075v.146h-1.61zm23.612-.633l.073.15a.307.307 0 01-.146 0 .756.756 0 01.073-.15zm-21.815-.188a1.84 1.84 0 00-.114.263c.005-.038 0-.07 0-.113 0-.075.071-.093.114-.15zm-33.325.292l-.073.15a.31.31 0 00.146 0 .756.756 0 00-.073-.15z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.895 25.538a1.052 1.052 0 00-.006.112c0 .317.008.583.021.835h.44a.213.213 0 01-.095-.18c0-.14.137-.224.274-.224.165 0 .312.131.312.3a.189.189 0 01-.032.104h.58c.007-.222.296-.217.513-.254.385-.066.563-.3.952-.3v-.304c-.325.042-.512 0-.842-.033-.002-.023-.011-.04-.02-.056H1.895zm23.76 0a.155.155 0 01.037.101c0 .164-.192.188-.33.263h-.146v-.337a.705.705 0 00.036-.026l.404-.001zm11.324 0a.195.195 0 00-.011.064v.375c0 .214-.14.313-.214.509h-.009c-.032-.143-.144-.247-.29-.247-.267 0-.44.113-.626.246h-.604c-.164-.243-.485-.328-.783-.508.316 0 .494.009.805-.038.17-.023.247-.15.421-.15.275 0 .385.262.66.262.27 0 .603-.06.603-.337a.252.252 0 00-.058-.175l.106-.001zm9.433 0c.019.031.036.064.074.064.098 0 .172-.03.244-.064h.448c.076.089.13.214.26.214h.44v.187c-.257.023-.435.089-.586.3l.073.075v.15c-.034.008-.07.014-.107.021H44.59c-.042-.154-.117-.28-.117-.471 0-.084.082-.098.146-.15h.183c.005.033.037.056.037.094 0 .07-.046.108-.073.168.137.103.233.188.402.188.174 0 .367-.085.367-.263 0-.238-.32-.309-.55-.262a.989.989 0 00-.097-.251h1.524zm13.473 0v.008c0 .36-.01.656-.027.94h-1.49a.287.287 0 00.016-.098c0-.229-.292-.224-.512-.262-.384-.065-.563-.3-.952-.3v-.288h2.965zm-.64.439c.137 0 .274.084.274.224 0 .192-.215.262-.402.262a.186.186 0 01-.183-.187c0-.168.146-.3.31-.3zm-13.051-.263c.082 0 .119.061.183.113v.224c-.06.01-.1.038-.165.038-.11 0-.146-.09-.238-.15.046-.112.1-.225.22-.225z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.946 24.608a.256.256 0 00-.057.162c0 .126.073.192.073.318 0 .19-.05.305-.067.469h2.097a.115.115 0 01-.017-.057c0-.243.307-.225.513-.337a.53.53 0 00-.146-.187c-.247.173-.43.262-.733.262-.137 0-.292-.066-.292-.206 0-.234.31-.178.512-.28l.27-.144h-.826l-.066.068a1.662 1.662 0 00-.142-.068H1.946zm32.781 0a.256.256 0 00.049 0h.618v.001c0 .127.114.225.238.225.128 0 .2-.075.33-.075.123 0 .187.075.31.075.275 0 .696-.393.696-.112v.412c0 .098.073.145.073.243 0 .075-.042.12-.062.18h-.106c-.16-.178-.614-.066-.93-.161-.554-.169-.847-.407-1.428-.45.004-.136.039-.215.066-.338h.146zm12.304 0c.004.044.005.085-.033.114-.183.14-.549.004-.549.243 0 .234.325.216.403.43a1.05 1.05 0 00-.476.113c.015.012.025.03.036.048h-1.525c-.06-.117-.122-.225-.122-.385 0-.164.096-.337.256-.337h.623c.153 0 .295-.09.346-.226h1.041zm12.847 0a.262.262 0 01.007.058c0 .126-.074.191-.074.318 0 .226.072.348.074.572h-2.966v-.015c.325.042.513 0 .842-.033.005-.047.037-.065.037-.112 0-.244-.307-.225-.513-.337a.53.53 0 01.147-.188c.247.174.43.263.732.263.137 0 .293-.066.293-.206 0-.234-.311-.178-.513-.281l-.072-.038 2.006-.001zm-34.627.95c.058-.044.103-.087.184-.087.084 0 .172.029.22.087h-.404zm21.477 0c.087-.041.17-.087.288-.087.071 0 .12.038.161.087h-.449z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.779 23.68v.453c.073.099.133.188.256.188h.806c.206 0 .33.126.439.3l-.007.006H4.1c.43-.23.735-.405 1.194-.718v-.15a1.031 1.031 0 01-.177-.08l-3.337.001zm23.965 0c.136.041.267.087.441.087.085 0 .141-.04.203-.087h.043c.043.184.074.37-.008.574-.013.042.005-.127-.036-.113-.156.052-.22-.187-.385-.187-.293 0-.402.3-.695.3-.179 0-.311-.173-.311-.356 0-.099.055-.171.135-.218h.613zm9.172 0c.063.104.132.176.22.274h.403v-.274h.666a.343.343 0 00-.007.068c0 .201.243.244.44.244.1 0 .182-.085.182-.188 0-.05-.015-.09-.039-.125l.253.001c.004.044.006.092.006.143 0 .126-.05.192-.073.318h-.55c-.2 0-.31-.112-.512-.112-.301 0-.511.29-.512.597h-.618c.114-.008.215-.085.215-.204 0-.22-.096-.468-.311-.468-.046 0-.05.056-.092.075.039-.126.07-.239.034-.35l.295.001zm10.887 0c.237.113.474.187.792.274.252.07.622-.005.622.262 0 .136-.142.173-.183.3-.012.036-.007.074-.004.11h-1.04a.318.318 0 00.02-.11c0-.061-.037-.09-.037-.15-.261-.023-.399-.112-.66-.112h-.512c-.206 0-.348-.117-.44-.3.162-.117.435-.097.472-.274h.97zm12.764 0v.087c-.06 0-.087.037-.147.037-.064 0-.087-.056-.146-.075v-.049H56.48v.124c.526.36.85.538 1.392.824l2.006-.001c-.03-.133-.164-.224-.305-.224-.398 0-.659 0-1.006.188l-.074-.075c.11-.173.234-.3.44-.3h.805c.124 0 .183-.089.256-.187v-.349h-1.427zm-55.503.948c-.282-.12-.523-.12-.865-.12-.1 0-.195.046-.254.12h1.119zm31.516 0a.905.905 0 00.007-.037c.051.013.089.032.14.037h-.147zm22.086-.45c-.005.048-.02.072-.038.115a.235.235 0 01.038-.114zm-53.46-.475v.168c.06 0 .088.038.147.038.064 0 .087-.057.146-.075v-.131c-.114.004-.178 0-.292 0zm1.906.58c.005.048.019.072.038.114a.236.236 0 00-.038-.114z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.879 22.751c.13.128.313.165.522.165.202 0 .311-.112.513-.112.178 0 .297.084.403.224h.512c.114 0 .174.08.256.15v.225c-.04.01-.064.037-.11.037-.324 0-.49-.15-.805-.224-.15-.033-.252-.019-.403 0-.384.047-.6.13-.988.15v.333h3.337c-.155-.086-.278-.184-.482-.184-.155 0-.228-.107-.329-.224.087-.108.233-.099.293-.225.048-.1.013-.224.036-.315H1.879zm56.766 0a.599.599 0 00-.188.173h-.513c-.114 0-.173.08-.256.15v.225c.041.009.064.037.11.037.325 0 .49-.15.805-.225.151-.032.252-.018.403 0 .384.047.6.132.989.15v.438h-1.428v-.082c-.115 0-.179.005-.293 0v.082H56.48v-.026c.256-.084.384-.262.659-.262.156 0 .229-.108.33-.225-.087-.107-.234-.098-.293-.224-.031-.066-.028-.141-.028-.211h1.497zm1.095 0a.928.928 0 01-.368.062c-.106 0-.187-.031-.268-.061h.636zm-34.608.95a.498.498 0 01.249-.065c.14 0 .254.03.364.064h-.613zm1.257-.002l.035-.026v-.45c-.069-.037-.105-.074-.183-.074-.421 0-.64.187-1.062.187-.142 0-.233-.051-.366-.005v-.182c.229-.155.385-.262.66-.262.37 0 .576.056.946.075-.121.26-.044.495.013.736h-.043zm8.232.001a.34.34 0 00-.028-.064h.288c.011.023.023.044.035.064h-.295zm.919-.001v-.063c-.298-.22-.522-.468-.879-.375.174-.21.385-.262.66-.262.09 0 .127.075.219.075h.659c.206 0 .306.15.512.15.115 0 .21-.047.256-.15 0 .248.05.407.068.624h-.253c-.068-.099-.206-.152-.254-.287h-.22c-.022.116-.083.185-.103.287h-.665zm9.294 0a.215.215 0 00.005-.045c0-.257-.334-.28-.586-.28v-.15h.293c.27.145.517.014.769.187.177.123.333.213.489.287h-.97z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.656 21.821l.013.14c.014.182.183.299.366.299.142 0 .206-.113.348-.113.124 0 .178.075.275.15-.188.323-.577.131-.88.337a.552.552 0 00.1.135h2.756a.14.14 0 01.074-.097c.15-.08.293.018.44-.075.187-.122.25-.337.475-.337.091 0 .151.065.183.15 0-.225.096-.342.146-.562-.068.032-.105.075-.183.075a.805.805 0 01-.396-.103l-1.944.001a.153.153 0 01.034.102c0 .121-.1.224-.22.224-.17 0-.222-.167-.29-.326H1.656zm23.756 0a.197.197 0 00-.013.072c0 .127.147.14.256.187.252.108.385.263.66.263.045 0 .064-.038.11-.038v.262c-.088-.056-.152-.112-.257-.112h-.55c-.104 0-.15-.075-.255-.075h-.403c-.265 0-.402-.288-.403-.559h.855zm11.702 0c.037.083.074.164.074.278 0 .098-.055.15-.074.244a.272.272 0 00-.073.037c-.091-.22-.256-.374-.494-.374-.453 0-.59.524-1.043.524-.302 0-.481-.094-.77-.005v-.482c.074.019.124.037.202.037.11 0 .238-.037.238-.15a.174.174 0 00-.037-.11l1.977.001zm19.806 0v.147c-.064.024-.096.075-.165.075-.091 0-.123-.07-.2-.112.018-.043.033-.075.046-.11h-.758c.052.173.125.288.125.484.032-.084.092-.15.183-.15.225 0 .289.216.476.337.147.094.289-.004.44.075.077.041.082.117.082.201h1.497a.443.443 0 01.214-.05c.096 0 .17.024.244.05h.636a.518.518 0 00.255-.238c-.302-.206-.69-.014-.878-.337.096-.075.15-.15.274-.15.142 0 .206.112.348.112.183 0 .352-.117.366-.3l.003-.035-1.33.001c-.053.12-.114.222-.247.222a.225.225 0 01-.22-.222H56.92zm.953.56l.076.037v.187h-.114c.01-.084.02-.14.038-.225zm-52.726-.533h-.292v.225c.064.023.096.074.164.074.092 0 .124-.07.202-.112a1.31 1.31 0 01-.074-.187zm-1.242.636l-.075.038v.187h.113a1.858 1.858 0 00-.038-.225z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.59 20.892c.006.312.032.59.066.948h1.297a1.331 1.331 0 00-.04-.086l.074-.075h.11c.11.068.264.081.332.162l1.944-.001a.484.484 0 01-.263-.423c0-.164.17-.262.33-.262.307 0 .718.192.622.487.083-.136.078-.253.183-.375v-.262c-.082.033-.128.075-.22.075a.767.767 0 01-.52-.188H2.816a.285.285 0 01.024.113c0 .225-.147.45-.367.45h-.366c-.04-.085-.11-.127-.11-.225 0-.142.064-.256.162-.338h-.57zm58.116 0a.417.417 0 01.069.233c0 .099-.069.14-.11.225H59.3c-.22 0-.366-.225-.366-.45v-.008h-2.814a.85.85 0 01-.371.084c-.092 0-.137-.043-.22-.075v.262c.105.122.101.239.183.374-.096-.294.316-.486.623-.486.16 0 .33.098.33.262 0 .332-.335.524-.66.524-.078 0-.114-.042-.183-.075.006.028.013.054.02.078h.76l.026-.078h.292v.078h1.392v-.003c0-.178.22-.173.366-.262h.11l.073.075c-.031.06-.056.127-.083.19h1.33c.035-.369.063-.648.074-.948h-.475zm-35.15.948v-.003c0-.206.165-.374.366-.374h.403c.115 0 .17.08.256.15-.054.093-.138.141-.17.226h-.855zm10.58 0c-.065-.085-.207-.115-.329-.115.028-.127.064-.225.005-.337h.251c.495.3.833.45 1.41.45.233 0 .522.008.567-.226.013.09.043.16.073.227h-1.977zm22.753-.752c.142 0 .238.136.238.28 0 .141-.155.207-.293.207-.105 0-.15-.075-.256-.075-.142 0-.206.112-.348.112-.146 0-.215-.098-.348-.15l.037-.337c.137.132.252.188.44.188.224 0 .306-.225.53-.225zm-54.006.104c-.142 0-.238.136-.238.281 0 .14.156.206.293.206.105 0 .151-.075.257-.075.141 0 .205.113.347.113.147 0 .215-.099.348-.15l-.037-.337c-.137.13-.251.187-.439.187-.224 0-.306-.225-.53-.225z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.634 19.963c.02.053.035.11.035.18 0 .085-.073.122-.073.207-.009.21-.01.39-.007.561h.571a.698.698 0 01.443-.15c.098 0 .177.063.215.15h2.686a.43.43 0 01-.137-.318c0-.126.146-.154.146-.28 0-.162-.213-.192-.22-.35H1.635zm58.477 0a.46.46 0 00-.006.076c0 .085.073.122.073.206.011.253.011.464.004.666h-.475c-.106-.164-.321-.254-.536-.254-.134 0-.233.117-.237.254H56.12c.166-.083.287-.226.287-.422 0-.127-.146-.155-.146-.281 0-.119.115-.166.179-.245h3.671zm-1.873.02h.366c.005.056.037.09.037.15 0 .098-.106.15-.202.15-.123 0-.174-.084-.274-.15l.073-.15zm-54.701.105H3.17c-.005.056-.037.088-.037.15 0 .098.105.15.202.15.123 0 .173-.085.274-.15l-.073-.15z"
-					fill="#fff"
-				/>
-				<path
-					d="M3.127 19.034c-.107.573-1.568-.003-1.568.66 0 .117.043.198.075.287h3.66v-.006c0-.248.292-.253.512-.356.22-.103.275-.341.513-.374.251-.033.407.004.659.037.024-.1.057-.17.092-.248h-.124c-.151.068-.32.061-.518.061-.108 0-.181-.039-.27-.06l-.47-.001a.962.962 0 00-.212.061c-.251.103-.274.45-.549.45-.201 0-.33-.207-.33-.413 0-.037.013-.07.033-.098H3.127zm53.246 0c.184.133.23.406.473.406.2 0 .328-.202.33-.407l1.512.001c.264.393 1.526-.06 1.526.556 0 .16-.08.251-.103.391h-3.672a.172.172 0 00.041-.11c0-.248-.293-.253-.512-.356-.22-.103-.275-.342-.513-.375-.251-.032-.407.005-.659.038a.982.982 0 00-.047-.144h1.624zm1.663.219c.183 0 .421.037.421.225 0 .238-.316.262-.549.262-.21 0-.325-.099-.512-.188v-.112c.233-.108.38-.187.64-.187zm1.263.3h.403v.224c-.069.038-.12.075-.201.075-.129 0-.193-.075-.312-.112.037-.07.074-.117.11-.188zm-55.562-.196c-.183 0-.42.038-.42.225 0 .239.315.262.548.262.21 0 .325-.098.513-.187v-.113c-.234-.107-.38-.187-.64-.187zm-1.263.3h-.403v.225c.069.037.12.074.202.074.128 0 .192-.074.311-.112-.037-.07-.073-.117-.11-.187z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.451 18.105c.02.22.21.323.401.447.142.094.179.262.348.262.229 0 .284-.3.512-.3.243 0 .422.22.422.469a.385.385 0 01-.007.07H4.63a.236.236 0 01.188-.09h.402c.124-.172.366-.196.366-.411 0-.2-.18-.262-.255-.447h-.809c-.085.015-.14.072-.235.072a.247.247 0 01-.173-.072h-.358a.78.78 0 01.073.11c-.146.08-.233.187-.403.187a.259.259 0 01-.256-.262c0-.013.003-.024.006-.035h-.723a.357.357 0 01-.216.072.266.266 0 01-.188-.071l-.598-.001zm4.857 0c-.228.03-.336.28-.392.522.302 0 .471-.075.769-.038.206.024.393.07.475.263a3.03 3.03 0 00.33-.712 2.482 2.482 0 01-.494.037c-.22 0-.353-.057-.551-.071l-.137-.001zm7.27 0c-.228.142-.423.268-.736.268-.113 0-.09-.15-.119-.267l.856-.001zm33.087 0a6.212 6.212 0 01-.839.043h-.402c-.147 0-.22.112-.366.112h-.476c-.293 0-.458-.112-.75-.112-.225 0-.344.075-.568.075-.098 0-.152-.058-.22-.119l3.62.001zm9 0c.098.102.156.261.192.418-.302 0-.471-.075-.769-.038-.205.023-.393.07-.475.262a2.9 2.9 0 01-.306-.642h1.358zm2.929 0a.257.257 0 01-.247.193c-.17 0-.256-.108-.403-.187l.004-.007-1.565.001c-.083.113-.196.185-.196.343 0 .215.242.238.366.412h.403c.11 0 .22.074.22.187v.005h1.512a.303.303 0 01-.048-.174c0-.248.178-.468.42-.468.23 0 .284.3.513.3.17 0 .206-.169.348-.262.157-.103.314-.19.376-.343h-1.703zm-52.436.949a.494.494 0 00-.114-.014c-.14 0-.247-.004-.357.014h.47zm.912-.001c.018-.038.036-.079.054-.126a.577.577 0 01-.178.126h.124zm47.68 0c-.03-.074-.066-.14-.1-.23.197.2.416.186.695.186.151 0 .234-.075.385-.075.224 0 .361-.009.567.075a.365.365 0 01.076.043H54.75zm2.81-.83c.092 0 .128.112.128.206 0 .126-.077.243-.201.243-.1 0-.201-.065-.201-.168 0-.155.123-.281.274-.281zm-53.347.104c-.091 0-.128.113-.128.206 0 .127.078.244.201.244.101 0 .202-.066.202-.169a.279.279 0 00-.275-.28z"
-					fill="#fff"
-				/>
-				<path
-					d="M2.47 17.524c.065.092.114.19.114.328a.33.33 0 01-.131.264h.723c.01-.026.027-.047.03-.077h.403c.064 0 .11.032.147.077h.358a.214.214 0 01-.065-.152c0-.206.131-.354.31-.44h-1.89zm3.025 0c.134.12.216.328.42.328.143 0 .147-.174.257-.262a.675.675 0 01.095-.065l-.772-.001zm2.11 0a.583.583 0 01.141.14c.042-.028.069-.051.11-.074v-.066h-.251zm5.17 0c.322.287.698.377 1.128.26v.15a3.155 3.155 0 00-.324.183h-.856a.24.24 0 00-.028-.07c-.132-.206-.389-.117-.549-.3-.058-.067-.064-.143-.067-.223h.696zm31.78 0c.111.207.404.298.668.298.124 0 .202-.117.202-.243a.268.268 0 00-.006-.055h2.527c-.394.373-.806.529-1.28.593h-3.621a.562.562 0 00-.037-.033c.246-.42.64-.442 1.053-.56h.494zm9.505 0a.668.668 0 00-.03.037l-.053-.037h.083zm2.124 0c-.089.109-.173.223-.326.223-.123 0-.142-.13-.217-.223h.543zm3.057 0a.497.497 0 00-.051.223c0 .192.16.337.348.337.251 0 .32-.276.384-.524h.146c.151.112.257.22.257.412a.38.38 0 01-.027.145h-1.703a.26.26 0 00.01-.07c0-.047-.033-.07-.038-.112h-.402c-.108 0-.162.09-.216.18l-1.565.001a.478.478 0 00.06-.107h.769c.11 0 .165.075.274.075.124 0 .238-.098.238-.224a.443.443 0 00-.153-.336h1.669zm-57.192.592c-.106-.097-.152-.28-.197-.452h-.146c-.151.113-.256.22-.256.412 0 .014 0 .027.002.04h.597zm3.282.001v-.002h-.77a.234.234 0 00-.038.003h.808zm1.114 0a1.892 1.892 0 00-.09-.002.353.353 0 00-.047.003h.137zm47.863.001a6.346 6.346 0 01-.024-.07c.187.033.302.038.494.038.252 0 .389-.076.64-.076a.33.33 0 01.248.108h-1.358z"
-					fill="#fff"
-				/>
-				<path
-					d="M2.151 16.944c.023.063.04.135.067.208.062.163.17.268.252.384h1.888a.914.914 0 01.386-.084h.55a.29.29 0 01.2.084h.773c.124-.07.254-.084.418-.084h.622c.119 0 .215.027.298.084h.251v-.121c-.17-.024-.16-.234-.256-.375a.322.322 0 00-.102-.096h-2.54c-.08.057-.138.134-.25.134h-.732c-.045-.053-.099-.091-.142-.134h-.412a.22.22 0 01-.179.096.502.502 0 01-.294-.095l-.798-.001zm10.632 0a.798.798 0 00.315.067.79.79 0 00.3-.067H14a.505.505 0 01.123.142c-.357 0-.526.224-.879.262-.215.023-.398-.042-.549.112.026.027.053.053.08.077h-.696c-.003-.076-.003-.153-.043-.227-.114-.21-.361-.187-.604-.187-.082 0-.119.075-.201.075-.092 0-.138-.038-.22-.075a.613.613 0 01.12-.18l1.652.001zm35.021 0c.002.165.066.325.22.366v.15a2.625 2.625 0 01-.078.077h-2.527c-.014-.067-.05-.115-.068-.19h.403c.186 0 .187-.245.27-.403h1.78zm1.686 0a.34.34 0 00-.039.142h-.695c-.097 0-.14-.074-.197-.142h.93zm8.335 0a.696.696 0 00-.027.03h-.732a.224.224 0 01-.112-.03h-2.785c-.09.14-.085.343-.251.366v.188c.022.012.04.024.058.037l.08.001a.493.493 0 01.41-.188h.623c.206 0 .357.023.513.15a.261.261 0 01.039.039h.542c.08-.097.161-.19.297-.19h.55c.198 0 .406.065.542.189h1.669c.08-.164.235-.279.315-.488.013-.035.024-.07.034-.104h-1.765zm-13.764.592c.149-.043.3-.098.448-.188 0 .071.017.134.046.189h-.494z"
-					fill="#fff"
-				/>
-				<path
-					d="M1.334 16.247c-.01.163-.022.317.116.395.187.108.361 0 .549.112.084.05.123.12.152.203h.798c-.09-.059-.175-.137-.291-.203-.12-.065-.293-.065-.293-.206 0-.037.036-.056.036-.093.147-.02.234-.038.385-.038.306 0 .677.098.677.412a.225.225 0 01-.041.13l.412-.001a.22.22 0 01-.078-.166c0-.188.142-.272.293-.375.071-.048.121-.113.171-.17H1.334zm10.958 0a.613.613 0 01-.183.066c-.384.07-.613.075-1.006.075-.151 0-.238-.01-.385-.038a.8.8 0 01.064-.103h1.51zm3.621 0c0 .12.069.22.004.328-.105.173-.293.15-.495.15-.334 0-.416-.412-.75-.412-.252 0-.389.075-.64.075a.81.81 0 01-.44-.141h2.321zm30.49 0c-.338.14-.67.138-.906.478v-.478h.906zm2.6 0c-.021.073-.077.123-.138.178h-.256v-.178h-1.354a.494.494 0 01-.147.328c-.275.267-.586.145-.952.262a.228.228 0 00-.132.12h1.78v-.008c-.188-.037-.44-.028-.44-.224 0-.103.119-.15.22-.15.15 0 .256.145.256.3.238-.047.412-.113.622 0 .04.02.07.051.096.083h.931c.032-.066.072-.125.072-.214 0-.195.002-.34-.039-.497h-.52zm8.643 0c.023.023.049.046.079.066.15.103.292.187.292.374 0 .132-.105.183-.193.271h1.766c.038-.127.065-.238.185-.308.187-.112.361-.005.549-.112.108-.061.124-.169.12-.291h-2.798zm-52.69.71a.308.308 0 01.117-.053c.307-.066.609.01.769-.262-.183-.171-.309-.343-.52-.396l1.164.001c-.109.103-.278.17-.278.339 0 .379.554.21.915.28.14.028.271.03.374.092h-2.54zm6.175 0c.241-.244.65-.195 1.051-.195.25 0 .41.119.6.196h-1.65zm2.266.002c.103-.04.204-.084.34-.084.103 0 .19.03.263.085h-.603zm40.77-.001l.006-.009c.11-.159.284-.15.476-.187.361-.07.915.099.915-.28 0-.108-.069-.175-.147-.235h.82c-.103.078-.194.184-.307.29.16.272.462.197.769.263.112.023.17.11.254.157h-2.786zm4.82-.645c.151 0 .238.019.384.037 0 .038.037.056.037.094 0 .14-.174.14-.293.206-.224.126-.33.3-.586.3a.225.225 0 01-.22-.225c0-.314.371-.412.678-.412zm-42.742.374h.622c-.045.136-.15.225-.293.225-.233 0-.361-.084-.585-.15.1-.023.15-.075.256-.075z"
-					fill="#fff"
-				/>
-				<path
-					d="M5.564 15.318c-.55.056-1.029.163-1.698.294-.682.131-1.071.206-1.758.3-.306.042-.768-.094-.768.224 0 .043-.003.087-.006.13H4.22c.062-.071.125-.13.232-.13.146 0 .22.113.366.113h.366a.57.57 0 01.14.016h1.162a.25.25 0 00.089-.185c0-.192-.293-.201-.293-.393 0-.173.33-.126.33-.3 0-.028-.01-.05-.027-.068l-1.021-.001zm1.194 0v.107c.041.01.064.037.11.037.216 0 .36-.068.523-.144h-.633zm10.46 0c.041.048.082.1.127.152-.096.117-.197.187-.348.187-.256 0-.43-.121-.567-.337l.002-.002h-.723c-.052.033-.109.068-.158.115.21.056.476.093.476.318 0 .3-.403.281-.696.281-.183 0-.402-.056-.402-.243 0-.16.178-.183.256-.319a1.55 1.55 0 00-.33-.037c-.366 0-.567.15-.933.15-.403 0-.604-.225-1.007-.225-.334 0-.42.337-.732.45-.532.195-1.097.039-1.401.457h1.51c.206-.112.337-.308.605-.308.295 0 .481.177.695.308h2.321c0-.015.002-.03.004-.046.032-.187.197-.3.384-.3.238 0 .348.188.586.188.334 0 .714-.145.714-.487a.427.427 0 00-.13-.302h-.253zm29.897 0c-.022.15-.01.345-.08.49-.087.182-.238.224-.403.336-.075.052-.152.09-.228.122h-.907v-.197c0-.07-.073-.098-.073-.168 0-.09.115-.084.183-.131.173-.121.278-.263.315-.452h1.193zm.818 0a.517.517 0 01-.098 0h.098zm7.083 0v.003c-.041.01-.064.038-.11.038a.891.891 0 01-.27-.04h.38zm1.915 0c.297.052.608.117.977.19.682.13 1.071.206 1.757.3.307.041.77-.094.77.224 0 .078.009.159.01.234h-2.799c-.104-.106-.167-.234-.324-.234-.146 0-.22.112-.366.112h-.366a.555.555 0 00-.351.122h-.82c-.103-.08-.22-.15-.22-.29 0-.192.293-.202.293-.393 0-.161-.283-.133-.325-.265h1.764zm-9.675.948c.011-.117-.001-.245-.001-.384.389-.019.6-.187.988-.187.311 0 .476.126.77.225.168.056.365-.014.438.15.033.069.056.133.072.195h-.52a.234.234 0 00.01-.065c0-.088-.097-.13-.184-.13-.091 0-.137.037-.22.074v.122h-1.353zm-9.556-.87c-.037.27-.05.435-.073.71-.16-.257-.302-.383-.513-.598.156.037.243.075.403.075.1 0 .16-.09.183-.188z"
-					fill="#fff"
-				/>
-				<path
-					d="M14.606 14.388c.03.037.066.07.066.126 0 .14-.1.262-.238.262-.233 0-.334-.187-.567-.187h-.403c-.187 0-.288.112-.476.112-.389 0-.695-.004-.951-.3v-.013h2.569zm2.484 0a.18.18 0 01.072.089c.078.234-.037.44.11.636.064.088.138.154.2.223h-.254c-.095-.106-.193-.185-.349-.185-.186 0-.3.064-.436.184h-.724c.093-.06.171-.117.171-.24 0-.234-.237-.272-.402-.431-.09-.087-.162-.186-.233-.276h1.845zm9.448 0c-.363.16-.57.576-1.03.576h-.439c-.306 0-.43-.258-.732-.3-.164-.023-.27.023-.439.037.055-.107.06-.262.183-.262.133 0 .197.113.33.113.228 0 .361-.005.585-.038.101-.014.212-.05.264-.126h1.278zm10.092 0c.058.09.197.126.32.126.188 0 .31-.069.456-.125l.22-.001v.276c-.068-.042-.118-.075-.2-.075-.11 0-.165.075-.275.075h-.33c-.227 0-.347-.18-.516-.277l.325.001zm11.472 0c-.003.05-.005.102-.005.164-.183 0-.274.112-.457.112-.124 0-.188-.075-.312-.075-.096 0-.22.033-.22.131 0 .127.079.244.202.244.115 0 .179-.075.293-.075.288 0 .393.224.64.374-.12.033-.198.066-.313.073h-.095c-.204-.025-.306-.223-.525-.223-.135 0-.176.095-.195.222h-1.193a.856.856 0 00.015-.166c0-.191-.078-.294-.146-.468.22 0 .476-.075.476-.3a.48.48 0 000-.012l1.835-.001zm-41.517.949a.168.168 0 00-.12-.044c-.339 0-.628.016-.901.044h1.02zm.806-.001c.085-.04.175-.083.283-.118.15-.052.27.014.402-.075-.087 0-.133-.038-.22-.038-.434 0-.736-.018-1.098.225v.006h.633zm47.245 0c-.177-.054-.321-.152-.535-.223-.151-.051-.27.014-.403-.075.087 0 .133-.037.22-.037.435 0 .736-.019 1.098.225v.11h-.38zm.531 0a.121.121 0 01-.005-.035c0-.075.074-.113.147-.113a8.6 8.6 0 011.622.147h-1.764zm-38.774-.897c.101 0 .138.094.22.15-.1.07-.146.187-.274.187-.101 0-.152-.065-.239-.112v-.112c.11-.052.17-.113.293-.113zm.367.226a.232.232 0 01.113.038c-.042-.01-.07-.024-.113-.038z"
-					fill="#fff"
-				/>
-				<path
-					d="M26.244 13.459c-.127.026-.223.079-.223.212 0 .174.215.225.384.225.23 0 .531-.028.531-.262a.268.268 0 00-.07-.176l1.213.001.065.1v.112c-.375.015-.466.37-.695.675h-.623a.706.706 0 00-.288.06h-1.279a.173.173 0 00.03-.098c0-.23-.343-.117-.513-.262-.229-.197-.352-.45-.659-.45v-.137h2.127zm7.795 0v.1h-.257c-.036-.028-.068-.056-.11-.075a.759.759 0 00.015-.025h-1.228c.131.059.182.212.335.212h.805c.238 0 .298.244.513.338.48.21.791.337 1.318.337h.659c.084 0 .153.024.215.06h.326a.177.177 0 01-.029-.098c0-.182.234-.182.403-.225.188-.046.28-.187.476-.187.096 0 .133.075.22.113.045-.122.086-.188.146-.3-.252.117-.416.187-.696.187-.247 0-.549-.084-.549-.337 0-.038.01-.072.027-.1h-.574c-.128.34-.586.475-.99.475-.156 0-.234-.07-.366-.15.047-.12.105-.23.173-.325h-.832zm4.02 0c-.033.056-.071.102-.14.102-.106 0-.183-.05-.257-.102h.398zm-26.022.948v-.249c.073.033.123.057.201.075l2.142-.187h.549l2.013-.187c0 .033.037.042.037.075 0 .117-.147.145-.147.262 0 .13.157.144.258.211h-1.845c-.123-.156-.245-.286-.463-.286-.11 0-.22.075-.22.187 0 .042.02.072.044.1l-2.57-.001zm25.369.001c.066-.025.136-.049.22-.062v.062h-.22zm8.86-.001c-.01-.252-.297-.361-.549-.361h-.293c-.16 0-.242.112-.402.112-.316 0-.472-.187-.788-.187-.196 0-.494-.07-.494.131 0 .047.06.052.073.094-.169-.108-.237-.225-.366-.375v-.187c.101-.014.17-.037.275-.037 1.29 0 1.995.271 3.276.449.44.06.696.178 1.135.112a2.162 2.162 0 00-.032.248h-1.835zm-28.628-.623a.548.548 0 00-.073.187.453.453 0 01-.147-.15c.087 0 .138-.01.22-.037z"
-					fill="#fff"
-				/>
-				<path
-					d="M26.022 12.529c-.068.093-.135.188-.257.188-.22 0-.348-.037-.568-.037-.208 0-.448-.036-.59-.151h1.415zm1.8 0c.088.023.182.04.286.076a.419.419 0 01-.366.187c-.225 0-.289-.262-.513-.262h-.476l-.027-.001h1.096zm7.99 0a.228.228 0 01-.126.038.3.3 0 01-.15-.038h-3.251c.367.293 1.131-.037 1.131.45 0 .127-.133.187-.256.187-.147 0-.22-.112-.366-.112-.339 0-.604.066-.806.337.151.08.275.024.44.075a.24.24 0 01.03.012h1.229c.043-.078.078-.162.168-.162.083 0 .12.061.183.113v.05h.833c.21-.294.516-.462.906-.462.156 0 .312.103.312.262 0 .073-.013.139-.035.199h.574c.047-.08.147-.124.248-.124.1 0 .137.112.238.112.105 0 .15-.112.256-.112.122 0 .205.064.291.124h.398c.026-.04.048-.086.08-.124-.275.056-.66-.047-.66-.337 0-.169.147-.244.147-.412a.228.228 0 00-.012-.075l-1.802-.001zm-11.694.95v-.2h.402c.21 0 .302-.187.513-.187.105 0 .15.074.256.074h.366c.174 0 .247-.15.421-.15.261 0 .417.174.531.413-.114.022-.25.025-.363.049h-2.126zm2.748 0c-.038-.05-.083-.096-.113-.163.266-.037.43-.004.696-.037.522-.07.81-.201 1.318-.337v.224a.21.21 0 00-.073.113c-.243-.056-.435.023-.622.187l.007.012h-1.213zm9.992-.874h.33c.004.047.036.08.036.13 0 .146-.17.207-.311.207-.115 0-.179-.056-.275-.113.041-.112.124-.159.22-.224z"
-					fill="#fff"
-				/>
-				<path
-					d="M27.175 11.601c.126.268.095.527.31.76a.64.64 0 00.338.185l-1.096.001c-.178-.016-.262-.186-.45-.186-.12 0-.187.092-.255.185l-1.414.001a.323.323 0 01-.124-.261c0-.337.622-.244.622-.58a.172.172 0 00-.033-.105h2.102zm5.544 0c.011.007.024.01.039.01a.213.213 0 00.065-.01h3.978a.187.187 0 00-.09.16c0 .281.435.108.66.263.086.056.031.163.073.262.044.11.137.165.17.262l-1.802-.001a.201.201 0 00.094-.168c0-.159-.174-.243-.33-.243-.128 0-.187.117-.256.225.077.07.139.143.217.186h-3.252a.492.492 0 01-.04-.037c.22-.051.44-.159.44-.393 0-.168-.152-.234-.294-.318.05-.059.073-.141.11-.198h.218zm6.07 0c-.232.194-.37.46-.705.46-.137 0-.192-.15-.33-.15-.068 0-.104.042-.164.075h-.146a.527.527 0 00-.188-.385h1.533zm-.833.01h.183v.15a1.18 1.18 0 01-.183 0v-.15zm-4.576.113c.206 0 .256.215.44.3.173.08.439.028.439.224 0 .15-.165.225-.312.225-.196 0-.196-.28-.384-.337-.174-.052-.44.019-.44-.169 0-.14.12-.243.257-.243zm-8.054.412c.073 0 .096.065.146.112a.42.42 0 00-.11.113h-.146a.209.209 0 00-.073-.113c.064-.047.1-.112.183-.112zm3.516.187v.114-.114z"
-					fill="#fff"
-				/>
-				<path
-					d="M26.029 10.672a.275.275 0 01-.08.06c-.248.112-.66-.038-.66.243 0 .14.12.243.256.243a.637.637 0 00.55-.262h.402c0 .173-.096.346.037.45.183.14.462-.085.586.112.021.034.04.067.055.101h-2.102c-.107-.147-.443-.143-.443-.363 0-.174.183-.239.183-.412a.43.43 0 00-.033-.172h1.249zm1.347 0v.056c-.142-.014-.224.005-.366.005a.523.523 0 01-.24-.06h.606zm6.004 0c.185.096.34.239.586.284.297.056.48.038.787.038.286 0 .648-.048.675-.322h2.32a.152.152 0 00-.011.06c0 .21.215.336.42.336.211 0 .422-.14.422-.355l-.001-.042h2.344v.135c.077.024.1.113.183.113.11 0 .22-.075.22-.188 0-.025-.01-.041-.019-.059h.439c.013.098.019.203.019.322 0 .206-.129.412-.33.412-.293 0-.526.06-.732-.15-.156-.16-.252-.337-.476-.337-.215 0-.165.318-.33.449-.27.215-.562-.014-.878.112a.845.845 0 00-.23.14h-1.532a.325.325 0 00-.197-.065c-.09 0-.188.02-.257.065h-3.979c.065-.02.11-.065.19-.065h.477c.188 0 .403-.07.403-.262 0-.463-.61-.487-1.062-.487.022-.044.042-.093.067-.135h.482zm-.878.95c.025-.04.058-.066.11-.066.048 0 .07.044.107.065h-.217zm3.898-.89c.147 0 .22.112.311.224-.155.127-.315.098-.512.075v-.187c.064-.061.11-.113.201-.113z"
-					fill="#fff"
-				/>
-				<path
-					d="M19.768 9.743c-.118.08-.264.206-.264.09 0-.033.003-.063.006-.09h.258zm5.116 0a.361.361 0 01-.144.034.487.487 0 01-.185-.034h.329zm1.108 0c.103.134.245.22.431.22h.403c.115 0 .16.113.275.113.154 0 .19-.164.198-.333h.012c.019.177.064.314.064.52-.146-.056-.206-.187-.366-.187-.137 0-.256.103-.256.243 0 .258.375.239.622.206v.165h-.606c-.157-.077-.283-.202-.492-.202-.124 0-.169.118-.249.202H24.78c-.11-.253-.467-.325-.773-.427v-.187h.623c.274 0 .384.262.659.262.274 0 .38-.23.549-.45v-.145h.154zm10.45 0a.736.736 0 01-.134.07v.08c.137.033.224-.004.366-.004.215 0 .44.117.44.337 0 .191-.23.196-.403.262-.124.047-.174.187-.311.187-.179 0-.311-.154-.311-.337 0-.173.114-.258.182-.412-.123-.083-.29-.11-.39-.184H33.04a.582.582 0 01-.063.184c.055-.033.082-.075.146-.075.39 0 .664.033.952.3.091.084.091.262.22.262.288 0 .407-.262.695-.262.256 0 .44.243.44.505l-.002.035h2.32c.039-.094.173-.118.28-.166.138-.06.143-.262.294-.262.188 0 .248.229.255.427h2.344v-.052h.366c.002.021.01.037.018.053h.44c-.035-.269-.12-.484-.275-.765-.03-.058-.04-.137-.072-.184h-.211a.251.251 0 01-.12.034.548.548 0 01-.195-.033h-4.43zm-3.544.95c.03-.053.07-.093.134-.093a.72.72 0 01.347.092h-.481zm6.723-.617c.193 0 .385.121.385.318 0 .192-.234.244-.421.244-.165 0-.348-.075-.348-.244 0-.197.192-.318.384-.318z"
-					fill="#fff"
-				/>
-				<path
-					d="M22.653 8.814v.044c-.302.014-.448.188-.75.188-.158 0-.229-.13-.329-.232H19.67c0 .127-.007.257-.019.419-.016.216-.115.336-.14.528h.257a.53.53 0 01.066-.041c.307-.155.55-.028.879-.113.462-.116.677-.411 1.153-.411.224 0 .32.159.494.3.256.21.508.186.842.186h.513c.45 0 .541-.59.822-.868h-.567c-.017.2-.201.344-.402.344-.223 0-.454-.124-.474-.344h-.441zm3.172 0c-.02.039-.06.068-.06.12 0 .295.067.618.227.827h-.154V9.72c-.307-.028-.48 0-.787 0-.067 0-.116.022-.167.042h-.328c-.084-.035-.146-.097-.146-.192 0-.183.133-.337.311-.337.243 0 .362.187.605.187.105 0 .22-.06.22-.168a.537.537 0 00-.231-.438h.51zm1.48 0v.046a.542.542 0 00-.01-.046h.01zm8.63 0c.027.03.044.07.044.12 0 .173-.179.262-.348.262-.156 0-.348-.047-.348-.206 0-.063.017-.126.047-.176h-1.563c.14.094.235.235.235.419 0 .239-.215.365-.44.412-.196.037-.315-.014-.512-.038a.952.952 0 01-.01.154h2.842a.203.203 0 01-.086-.172c0-.365.412-.506.769-.506.105 0 .11.155.11.262a.483.483 0 01-.233.417h4.43a.638.638 0 01-.28-.23c.215-.285.805-.084.805-.449 0-.047-.027-.07-.037-.112h-.622c-.023-.047-.073-.075-.073-.131 0-.01 0-.019.002-.027h-.36c.118.026.211.068.211.195 0 .192-.196.3-.384.3-.261 0-.343-.244-.568-.375-.106-.061-.242-.08-.347-.119h-1.246a.684.684 0 01.202.27h.732c.115 0 .16-.113.275-.113.124 0 .201.117.201.243 0 .342-.343.543-.677.543-.522 0-.705-.561-1.226-.561-.124 0-.188.074-.312.074-.178 0-.347-.117-.347-.3a.23.23 0 01.052-.157h-.938zm-8.633.947c.002-.052.002-.105.002-.154 0 .056.004.107.01.155h-.012zm13.885.002c.04-.02.08-.043.137-.043.034 0 .057.017.074.042h-.211zM34.33 9.27c.069.01.115.02.183.038a.346.346 0 01-.183.15.118.118 0 00-.073-.075.429.429 0 01.073-.113z"
-					fill="#fff"
-				/>
-				<path
-					d="M22.087 8.232c.179.154.295.413.566.413v.18h.44a.484.484 0 01.475-.517c.242 0 .403.24.403.487 0 .01 0 .02-.002.03h.567a.452.452 0 01.332-.142c.162 0 .323.051.447.143h.51a.103.103 0 00.013-.05c0-.051-.032-.079-.037-.13h-.403c-.105 0-.15-.113-.256-.113h-.988c-.032-.056-.074-.084-.074-.15 0-.057.016-.108.042-.15h-1.181a.34.34 0 01-.252.113.566.566 0 01-.336-.114h-.266zm13.332.001a.248.248 0 01-.051 0h.051zm1.621 0v.15c-.073 0-.11.038-.183.038-.1 0-.146-.085-.22-.15a.414.414 0 00.034-.04l-.84.001.001.001c-.128.103-.27.061-.402.15-.124.084-.142.262-.293.262-.238 0-.348-.187-.586-.187h-.293c-.105 0-.155.066-.256.075-.28.028-.449-.005-.732-.005v.155c.177 0 .359.048.497.142h1.563a.2.2 0 01.173-.105c.152 0 .346.011.432.105h.939c.135-.156.503-.105.77-.105.13 0 .24.038.329.106h1.246c-.095-.035-.165-.086-.165-.2 0-.13.165-.168.293-.168.279 0 .398.201.659.3.092.035.204.046.3.067h.36c.04-.281.731-.034.731-.348 0-.108-.075-.17-.129-.245h-1.72a.742.742 0 01-.238.039.41.41 0 01-.176-.038h-.526c-.046.188-.342.188-.56.188-.192 0-.443-.005-.458-.189h-.55zm-17.37.592a7.302 7.302 0 00-.02-.517h.513c.063 0 .119-.03.158-.076h.5a.322.322 0 00-.109.245c0 .29.426.168.696.243.069.02.12.06.166.105h-1.905zm7.622 0c-.021-.08-.057-.142-.1-.217.069-.07.12-.112.183-.187-.05.153-.07.253-.072.404h-.01zm-.667-.292c.091 0 .15.01.238.038-.1.112-.197.187-.348.187-.05 0-.078-.033-.128-.038 0-.037-.037-.056-.037-.093 0-.117.16-.094.275-.094z"
-					fill="#fff"
-				/>
-				<path
-					d="M22.889 7.652c.1.067.167.162.167.293a.443.443 0 01-.115.299h1.181c.075-.122.24-.186.398-.186h.952c.036-.148 0-.25 0-.406h-.82c-.107.17-.328.219-.553.219-.179 0-.4-.047-.42-.218h-.79zm12.522 0a.13.13 0 01.02.069c0 .089-.079.15-.166.15a.166.166 0 01-.164-.169c0-.018.004-.034.01-.05h-.48c-.123.231-.59.138-.922.219.298.182.531.187.879.187h.476c.137 0 .178.164.303.185h.05c.11-.016.15-.148.27-.148.08 0 .1.083.144.148h.84c.054-.07.095-.148.187-.148.082 0 .114.075.183.113v.036h.55a.21.21 0 01-.001-.018c0-.248.307-.318.55-.318.214 0 .475.06.475.28 0 .02-.002.039-.006.055h.525c-.085-.037-.157-.098-.263-.148.06-.239.353-.173.586-.224.196-.046.297-.189.478-.22l-2.236.001c-.24.115-.477.219-.805.219a.582.582 0 01-.467-.219H35.41zm5.038 0c.027.048.058.08.106.106-.06.244-.306.304-.549.337-.186.026-.31.102-.459.15l1.721-.001a.244.244 0 01-.054-.149v-.443h-.765zm-20.127.592a.261.261 0 00.062-.167c0-.283-.506-.212-.606-.425h.54c.044.19.294.256.505.256.316 0 .49-.112.805-.112.332 0 .493.283.725.447h-.265a.41.41 0 00-.276-.11c-.07 0-.097.075-.165.075-.156 0-.229-.113-.385-.113-.16 0-.332.048-.44.148h-.5zm15.73-.486h.22v.225a.204.204 0 00-.11.075c-.068-.07-.119-.113-.183-.187.028-.038.055-.07.073-.113zm-2.928.075c.141 0 .224.023.366.038-.11.065-.183.084-.293.15a.55.55 0 00-.073-.188zm-6.883.225c-.005.047-.018.071-.037.114h-.11a.492.492 0 01.147-.114z"
-					fill="#fff"
-				/>
-				<path
-					d="M21.058 6.955c-.127.096-.257.131-.418.218.11.154.27.15.457.15.142 0 .206-.113.348-.113h.55c.109.14.187.234.219.412.073-.033.119-.075.201-.075.165 0 .346.036.474.12h.79a.212.212 0 010-.026c0-.098.082-.14.146-.206h.402c.188 0 .298.051.476.112a.412.412 0 01-.05.119h.82V7.66c0-.313.178-.597.443-.704H25c.02.014.043.024.07.03v.262c-.046 0-.065.037-.11.037-.106 0-.156-.098-.257-.112a2.659 2.659 0 00-.695-.005v-.182a.986.986 0 01.035-.031h-.523a.922.922 0 01-.263.03.472.472 0 01-.166-.03h-2.033zm3.651.001a.484.484 0 00.11 0h-.11zm2.364 0c.1.187.456.155.705.254-.361.145-.526.412-.915.412-.15 0-.33-.051-.33-.206 0-.084.074-.121.074-.206 0-.128-.087-.21-.202-.254h.668zm13.122 0c.057.066.104.122.104.217 0 .196-.21.239-.293.412-.293-.07-.732.047-.732-.262 0-.18.126-.332.293-.367h-6.053a.409.409 0 01-.024.03c.046.08.069.187.165.187.146 0 .22-.112.366-.112.306 0 .64.154.64.468 0 .055-.01.1-.03.137h.48a.153.153 0 01.136-.081c.066 0 .133.03.164.081h1.016a.417.417 0 01-.082-.25c0-.3.33-.43.623-.43.265 0 .412.112.677.112.146 0 .22-.075.366-.075.201 0 .316.117.458.262v.113c-.311.043-.541.16-.77.269h2.237a.376.376 0 01.07-.007h.44a.195.195 0 00.003.006h.765V7.51c0-.098.11-.126.11-.225 0-.107-.11-.154-.11-.262v-.068h-1.019zm-20.417.71a.191.191 0 01-.017-.081c0-.08.074-.108.074-.187V7.06c0-.035 0-.07.002-.107l.474.001c.008.124.22.127.22.255 0 .174-.22.22-.22.394 0 .022.002.043.007.063h-.54zm15.853-.643c.16 0 .385.005.385.169 0 .168-.183.243-.348.243-.16 0-.348-.06-.348-.225 0-.14.174-.187.311-.187zm-12.374.187c.124 0 .151.136.238.225v.15c-.073.019-.123.037-.201.037-.128 0-.275-.056-.275-.187 0-.126.115-.225.238-.225zm4.595 0h.146a.633.633 0 01-.11.114c-.018-.042-.032-.066-.036-.114z"
-					fill="#fff"
-				/>
-				<path
-					d="M21.901 6.375l.02.023h.403c.155 0 .21.15.366.15.159 0 .327-.06.414-.174h-1.203zm1.721 0c-.03.009-.06.016-.09.023v.112c.068.099.146.155.146.281 0 .097-.068.149-.158.177l.523-.001a.3.3 0 01.203-.082c.179 0 .32.067.463.082h.106a.414.414 0 00.145-.045.174.174 0 00.04.045h.916a.628.628 0 01.233-.045c.086 0 .177.014.256.045h.668a.248.248 0 01-.027-.12c0-.135.147-.182.147-.318a.19.19 0 00-.08-.155h-.606c.017.033.027.071.027.118 0 .126-.179.098-.293.13-.37.113-.563.3-.952.3-.339 0-.485-.23-.805-.337a.686.686 0 01.108-.21h-.97zm10.076 0a.296.296 0 01-.117.024.227.227 0 01-.105-.024h.222zm1.182 0a.414.414 0 01.037.06c.097.188.101.45.312.45.239 0 .317-.252.34-.51h.822c-.022.03-.046.054-.046.098 0 .365.49.29.806.45.137-.24.549-.17.549-.45a.286.286 0 00-.017-.098h.697c.173.24.697.173 1.077.173.268 0 .462-.03.665-.174h1.092v.593h-1.019a.179.179 0 00-.006-.007h-.549a.35.35 0 00-.073.007h-6.053c.078-.113.094-.27.232-.27.064 0 .082.076.146.076.124 0 .22-.118.22-.244a.317.317 0 00-.04-.154h.808zm-15.043.592c.012-.233.072-.466.29-.494.282-.04.688.074.808-.098h.416c.054.1.228.102.35.173.202.112.298.3.531.3.147 0 .22-.075.366-.075.208 0 .327.133.493.194h-2.033a.735.735 0 00.094-.082.466.466 0 00-.42-.225c-.202 0-.422.094-.422.3v.007h-.473zm18.43-.457c.091 0 .165.075.165.169a.166.166 0 01-.165.168c-.1 0-.201-.065-.201-.168s.1-.169.201-.169zm2.178.15h.257c0 .047.036.066.036.113 0 .117-.142.15-.256.15-.078 0-.115-.043-.183-.076a.933.933 0 01.146-.187z"
-					fill="#fff"
-				/>
-				<path
-					d="M22.221 5.794a.794.794 0 01-.044.054h-.659a.767.767 0 01-.044-.054h-1.566a.86.86 0 000 .035c0 .094.072.169.164.169.15 0 .233-.075.384-.075.238 0 .513.093.513.337a.217.217 0 01-.035.127h.417a.145.145 0 01-.016-.071c0-.108.114-.169.22-.169.163 0 .242.116.346.24h1.203a.295.295 0 00.061-.183c0-.14-.073-.216-.073-.356 0-.02.002-.037.005-.054h-.876zm1.651 0c.06.062.099.145.099.241 0 .193-.166.296-.349.351h.97c.213-.275.612-.313 1.008-.313.11 0 .215.042.238.15.248 0 .574-.014.669.163h.607c-.136-.109-.397-.134-.397-.332 0-.103.064-.206.164-.206.48 0 .746.187 1.227.187.14 0 .189-.124.255-.242h-4.491zm9.959 0a.516.516 0 00-.012.11c0 .131.023.23.11.318-.087.06-.15.128-.23.164h-.223c-.088-.044-.144-.138-.243-.201-.276-.179-.473-.247-.739-.391h1.337zm3.377 0a.34.34 0 01-.186.054.726.726 0 01-.28-.055h-1.584c.048.058.099.116.162.167.096.08.256.094.256.225 0 .064-.002.133-.008.201h.823a.167.167 0 00.027-.051h.66c.155 0 .228-.113.384-.113.103 0 .186.07.22.165h.698a.336.336 0 01-.058-.202c0-.173.16-.3.33-.3.288 0 .411.216.695.262.33.057.705-.121.842.188a1.225 1.225 0 01-.068.051h1.093v-.388c-.558 0-.874.056-1.428 0-.149-.015-.303-.083-.37-.205h-2.208zm-3.137.592c-.02-.04-.046-.078-.07-.126.074-.014.124-.038.202-.038.268 0 .532-.031.677.164h-.809z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.676 5.097v.016c.069.037.146.065.146.15 0 .154-.141.262-.292.262-.183 0-.293-.188-.293-.375 0-.022.005-.039.015-.053h-.245a.41.41 0 01.01.09c0 .246-.1.385-.11.62h1.567c-.167-.22-.162-.457-.29-.71h-.508zm3.526 0c.038.045.062.1.062.166 0 .243-.339.177-.513.337a.367.367 0 00-.22-.075c.014.056.028.093.037.15a.452.452 0 01.304.132h4.49a.415.415 0 01.149-.17c.247-.145.453-.047.732-.112 0-.277.297-.412.567-.412.526 0 .773.374 1.3.374.082 0 .119-.075.201-.075.078 0 .105.075.183.075.197 0 .325.024.494-.075.108-.061.077-.189.082-.315h-4.607c.105.03.205.075.205.184 0 .178-.233.174-.402.206-.348.066-.522.262-.879.262-.284 0-.494-.271-.476-.561a4.061 4.061 0 01-.44-.09h-1.269zm11.811 0a.285.285 0 01.039.147c0 .178-.21.243-.384.243-.202 0-.421-.112-.421-.318 0-.025.003-.049.01-.071h-1.454a.24.24 0 01.016.09c0 .21-.215.337-.421.337-.083 0-.12-.075-.201-.075-.156 0-.216.122-.367.15-.173.032-.347-.005-.439.15l.103.058h1.337c.029-.133.11-.246.244-.246h.696c.184 0 .278.118.387.246h1.584c-.122-.053-.214-.146-.214-.283 0-.183.206-.211.33-.337.029-.03.056-.06.081-.09h-.926zm2.77 0c.01.029.015.059.015.09 0 .104-.083.15-.147.225h-.476c-.141 0-.206-.112-.347-.112-.308 0-.396.367-.62.508h2.208a.279.279 0 01-.033-.133c0-.155.147-.235.293-.263.563-.112.925.192 1.465.005v-.192a.865.865 0 01-.028-.128h-2.33zm-15.686.71c.017-.085.068-.147.068-.245 0-.206-.256-.262-.457-.262-.275 0-.323.287-.487.508h.876z"
-					fill="#fff"
-				/>
-				<path
-					d="M21.501 4.516c-.13.291-.48.458-.825.458v.135h.509l-.033-.06c.156-.159.27-.3.494-.3.266 0 .421-.018.677.038.39.085.586.3.989.3.22 0 .33-.112.549-.112a.46.46 0 01.341.134h1.27c-.262-.066-.493-.134-.824-.134-.123 0-.188.074-.311.074-.133 0-.197-.102-.293-.187.069-.134.148-.25.24-.346h-.982c-.045.146-.242.196-.411.196-.21 0-.364-.091-.51-.195h-.88zm4.137 0c.194.149.368.27.383.533.105.01.17.038.275.038.169 0 .233-.187.402-.187.275 0 .417.117.677.187.029.007.058.014.088.023l4.607-.001a.348.348 0 01.028-.134c.12-.253.366-.3.64-.3.18 0 .257.122.422.187.23.09.566.058.643.247h1.454c.036-.125.165-.21.3-.21.18 0 .375.07.456.21h.926c.14-.165.243-.322.467-.322.083 0 .133.052.184.113h.878c.138 0 .27.084.314.209h2.33c-.02-.126-.035-.247-.154-.247-.165 0-.202.225-.366.225-.193 0-.211-.3-.403-.3-.142 0-.206.113-.348.113-.288 0-.471-.136-.64-.375l.006-.009h-.73a.27.27 0 01-.39 0H35.52a.313.313 0 01-.073.01.906.906 0 01-.13-.01h-.806a1.136 1.136 0 01-.309.047.589.589 0 01-.255-.047h-.74a.189.189 0 01-.158.084c-.064 0-.105-.037-.142-.084h-3.775v.047c.096.028.22.06.22.168 0 .131-.184.131-.312.131-.206 0-.297-.145-.494-.187-.165-.038-.343.103-.44-.038-.023-.036-.023-.078-.017-.121h-.798a.621.621 0 01-.32.084c-.148 0-.345.012-.414-.083h-.92zm15.1 0a.66.66 0 01-.33.086c-.061 0-.095-.038-.124-.086h.454zm-20.73.593c-.033-.148-.137-.242-.137-.415 0-.075.036-.128.076-.178h.456a.364.364 0 00-.02.121c0 .15.046.23.074.375-.073.03-.168.042-.204.096h-.246zm15.861-.584c.202 0 .312.01.513 0v.187c-.215.014-.334.01-.55 0-.004-.046-.018-.07-.036-.112a.284.284 0 00.073-.075zm-5.436.15c.078 0 .105.07.165.112a.396.396 0 01-.11.15h-.22c0-.033-.036-.042-.036-.075 0-.107.096-.187.2-.187zm3.972.225h.183a.12.12 0 00.037.076c-.101.01-.16.01-.257 0a.288.288 0 00.037-.076z"
-					fill="#fff"
-				/>
-				<path
-					d="M22.73 3.936l.106.076c.197.15.476.192.476.45a.22.22 0 01-.01.066h.982a.92.92 0 01.693-.29c.174 0 .252.116.421.149-.036-.061-.077-.098-.146-.112 0-.047-.037-.066-.037-.113 0-.15.165-.225.312-.225.123 0 .173.085.274.15V4.2a.869.869 0 00-.293.224c.044.038.087.072.13.105l.92-.001a.141.141 0 01-.025-.085c0-.178.033-.384.129-.508h-1.648a.78.78 0 01-.348.077H24.3c-.26 0-.361.263-.622.263-.206 0-.28-.175-.383-.34h-.565zm4.378 0c.056.053.112.114.194.151h.11a.691.691 0 01.055-.152h-.359zm3.452 0v.02c0 .36-.434.468-.787.468-.169 0-.251-.112-.42-.112-.097 0-.138.066-.22.112v.104h3.775c-.026-.033-.05-.07-.078-.104.069-.07.12-.15.22-.15.096 0 .183.07.183.17 0 .03-.009.06-.024.084h.739c-.055-.03-.093-.073-.093-.141 0-.122.1-.225.22-.225h.512c.17 0 .261-.11.367-.227h-1.322a.431.431 0 01-.18.04.33.33 0 01-.157-.04h-1.23a.413.413 0 01-.242.077.295.295 0 01-.199-.077H30.56zm7.213 0v.001c0 .155-.165.183-.293.263-.078-.117-.151-.225-.293-.225-.32 0-.467.262-.787.262-.129 0-.202-.075-.33-.075-.253 0-.33.308-.549.366h2.566a.205.205 0 01-.058-.141c0-.112.128-.15.238-.15.137 0 .215.089.31.187-.035.037-.067.073-.101.104h.731c.12-.153.198-.328.396-.328.22 0 .361.037.55.15.066.04.094.117.13.178h.455c.107-.062.183-.16.183-.29 0-.066-.041-.1-.073-.15h-.842c-.092 0-.128-.076-.22-.076h-.293c-.11 0-.165.075-.274.075-.22 0-.46-.043-.632-.15l-.815-.001zm-17.827.592c.05-.063.107-.122.107-.216v-.376h.78c-.005.082.014.173.025.264-.2.054-.395.15-.456.327l-.456.001zm2.434 0c-.19-.135-.368-.29-.643-.29-.073 0-.114.023-.183.037 0 .092-.02.177-.054.253h.88zm5.71 0c.011-.08.043-.161-.02-.216-.11-.094-.233-.06-.366-.112h-.183a.47.47 0 01-.23.329l.799-.001zm7.228.001c-.142-.02-.252-.067-.42-.067-.154 0-.268.036-.387.068h.807z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.693 3.239c.211.124.426.027.679.113.201.066.297.225.512.225h.403c.105 0 .17-.061.256-.112.165.168.353.14.55.262.09.055.15.138.203.222h-.566c-.321-.214-.612-.26-1.029-.26-.302 0-.521-.098-.769.075a.23.23 0 00-.097.186h-.781v-.148c.224-.043.549.009.549-.225 0-.183-.265-.183-.44-.15 0-.033-.036-.042-.036-.075 0-.046.02-.08.043-.114h.523zm6.598 0a.557.557 0 00.341.113c.43 0 .677-.004 1.098-.075a.514.514 0 00.13-.039h3.058l-.003.002c-.183.14-.402.21-.402.45a.35.35 0 00.111.26H30.56c-.007-.124-.076-.196-.146-.298h-.512c-.201 0-.293.15-.494.15-.357 0-.54-.188-.897-.188h-.476a.613.613 0 00-.567.335h-.36c-.062-.06-.125-.11-.226-.11a.267.267 0 00-.22.11h-1.647c.117-.06.2-.157.2-.297 0-.224-.189-.307-.377-.413h2.453zm6.466 0a.435.435 0 01-.124.711h1.322c.034-.038.069-.076.109-.11.357-.32.6-.6 1.08-.6.21 0 .265.262.476.262.265 0 .375-.225.64-.225.156 0 .165.192.22.337.06.159.29.164.293.335h.815a.445.445 0 01-.23-.39c0-.132.078-.193.11-.32a.206.206 0 00-.008-.001h-4.703zm7.055 0c0 .263-.348.263-.605.263-.2 0-.398.028-.494-.15-.019-.035-.032-.075-.045-.112l1.144-.001zm-7.517.71c-.218-.115-.305-.447-.592-.447-.303 0-.423.303-.638.448h1.23zm-9.946-.71c.201 0 .334.024.512.113-.27.024-.43.033-.695 0 .064-.042.1-.112.183-.112zm3.02 0c.133 0 .311.015.311.15 0 .174-.16.3-.33.3-.136 0-.292-.065-.292-.206 0-.154.16-.243.311-.243zm-.897.226h.22a.12.12 0 00.036.076.61.61 0 01-.292 0 .29.29 0 00.036-.076z"
-					fill="#fff"
-				/>
-				<path
-					d="M24.191 2.658c.004.028.036.038.036.07 0 .102-.119.149-.22.149-.123 0-.169-.112-.292-.112h-.403a.605.605 0 00-.11.224c-.123-.065-.206-.112-.348-.112-.21 0-.302.187-.512.187-.137 0-.206-.107-.275-.225v-.15l.087-.03h2.037zm1.988 0a.21.21 0 00-.012.07c0 .172.16.299.33.299.16 0 .229-.113.366-.188h.183c.032.171.12.318.245.412h-2.453a2.924 2.924 0 01-.062-.037.285.285 0 00.037-.075h.403c.33 0 .805.075.805-.262 0-.142-.157-.175-.297-.218h.455zm4.412 0c.096.05.187.107.317.107.092 0 .129-.075.22-.075.16 0 .275.154.275.318 0 .117-.106.206-.22.206-.128 0-.183-.112-.311-.112-.224 0-.343.112-.568.112-.201 0-.439-.075-.439-.28a.32.32 0 01.16-.275h-.275a.245.245 0 01.005.05c0 .163-.215.163-.366.205-.232.063-.334.248-.528.336h3.057c.12-.094.034-.266.107-.41a.458.458 0 01.152-.181H30.59zm2.068 0c.173.163.23.519.501.519h.403c.075 0 .141.028.194.074h4.703a1.61 1.61 0 00-.211-.037c-.17.042-.243-.187-.421-.187-.106 0-.151.075-.257.075-.105 0-.2-.08-.2-.188 0-.133.078-.209.19-.255h-1.183c-.093.057-.213.068-.344.068a.516.516 0 01-.263-.07l-.87.001c.03.048.053.097.053.163 0 .28-.348.356-.622.356-.417 0-.678-.127-.989-.412a.388.388 0 01.083-.108l-.767.001zm7.103 0a.193.193 0 00-.012.07c0 .172.16.299.33.299h.585c.06.074.146.121.146.223h-1.143c-.027-.078-.055-.148-.138-.148-.16 0-.252.023-.403.075a.273.273 0 00-.073-.038c.064-.08.11-.154.11-.262a.408.408 0 00-.064-.22l.662.001zM20.17 3.25c.031-.045.067-.087.067-.148 0-.117-.087-.169-.146-.263v-.112c.032 0 .041-.037.073-.037.147 0 .22.06.366.075 0 .177-.027.341.11.45a.512.512 0 00.053.035h-.523zm15.388-.448c.065 0 .087.066.129.112v.15c-.046.01-.06.075-.11.075a.186.186 0 01-.183-.187c0-.09.077-.15.164-.15z"
-					fill="#fff"
-				/>
-				<path
-					d="M21.589 2.077c-.201.19-.38.362-.693.362-.302 0-.421-.206-.696-.337a.52.52 0 00.03-.025h1.359zm3.18.001a6.534 6.534 0 00-.286.025.861.861 0 01-.008-.024h.294zm5.9 0c-.089.067-.162.136-.291.136h-1.392c-.258 0-.445-.058-.637-.136h-1.238c-.073.078-.179.139-.248.249.307 0 .462.187.769.187.187 0 .27-.15.457-.15.316 0 .472.187.787.187.17 0 .252-.112.421-.112.19 0 .416.06.453.231h.275a.57.57 0 01.298-.081c.107 0 .189.038.268.08h1.586a.503.503 0 01.287-.08c.081 0 .143.032.195.08h.767a.178.178 0 01.1-.043c.508-.06.805-.023 1.318-.037.02.029.04.055.057.081h.87c-.075-.042-.146-.098-.232-.156.138-.196.379-.264.473-.436h-.63c-.093.092-.248.136-.392.136-.186 0-.378-.046-.528-.137h-.454c.085.043.14.11.14.231 0 .16-.21.169-.366.169-.412 0-.61-.329-.976-.4h-.242c-.305.081-.406.512-.759.512-.137 0-.293-.066-.293-.206 0-.143.158-.192.206-.306h-1.05zm7.169 0c-.113.115-.413.054-.614.136-.101.043-.129.15-.238.15-.124 0-.188-.075-.312-.075s-.123.16-.183.262a.335.335 0 01-.113.119h1.183c.124-.051.29-.066.431-.081.394-.047.627-.024 1.025-.038.032.041.06.079.083.118h.662c.036-.093.14-.155.244-.155h.622c.087 0 .183-.042.183-.131 0-.119-.01-.213-.031-.306h-.246c-.229.316-.533.512-.95.512-.205 0-.233-.257-.42-.337-.159-.069-.377-.07-.469-.174h-.857zm-15.684.592c.291-.11.48-.23.81-.23.334 0 .49.224.824.224h.402v.005h-2.036zm3.57 0c-.113-.035-.216-.078-.216-.193 0-.113.11-.188.22-.188.26 0 .43.028.659.15-.075.085-.176.135-.209.23h-.454z"
-					fill="#fff"
-				/>
-				<path
-					d="M20.23 2.099c.065-.06.112-.125.209-.125s.141.075.238.075h.402c.032-.056.073-.085.073-.15 0-.314-.416-.253-.695-.375-.188-.08-.229-.252-.293-.45.151-.079.265-.074.44-.074.21 0 .32.075.53.075.234 0 .37-.075.604-.075.645 0 1.002.15 1.648.15 1.871.01 2.928.014 4.796-.075.086 0 .128-.075.22-.075h9.371c.018.056.032.089.037.15.032 0 .041.037.073.037.092 0 .105-.107.146-.187h2.71v.225c-.138.112-.284.065-.44.15-.155.084-.15.337-.33.337-.228 0-.301-.263-.53-.263-.256 0-.357.192-.568.338-.086.06-.22.074-.22.187a.18.18 0 00.045.125h-.858a.168.168 0 00.045-.125c0-.187-.11-.375-.293-.375-.252 0-.37.188-.622.188-.28 0-.38-.3-.66-.3-.155 0-.329.084-.329.243 0 .085.073.122.073.206a.33.33 0 01-.04.163h-.63a.272.272 0 00.085-.2c0-.398-.49-.487-.879-.487-.178 0-.403.042-.403.225a.53.53 0 00.278.462h-.455c-.183-.093-.504-.071-.701-.238-.105-.088-.041-.257-.147-.337-.242-.192-.48-.112-.787-.112-.316 0-.75-.023-.75.3 0 .13.11.187.11.318a.176.176 0 01-.014.069h-1.05a.616.616 0 01.148-.088c.128-.047.293-.065.293-.206 0-.234-.325-.164-.549-.206-.229-.042-.334-.187-.567-.187-.275 0-.421.15-.696.15-.352 0-.545-.15-.897-.15-.366 0-.879-.037-.879.337 0 .14.202.108.33.15.194.066.347.14.498.2h-1.239a.281.281 0 00.082-.2c0-.234-.265-.337-.494-.337-.096 0-.142.075-.238.075h-.403c-.274 0-.426.206-.549.45-.292 0-.496-.003-.74.012h-.293c-.076-.225-.106-.5-.34-.5-.206 0-.33.005-.53.038-.101.014-.17.06-.257.112h-.33l-.073-.075c.028-.056.074-.084.074-.15 0-.093-.092-.15-.184-.15-.16 0-.045.295-.183.375-.242.14-.494-.08-.732.075a2.34 2.34 0 00-.332.274H20.23zm12.576 0a.776.776 0 00-.14-.012.393.393 0 00-.102.013h.242zm7.728 0c.061-.084.117-.176.168-.275.033.104.06.189.078.275h-.246z"
-					fill="#fff"
-				/>
-				<path
-					d="M30.135 6.22c1.487 0 3.13.693 3.13 2.185 0 .413-.183 1.377-.183 1.377-.041.31-.073.485-.146.79-.09.371-.268.898-.293.936-.137.438-1.282 1.45-1.282 1.872 0 .551.586.845 1.135.845h.366c.49 0 1.771.656 2.16.77.98.295 1.543.552 2.563.552.179 0 .147-.262.147-.441v-.771c0-.33.695-1.198.951-1.469.453-.472.806-.674 1.428-.88.197-.065.302-.185.513-.185h.549c.357 0 .503.257.805.441.66.404 1.02.716 1.501 1.322.44.55.504 1.358 1.208 1.358.234 0 .357-.105.586-.147.352-.064.613.046.915-.146.33-.211.43-.588.824-.588.188 0 .275.216.275.404v.477c0 1.125-1.094 1.556-2.16 1.91-.77.257-1.19.486-1.977.66-.375.083-.879.088-.879.478v.514c.627.418 1.13.477 1.886.477 1.006-.082 1.578-.114 2.58-.22 1.277-.138 2.046-.33 3.149-.991.412-.248.879-.381.879-.863 0-.822-.39-1.35-1.062-1.818-.558-.39-1.501-.215-1.501-.9 0-.22.014-.353.037-.568.141-1.143.21-1.795.512-2.901.092-.33.055-.574.256-.845.188-.247.463-.188.77-.257.48-.11.75-.22 1.244-.22 1.016 0 1.583.317 2.49.771.75.376 1.464.753 1.464 1.597 0 .478-.193.73-.403 1.157-.499 1-1.025 1.56-1.025 2.68v.808c0 .711.512.991.842 1.616.224.426.445 1.173.476 1.23.348.628.659 1.423.769 1.634.137.257.444 1.556.586 2.092.27 1.024.545 1.589.658 2.644 0 0 .147 1.024.147 1.359v.863c0 .472-.019.865-.128 1.325l-.092.382c-.19.8-.924 4.13-1.025 4.736-.15.91-.356 1.404-.512 2.314-.318 1.85-.508 2.836-.952 4.626-.449 1.804-.59 2.84-.952 4.663a44.312 44.312 0 01-.513 2.313c-.141.583-.169.927-.329 1.506-.073.27-.192.55-.476.55-.078 0-.119-.036-.183-.073v-4.48c0-.904-.037-1.413-.11-2.313-.032-.399-.11-.624-.11-1.028v-6.609c0-.303.11-1.322.11-1.322h-.22c-.622.39-.553 1.038-.695 1.763-.256 1.294-.44 2.033-.55 3.341-.15 1.772-.086 2.79-.329 4.553-.251 1.836-.585 2.85-.952 4.663-.517-.239-.832-.606-.878-1.175-.096-1.157-.196-3.553-.22-3.929l-.11-1.689s-.08-1.262-.11-1.652c0 0-.082-1.088-.11-1.432l-.219-2.717v-2.02c0-.807.33-4.185.33-4.185.03-.379.146-1.684.146-1.763 0-.463.15-2.331.293-3.047.027-.142.11-1.065.11-1.065.028-.274-.339-.29-.586-.404-.504-.234-.838-.27-1.355-.477a9.155 9.155 0 00-1.244-.404c-.307-.078-.467-.257-.788-.257-.173 0-.2.23-.2.404 0 .124.255 2.152.438 3.047 0 0 .218 1.092.293 1.432 0 0 .293 1.143.293 1.322 0 .064.32 1.946.366 2.35.133 1.138.11 1.795.22 2.938.027.284.087.446.11.734l.11 1.395c.091 1.07.114 1.685.183 2.754-.453.179-.76.142-1.245.184-.71.064-1.121.04-1.83.11-.188.018-.376-.014-.477.147-.22.348-.096.651-.146 1.064a39.35 39.35 0 00-.22 2.13s-.105 1.11-.11 1.469c-.036 2.033-.004 3.18-.004 5.213-.417 0-.622-.21-1.02-.33-.522-.156-.86-.073-1.41-.073-.252 0-.229-.34-.275-.588-.073-.385-.732-3.02-.732-3.103 0-.091-.36-2.767-.366-2.845 0 0-.11-1.35-.11-1.432 0-.445-.014-.698.037-1.138.009-.088-.037-.881-.037-1.304 0-.294-.357-.307-.622-.422a20.35 20.35 0 01-1.83-.881c-.38-.202-.806-.335-.806-.771 0-.74.513-2.607.513-2.607l.659-3.048s.306-1.85.439-2.35c.128-.486.366-1.946.366-1.946.054-.288.293-3.47.293-3.561V21.31c0-.293-.37-.349-.66-.367-.622-.041-.983-.023-1.61-.037-.046.078-.073.726-.073.937 0 .092-.073.642-.073.936 0 .106-.11 1.248-.11 1.248-.076.86-.175 1.185-.257 2.167-.171 2.058.05 3.25-.256 5.287-.009.087-.114 1.414-.183 2.166-.096 1.042-.201 1.625-.33 2.662.019.042-.073.753-.073.753-.01.105-.146 1.359-.146 1.359-.058.54-.22 1.5-.22 1.597 0 .266-.732 4.686-.732 4.755 0 .463-.769 2.97-1.318 4.002h-.44c-.15-.29-.104-.514-.109-.844-.078-5.297.023-8.294-.11-13.585 0-.097-.068-.161-.146-.22-.028.605-.165.94-.22 1.541a46.595 46.595 0 00-.183 2.644c-.17 3.933-.183 6.164-.366 10.097a.924.924 0 01-.11.44c-.42-.224-.503-.6-.622-1.064-.224-.845-.412-1.318-.623-2.167-.315-1.266-.485-1.978-.695-3.268-.302-1.817-1.062-5.842-1.062-6.609v-8.15c0-1.933-.014-3.03.11-4.958l.073-1.028c-.055-.046-.128-.326-.146-.367-.838 0-1.304-.147-2.142-.147-.119 0-.165.138-.165.257.21 1.4.385 2.194.55 3.598.173 1.469.33 2.288.33 3.782v.661l.21 2.644v3.745c0 1.262-.092 1.978-.21 3.231-.079.799-.317 1.216-.55 1.983-.398 1.308-1.004 3.803-1.135 4.039-.405.73-1.043 3.107-1.172 3.892-.036.23-.022.55-.256.55-.613 0-.805-.743-.805-1.358.11-1.097.142-1.735.293-2.827.009-.087.1-5.191.036-8.372-.009-.537-.21-.872-.585-1.248-.133.082-.22.174-.22.33v1.322c0 .344-.11 1.359-.11 1.359-.111 1.376-.348 2.148-.44 3.525-.164 2.469-.136 3.873-.219 6.352-.041.009-.064.036-.11.036-1.345 0-1.345-1.735-1.647-3.047-.179-.785-.16-1.258-.293-2.056-.096-.583-.125-.921-.22-1.506 0 0-.242-1.822-.439-2.717-.398-1.804-.462-2.86-.879-4.663-.302-1.317-.491-2.072-.659-3.415 0 0-.068-.5-.073-.587-.082-1.441.01-2.268-.073-3.709 0-.046-.027-.67 0-1.028-.247-.197-.471-.174-.769-.294h-.22l-.279 10.207v5.288c0 .81-.077 1.257-.16 2.056-.155 1.45-.311 2.267-.549 3.708-.183 1.106-.188 1.822-.805 2.754-.188-.115-.128-.335-.183-.55-.124-.473-.151-.763-.183-1.249-.11-1.63-.101-2.556-.22-4.186-.133-1.84-.366-2.891-.989-4.626-.357-1-.402-1.69-1.098-2.497-.393.395-.242.85-.366 1.395-.215.937-.27 1.483-.476 2.424-.192.885-.146 1.418-.293 2.313-.096.596-.22 1.91-.22 1.91-.03.26-.11 1.1-.11 1.1a28.863 28.863 0 01-.438 2.975c-.179.927-.142 1.638-.843 2.276-.082-1.629-.087-2.56-.22-4.186-.16-1.95-.356-3.047-.548-4.993-.142-1.441-.225-2.263-.33-3.709-.064-.9-.07-1.41-.11-2.313 0 0-.08-2.13-.11-2.46l-.11-1.248s-.075-1.021-.11-1.322c0 0-.17-1.347-.219-1.91 0 0-.131-1.868-.22-2.57 0 0-.219-1.202-.219-1.744 0-.151-.147-3.63-.147-3.726 0-.63.197-.983.476-1.543.151-.307 1.062-1.858 1.575-2.46.393-.463.878-.693.878-1.303 0-.11-.073-.606-.073-.606-.142-1.171-1.172-1.611-1.172-2.79v-.441c0-.473.344-.68.696-.991.837-.74 1.446-1.286 2.563-1.286 1.016 0 2.196.376 2.196 1.396v3.01c0 .35.11 1.286-.073 1.396-.252.151-.457.11-.732.22-.53.211-.879.601-.879 1.175 0 .372.165.574.366.881.504.758.897 1.12 1.538 1.763.444.445.764.66 1.318.954.49.262.76.551 1.318.551h.366c.307 0 .302-.385.44-.66.137-.276.402-.386.402-.698 0-.368-.512-.303-.879-.33-.631-.051-1.464-.065-1.464-.698 0-.193.146-.285.146-.478.353-.032.563-.004.916-.037.85-.082 1.313-.266 2.16-.403.92-.152 1.867-.57 1.867-1.506 0-.546-.33-.794-.623-1.248-.196-.308-.438-.43-.512-.79-.052-.252-.008-.406.036-.66.14-.805.503-1.225 1.099-1.781.436-.408.726-.644 1.3-.808.453-.13.874-.074 1.208-.074.448 0 .709-.013 1.153.037.833.101 1.441.707 1.5 1.542.037.542-.027.854.037 1.395.005.051-.073.446-.073.661v.478c0 .6-.444 1.138-1.043 1.138-.234 0-.371-.073-.604-.073-.151 0-.257.142-.257.293 0 .198.06.441.257.441.782 0 1.194-.367 1.977-.367h.622c.265 0 .403-.115.659-.184.718-.197 1.149-.294 1.83-.587.394-.17.6-.308.952-.551.188-.129.366-.211.403-.44.114-.67.11-1.066.11-1.745 0-.555-.55-.679-.879-1.12-.49-.652-.659-1.18-.659-2 0-.556 0-.905.22-1.414.128-.299.865-1.3 1.281-1.506.444-.22.764-.22 1.263-.22z"
-					stroke="#CC1E1C"
-					strokeWidth=".837"
-				/>
-				<path
-					d="M30.135 6.22c1.487 0 3.13.693 3.13 2.185 0 .413-.183 1.377-.183 1.377-.041.31-.073.485-.146.79-.09.371-.268.898-.293.936-.137.438-1.282 1.45-1.282 1.872 0 .551.586.845 1.135.845h.366c.49 0 1.771.656 2.16.77.98.295 1.543.552 2.563.552.179 0 .147-.262.147-.441v-.771c0-.33.695-1.198.951-1.469.453-.472.806-.674 1.428-.88.197-.065.302-.185.513-.185h.549c.357 0 .503.257.805.441.66.404 1.02.716 1.501 1.322.44.55.504 1.358 1.208 1.358.234 0 .357-.105.586-.147.352-.064.613.046.915-.146.33-.211.43-.588.824-.588.188 0 .275.216.275.404v.477c0 1.125-1.094 1.556-2.16 1.91-.77.257-1.19.486-1.977.66-.375.083-.879.088-.879.478v.514c.627.418 1.13.477 1.886.477 1.006-.082 1.578-.114 2.58-.22 1.277-.138 2.046-.33 3.149-.991.412-.248.879-.381.879-.863 0-.822-.39-1.35-1.062-1.818-.558-.39-1.501-.215-1.501-.9 0-.22.014-.353.037-.568.141-1.143.21-1.795.512-2.901.092-.33.055-.574.256-.845.188-.247.463-.188.77-.257.48-.11.75-.22 1.244-.22 1.016 0 1.583.317 2.49.771.75.376 1.464.753 1.464 1.597 0 .478-.193.73-.403 1.157-.499 1-1.025 1.56-1.025 2.68v.808c0 .711.512.991.842 1.616.224.426.445 1.173.476 1.23.348.628.659 1.423.769 1.634.137.257.444 1.556.586 2.092.27 1.024.545 1.589.658 2.644 0 0 .147 1.024.147 1.359v.863c0 .472-.019.865-.128 1.325l-.092.382c-.19.8-.924 4.13-1.025 4.736-.15.91-.356 1.404-.512 2.314-.318 1.85-.508 2.836-.952 4.626-.449 1.804-.59 2.84-.952 4.663a44.312 44.312 0 01-.513 2.313c-.141.583-.169.927-.329 1.506-.073.27-.192.55-.476.55-.078 0-.119-.036-.183-.073v-4.48c0-.904-.037-1.413-.11-2.313-.032-.399-.11-.624-.11-1.028v-6.609c0-.303.11-1.322.11-1.322h-.22c-.622.39-.553 1.038-.695 1.763-.256 1.294-.44 2.033-.55 3.341-.15 1.772-.086 2.79-.329 4.553-.251 1.836-.585 2.85-.952 4.663-.517-.239-.832-.606-.878-1.175-.096-1.157-.196-3.553-.22-3.929l-.11-1.689s-.08-1.262-.11-1.652c0 0-.082-1.088-.11-1.432l-.219-2.717v-2.02c0-.807.33-4.185.33-4.185.03-.379.146-1.684.146-1.763 0-.463.15-2.331.293-3.047.027-.142.11-1.065.11-1.065.028-.274-.339-.29-.586-.404-.504-.234-.838-.27-1.355-.477a9.155 9.155 0 00-1.244-.404c-.307-.078-.467-.257-.788-.257-.173 0-.2.23-.2.404 0 .124.255 2.152.438 3.047 0 0 .218 1.092.293 1.432 0 0 .293 1.143.293 1.322 0 .064.32 1.946.366 2.35.133 1.138.11 1.795.22 2.938.027.284.087.446.11.734l.11 1.395c.091 1.07.114 1.685.183 2.754-.453.179-.76.142-1.245.184-.71.064-1.121.04-1.83.11-.188.018-.376-.014-.477.147-.22.348-.096.651-.146 1.064a39.35 39.35 0 00-.22 2.13s-.105 1.11-.11 1.469c-.036 2.033-.004 3.18-.004 5.213-.417 0-.622-.21-1.02-.33-.522-.156-.86-.073-1.41-.073-.252 0-.229-.34-.275-.588-.073-.385-.732-3.02-.732-3.103 0-.091-.36-2.767-.366-2.845 0 0-.11-1.35-.11-1.432 0-.445-.014-.698.037-1.138.009-.088-.037-.881-.037-1.304 0-.294-.357-.307-.622-.422a20.35 20.35 0 01-1.83-.881c-.38-.202-.806-.335-.806-.771 0-.74.513-2.607.513-2.607l.659-3.048s.306-1.85.439-2.35c.128-.486.366-1.946.366-1.946.054-.289.293-3.47.293-3.561V21.31c0-.294-.37-.349-.66-.367-.622-.041-.983-.023-1.61-.037-.046.078-.073.726-.073.937 0 .092-.073.642-.073.936 0 .106-.11 1.248-.11 1.248-.076.86-.175 1.185-.257 2.167-.171 2.058.05 3.25-.256 5.287-.009.087-.114 1.414-.183 2.166-.096 1.042-.201 1.625-.33 2.662.019.042-.073.753-.073.753-.01.105-.146 1.359-.146 1.359-.058.54-.22 1.5-.22 1.597 0 .266-.732 4.686-.732 4.755 0 .463-.769 2.97-1.318 4.002h-.44c-.15-.29-.104-.514-.109-.844-.078-5.297.023-8.294-.11-13.585 0-.097-.068-.161-.146-.22-.028.605-.165.94-.22 1.541a46.595 46.595 0 00-.183 2.644c-.17 3.933-.183 6.164-.366 10.097a.924.924 0 01-.11.44c-.42-.224-.503-.6-.622-1.064-.224-.845-.412-1.318-.623-2.167-.315-1.266-.485-1.978-.695-3.268-.302-1.817-1.062-5.842-1.062-6.609v-8.15c0-1.933-.014-3.03.11-4.958l.073-1.028c-.055-.046-.128-.326-.146-.367-.838 0-1.304-.147-2.142-.147-.119 0-.165.138-.165.257.21 1.4.385 2.194.55 3.598.173 1.469.33 2.288.33 3.782v.661l.21 2.644v3.745c0 1.262-.092 1.978-.21 3.231-.079.799-.317 1.216-.55 1.983-.398 1.308-1.004 3.803-1.135 4.039-.405.73-1.043 3.107-1.172 3.892-.036.23-.022.55-.256.55-.613 0-.805-.743-.805-1.358.11-1.097.142-1.735.293-2.827.009-.087.1-5.191.036-8.372-.009-.537-.21-.872-.585-1.248-.133.082-.22.174-.22.33v1.322c0 .344-.11 1.359-.11 1.359-.111 1.375-.348 2.148-.44 3.525-.164 2.469-.136 3.873-.219 6.352-.041.009-.064.036-.11.036-1.345 0-1.345-1.735-1.647-3.047-.179-.785-.16-1.258-.293-2.056-.096-.583-.125-.921-.22-1.506 0 0-.242-1.822-.439-2.717-.398-1.804-.462-2.86-.879-4.663-.302-1.317-.491-2.072-.659-3.415 0 0-.068-.5-.073-.587-.082-1.441.01-2.268-.073-3.709 0-.046-.027-.67 0-1.028-.247-.197-.471-.174-.769-.294h-.22l-.279 10.207v5.288c0 .81-.077 1.257-.16 2.056-.155 1.45-.311 2.267-.549 3.708-.183 1.106-.188 1.822-.805 2.754-.188-.115-.128-.335-.183-.55-.124-.473-.151-.763-.183-1.249-.11-1.63-.101-2.556-.22-4.186-.133-1.84-.366-2.891-.989-4.626-.357-1-.402-1.69-1.098-2.497-.393.395-.242.85-.366 1.395-.215.937-.27 1.483-.476 2.424-.192.885-.146 1.418-.293 2.313-.096.596-.22 1.91-.22 1.91-.03.26-.11 1.1-.11 1.1a28.855 28.855 0 01-.438 2.975c-.179.927-.142 1.638-.843 2.276-.082-1.629-.087-2.56-.22-4.186-.16-1.95-.356-3.047-.548-4.993-.142-1.441-.225-2.263-.33-3.709-.064-.9-.07-1.41-.11-2.313 0 0-.08-2.13-.11-2.46l-.11-1.248s-.075-1.021-.11-1.322c0 0-.17-1.347-.219-1.91 0 0-.131-1.868-.22-2.57 0 0-.219-1.202-.219-1.744 0-.151-.147-3.63-.147-3.726 0-.63.197-.983.476-1.543.151-.307 1.062-1.858 1.575-2.46.393-.463.878-.693.878-1.303 0-.11-.073-.606-.073-.606-.142-1.171-1.172-1.611-1.172-2.79v-.441c0-.473.344-.68.696-.991.837-.74 1.446-1.286 2.563-1.286 1.016 0 2.196.376 2.196 1.396v3.01c0 .35.11 1.286-.073 1.396-.252.151-.457.11-.732.22-.53.211-.879.601-.879 1.175 0 .372.165.574.366.881.504.758.897 1.12 1.538 1.763.444.445.764.66 1.318.954.49.262.76.551 1.318.551h.366c.307 0 .302-.385.44-.66.137-.276.402-.386.402-.698 0-.368-.512-.303-.879-.33-.631-.051-1.464-.065-1.464-.698 0-.193.146-.285.146-.478.353-.032.563-.004.916-.037.85-.082 1.313-.266 2.16-.403.92-.152 1.867-.57 1.867-1.506 0-.546-.33-.794-.623-1.248-.196-.308-.438-.43-.512-.79-.052-.252-.008-.406.036-.66.14-.805.503-1.225 1.099-1.781.436-.408.726-.644 1.3-.808.453-.13.874-.074 1.208-.074.448 0 .709-.013 1.153.037.833.101 1.441.707 1.5 1.542.037.542-.027.854.037 1.395.005.051-.073.446-.073.661v.478c0 .6-.444 1.138-1.043 1.138-.234 0-.371-.073-.604-.073-.151 0-.257.142-.257.293 0 .198.06.441.257.441.782 0 1.194-.367 1.977-.367h.622c.265 0 .403-.115.659-.184.718-.197 1.149-.294 1.83-.587.394-.17.6-.308.952-.551.188-.129.366-.211.403-.44.114-.67.11-1.066.11-1.745 0-.555-.55-.679-.879-1.12-.49-.652-.659-1.18-.659-2 0-.556 0-.905.22-1.414.128-.299.865-1.3 1.281-1.506.444-.22.764-.22 1.263-.22"
-					fill="#fff"
-				/>
-			</g>
-		</svg>
-	);
-};
