--- src/components/iconSet/ts/logos/IconRaphaelswerk.tsx
@@ -1,45 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconRaphaelswerk = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			viewBox="0 0 390.9 345.1"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<path
-				className="st0"
-				fill="#fff"
-				d="M385.4 334.3h5.5v-5.8h-5.5v5.8zm.6-36.1h-5.2L370 329.7h-.1l-10.6-31.5h-5.5l12.9 36.1h6.1l13.2-36.1zm-37.6 36.1h5.5v-5.8h-5.5v5.8zm-23.7-15.8c.3-3.4 2.2-7.3 6.7-7.3 4 0 6.2 3.3 6.2 7.3h-12.9zm16 9.8c-1.1 1-4.7 2.6-7.7 2.6-4.4 0-8.2-3.6-8.2-8.8h17.9c0-8.5-2.7-14.9-11-14.9-7.1 0-11.7 5.5-11.7 13.3 0 8.6 4.1 14.3 13 14.3 4.1 0 6.1-1 7.9-1.7v-4.8zm-54.3 6h4.6v-13.5l11.2 13.5h6.6l-13.1-14.7 11.6-11.7H301l-10 10.7v-23.1h-4.6v38.8zm-18.6 0h4.6v-12c0-6.6 2.7-10.8 6.8-10.8.8 0 1.7.1 2.6.5v-4.3c-1-.2-1.7-.4-3-.4-2.7 0-5.3 2-6.5 4.7h-.1v-4.1h-4.3v26.4zm-23.6-15.8c.3-3.4 2.2-7.3 6.7-7.3 4 0 6.2 3.3 6.2 7.3h-12.9zm16 9.8c-1.1 1-4.7 2.6-7.7 2.6-4.4 0-8.2-3.6-8.2-8.8h17.9c0-8.5-2.7-14.9-11-14.9-7.1 0-11.7 5.5-11.7 13.3 0 8.6 4.1 14.3 13 14.3 4.1 0 6.1-1 7.9-1.7v-4.8zM237 307.9h-4.5l-6.1 21.8h-.1l-6.3-21.8h-5.5l-6.4 21.8h-.1l-5.7-21.8h-5.1l8.1 26.4h5.4l6.3-22.1h.1l6.5 22.1h5.4l8-26.4zm-57.1 25.5c2.1 1.2 4.5 1.4 6.9 1.4 4.5 0 8.7-2.6 8.7-8.2 0-8.4-10.9-6.9-10.9-12.2 0-2.2 2.1-3.2 4.4-3.2 1.1 0 3.8.7 4.9 1.3l.4-4.2c-1.8-.6-3.6-1-5.7-1-5 0-8.8 2.6-8.8 8.1 0 7.4 10.9 7 10.9 11.8 0 2.8-2.6 3.8-4.7 3.8-1.3 0-3.9-.6-5.7-1.9l-.4 4.3zm-10.5.9h4.6v-38.8h-4.6v38.8zm-24.2-15.8c.3-3.4 2.2-7.3 6.7-7.3 4 0 6.2 3.3 6.2 7.3h-12.9zm16 9.8c-1.1 1-4.7 2.6-7.7 2.6-4.4 0-8.2-3.6-8.2-8.8h17.9c0-8.5-2.7-14.9-11-14.9-7.1 0-11.7 5.5-11.7 13.3 0 8.6 4.1 14.3 13 14.3 4.1 0 6.1-1 7.9-1.7v-4.8zm-31.1-4.3c0 3.9-2.4 6.9-7 6.9-2.1 0-4.7-1.5-4.7-4.3 0-4.6 6.2-5.1 8.7-5.1 1 0 2 .1 2.9.1v2.4zm-13.5-10.1c1.9-1.7 4.7-2.7 7.3-2.7 4.6 0 6.2 2.3 6.2 6.9-1.8-.1-3.1-.1-4.9-.1-4.8 0-11.6 2.1-11.6 8.7 0 5.7 3.8 8.2 9.2 8.2 4.2 0 6.7-2.4 7.8-4h.1v3.4h4.3c-.1-.8-.2-2.2-.2-5.2v-10.9c0-7.2-2.9-10.9-10.1-10.9-3.2 0-5.9 1-8.2 2.5l.1 4.1zm-30.9 20.4h4.6v-12.2c0-6.4 2.4-10.9 7.6-10.9 4 .3 5.3 3.3 5.3 9.4v13.6h4.6v-15.7c0-7.1-2.9-11.3-9.4-11.3-3.5 0-6.4 1.6-8.1 4.4h-.1v-16.2h-4.6v38.9zM75.1 321c.1 4.1-2 10-6.9 10-5.2 0-7.4-5.4-7.4-9.8 0-4.9 2.4-9.9 7.4-9.9 5-.1 7 4.9 6.9 9.7m-18.9 24.1h4.6v-14.5h.1c1.4 2.4 3.9 4.2 8.3 4.2 7.6 0 10.8-7 10.8-14.3 0-7.1-3.5-13.3-10.8-13.3-4.4 0-7.3 2.4-8.5 4.3h-.1v-3.7h-4.4v37.3zm-12-21.1c0 3.9-2.4 6.9-7 6.9-2.1 0-4.7-1.5-4.7-4.3 0-4.6 6.2-5.1 8.7-5.1 1 0 2 .1 2.9.1v2.4zm-13.4-10.1c1.9-1.7 4.7-2.7 7.3-2.7 4.6 0 6.2 2.3 6.2 6.9-1.8-.1-3.1-.1-4.9-.1-4.8 0-11.6 2.1-11.6 8.7 0 5.7 3.8 8.2 9.2 8.2 4.2 0 6.7-2.4 7.8-4h.1v3.4h4.3c-.1-.8-.2-2.2-.2-5.2v-10.9c0-7.2-2.9-10.9-10.1-10.9-3.2 0-5.9 1-8.2 2.5l.1 4.1zM4.9 302.7h4.7c4.5 0 6.9 2.1 6.9 5.3 0 3.5-2.7 5.6-7.4 5.6H4.9v-10.9zM0 334.3h4.9v-16h2.9c2.6 0 4.2 0 5.7 3.5l5.3 12.5h5.6L18 319.9c-1.1-2.2-1.8-3-3.3-3.6v-.1c4-.7 7.1-3.8 7.1-8.2 0-8.6-6.5-9.8-13.3-9.8H0v36.1z"
-			/>
-			<path
-				className="st1"
-				d="M218.2 246.6v13.3h-25v-43.5H186c-13.6 0-17.5 20.6-28.3 43.5h-27.6c14.1-31.4 17.2-44.6 27.6-50.1-11.5-5.7-17.2-14.5-17.2-26.2 0-16 12.2-30.8 32.9-30.8h21.9m-2.1 19.5h-12.7c-9.4 0-14.3 4.1-14.3 12.1 0 9.1 6.8 12.3 13.7 12.3h13.3v-24.4z"
-			/>
-			<path
-				className="st1"
-				fill="none"
-				d="M196.3 159.4s-5-7.1 5.1-15.9c14.5-12.6 5-34.8 39.6-73.6 29.4-33 11.3-70.7 13.6-68.9 22.3 17.2 13.7 58 13.7 58s6-21.7 6.8-30c.7-7.4 8.1 9.6 4.5 40.2-1.5 12.4-6 25.7-6 25.7s10.2-18.8 11.4-23.4c.5-2 2.8 20.4-4.7 40.2-2.9 7.8-5.4 14-5.4 14s5-6.8 9.1-13.4c5.3-8.6 2.5 12-.3 19.3-4.1 10.4-8.3 16.4-8.3 16.4s2.1 1 6.3-2.5c2-1.6 2.2 3.3-.1 10.2-1.5 4.5-8.2 11.6-8.2 11.6s6.1-1.1 8.9-3.9c1-1 1.2 3-1.4 9-3.6 8.5-15.4 18.9-15.4 18.9s5 .5 10.2-3.7c2.1-1.7-.7 6.6-4.3 10.6-4.8 5.3-10.7 7.9-10.7 7.9s10.4.3 1.6 10.3c-6.3 7.1-15 7.4-15 7.4s-1.7 13.4-14.7 10.8c0 0-7.3 14.7-20.3 11.3 0 0-2.5-.5-2.4-2.4.3-3.4 5.2-6.9 5.2-12.9 0-4-2.2-7.4-2.2-7.4"
-			/>
-			<path
-				d="M218.2 246.6v13.3h-25v-43.5H186c-13.6 0-17.5 20.6-28.3 43.5h-27.6c14.1-31.4 17.2-44.6 27.6-50.1-11.5-5.7-17.2-14.5-17.2-26.2 0-16 12.2-30.8 32.9-30.8h21.9m-2.1 19.5h-12.7c-9.4 0-14.3 4.1-14.3 12.1 0 9.1 6.8 12.3 13.7 12.3h13.3v-24.4zm3.1-12.9s-5-7.1 5.1-15.9c14.6-12.7 5-34.8 39.6-73.6 29.4-33 11.3-70.7 13.6-68.9 22.3 17.2 13.7 58 13.7 58s6-21.7 6.8-30c.7-7.4 8.1 9.6 4.5 40.2-1.5 12.4-2.9 18.5-6.2 26.2 0 0 10.3-19.2 11.5-23.8.5-2 2.8 20.4-4.7 40.2-2.4 6.4-6.3 15.4-6.3 15.4s5.9-7.3 10-14.8c4.8-8.9 2.5 12-.3 19.3-4.1 10.4-8.3 16.4-8.3 16.4s2.1 1 6.3-2.5c2-1.6 2.2 3.3-.1 10.2-1.5 4.5-8.2 11.6-8.2 11.6s6.1-1.1 8.9-3.9c1-1 1.2 3-1.4 9-3.6 8.5-15.4 18.9-15.4 18.9s5 .5 10.2-3.7c2.1-1.7-.7 6.6-4.3 10.6-4.8 5.3-10.7 7.9-10.7 7.9s10.4.3 1.6 10.3c-6.3 7.1-15 7.4-15 7.4s-1.7 13.4-14.7 10.8c0 0-7.3 14.7-20.3 11.3 0 0-2.5-.5-2.4-2.4.3-3.4 5.2-6.9 5.2-12.9 0-4-2.9-7.2-2.9-7.2m48.2-141.8s5.4-10.1 8.2-24m-6.3 56.8c8-11.5 12.1-20.9 12.1-20.9m-9.9 46.6c6.2-7.8 9.5-12.9 9.5-12.9"
-				opacity=".8"
-				fill="#fff"
-			/>
-			<path
-				className="st0"
-				d="M218.2 260.8h-25c-.6 0-1-.4-1-1v-42.5H186c-9.6 0-14.1 11.2-20.4 26.8-2 5-4.3 10.6-6.9 16.2-.2.4-.5.6-.9.6h-27.6c-.3 0-.7-.2-.8-.5s-.2-.6-.1-1c3.4-7.5 6.1-13.9 8.5-19.6 7-16.6 10.8-25.4 17.9-30.1-10.7-5.8-16.2-14.7-16.2-26.3 0-8.4 3.2-16.3 9-22.1 6.3-6.3 14.9-9.7 24.9-9.7h21.2c.7-2.6 2.5-5.6 6.2-8.9 5.7-4.9 7.5-11.3 10.1-20.1 3.7-12.7 8.7-30.2 29.4-53.4 15.8-17.7 20.4-39.7 13.7-65.4-.7-2.6-.8-3.3-.1-3.8.3-.2.8-.3 1.3.1 9.3 7.2 14.7 19.7 15.6 36.2.2 3.2.1 6.2 0 9 1.5-6 2.9-12.7 3.3-16.4.1-1 .3-2.1 1.2-2.5.5-.2 1-.1 1.5.2 2.8 2 6.6 19 3.8 42.7-.7 6.1-1.4 10.7-2.3 14.7 2.6-5.1 5.2-10.5 5.7-12.7.2-.6.6-.8.8-.9.4-.1.8.1 1 .3 1.2 1.4 2.3 23.1-4.6 41.3-.9 2.4-2.1 5.3-3.1 7.9 1.7-2.4 3.5-5.3 5-8.1.5-1 1.8-3.3 3.3-2.6 3.1 1.4-.1 18.3-1.9 22.8-2.9 7.4-5.8 12.6-7.4 15 .9-.2 2.1-.8 3.8-2.2 1.1-.9 1.9-.5 2.2-.3 1.9 1.3.8 7-.7 11.5-1 3-4.1 7-6.2 9.5 1.8-.6 3.9-1.5 5.3-2.8.6-.6 1.3-.5 1.7-.2 1.7 1.1 0 6.8-1.5 10.3-2.8 6.4-9.9 13.8-13.5 17.2 1.8-.4 4.3-1.2 6.8-3.2 1.1-.9 1.8-.2 2-.1 1.5 1.6-2.4 9.3-4.9 12.1-3 3.3-6.3 5.5-8.5 6.8 1.2.4 2.3 1.2 2.9 2.5 1 2.2-.2 5.1-3.5 8.8-5.4 6.1-12.5 7.4-14.9 7.6-.4 1.8-1.7 6.4-5.4 9.1-2.6 1.9-5.8 2.5-9.6 1.9-1.3 2.3-6 9.6-13.7 11.3v12.7c-.2.9-.6 1.3-1.2 1.3zm-24-2h23v-11.3c-1.7.2-3.5 0-5.4-.5-1.1-.2-3.3-1.2-3.1-3.5.2-1.6 1.2-3.2 2.2-4.8 1.4-2.2 3-4.7 3-8 0-3.6-2.6-6.5-2.6-6.6-.4-.4-.3-1 .1-1.4s1-.3 1.4.1c.1.1 3.1 3.5 3.1 7.9 0 3.9-1.9 6.7-3.4 9.1-1 1.5-1.8 2.8-1.9 3.9-.1 1 1.5 1.3 1.5 1.3 12.2 3.1 19.2-10.7 19.2-10.8.2-.4.7-.6 1.1-.5 3.5.7 6.5.2 8.8-1.4 4-2.9 4.8-8.5 4.8-8.5.1-.5.5-.9 1-.9.1 0 8.4-.4 14.3-7 2.6-3 3.7-5.3 3.1-6.7s-3.1-1.8-4-1.9c-.5 0-.9-.4-.9-.8-.1-.5.2-.9.6-1.1.1 0 5.8-2.6 10.4-7.6 2.2-2.4 3.9-6.2 4.5-8.3-5.1 3.5-9.6 3.1-9.8 3-.4 0-.7-.3-.9-.7-.1-.4 0-.8.3-1.1.1-.1 11.7-10.3 15.2-18.5 1.4-3.2 1.9-5.6 2-7-3.2 2.4-8.1 3.3-8.4 3.3-.4.1-.9-.1-1.1-.5s-.1-.8.1-1.2c.1-.1 6.5-7 8-11.3 1.6-4.7 1.8-7.8 1.5-8.9-4.5 3.6-6.9 2.6-7.2 2.5s-.5-.4-.5-.6c-.1-.3 0-.6.1-.8 0-.1 4.2-6.1 8.2-16.2 2.2-5.7 3.5-16.9 2.9-19.8-.2.3-.4.6-.8 1.3-3.9 7.2-9.5 14.3-10 14.9-.3.5-3.7 5.5-9.5 12.9-.3.4-1 .5-1.4.2s-.5-1-.2-1.4c5.7-7.2 9-12.1 9.4-12.7.4-.9 3.9-9.4 6.2-15.3 5.3-14 5.7-29.3 5.3-36-3.1 7.1-9.9 19.8-10.2 20.5-1.7 3.4-5.5 10.7-11.3 19-.3.5-.9.6-1.4.3s-.6-.9-.3-1.4c5.7-8.3 9.5-15.5 11.1-18.9 3.2-7.5 4.6-13.3 6.1-25.9 2.6-22.3-.6-36.8-2.5-40.2v.2c-.8 8.3-6.6 29.3-6.8 30.2 0 .1 0 .1-.1.2-2.9 13.1-7.9 22.5-7.9 22.6-.3.5-.9.7-1.4.4s-.7-.9-.4-1.4c.1-.1 5.3-10.1 8.1-23.7.5-3.1 1.7-11.5 1.2-21-.9-17.9-7.1-27.8-12.9-33 1.2 4.7 3.7 14.5 3 26.6-1 15.2-6.7 28.8-17.1 40.5-20.4 22.9-25.4 40.1-29 52.6-2.5 8.8-4.5 15.7-10.7 21-9.3 8.1-5.1 14.3-4.9 14.5.3.5.2 1.1-.2 1.4s-1.1.2-1.4-.2c0 0-1.7-2.5-1.4-6.3h-20.8c-20.9 0-31.9 15-31.9 29.8 0 11.4 5.6 19.9 16.7 25.4.3.2.6.5.6.9s-.2.7-.5.9c-7.5 4-11.2 12.7-18.6 30-2.3 5.3-4.8 11.3-7.9 18.2H157c2.5-5.4 4.6-10.5 6.6-15.5 6.4-15.6 11.4-28 22.3-28h7.2c.6 0 1 .4 1 1v42.5zm88.1-112.3zm-89.1 51.2h-13.3c-9.2 0-14.7-5-14.7-13.3 0-6 2.7-13.1 15.3-13.1h12.7c.6 0 1 .4 1 1v24.4c0 .5-.5 1-1 1zm-12.7-24.4c-8.9 0-13.3 3.6-13.3 11.1 0 10.2 8.9 11.3 12.7 11.3h12.3v-22.4h-11.7z"
-			/>
-			<path
-				className="st1"
-				d="M227.7 138.4c4.2-.3 9.2-6.7 9.2-6.7 4.1 3.7-2 13.9-2 13.9s5.1-2.9 8.6-5.4c1.1-.8-.1 13-5.6 21.4 0 0 4-.4 8.3-3.3 3-2.1-.7 16.2-11.1 27.3"
-			/>
-			<path
-				className="st0"
-				d="M235 186.6c-.2 0-.5-.1-.7-.3-.4-.4-.4-1 0-1.4 8.3-8.9 12-22.1 11.8-25.4-4.2 2.7-8.1 3.1-8.2 3.1-.4 0-.8-.2-.9-.5-.2-.3-.2-.7 0-1.1 4.2-6.4 5.8-15.8 5.9-19.2-3.3 2.3-7.3 4.5-7.5 4.6-.4.2-.9.2-1.2-.2-.3-.3-.4-.8-.1-1.2 1.4-2.4 4.4-8.7 2.8-11.8-1.7 1.9-5.6 5.9-9.1 6.1-.5 0-1-.4-1.1-.9 0-.6.4-1 .9-1.1 3.1-.2 7.2-4.7 8.5-6.4.2-.2.4-.4.7-.4s.6.1.8.3c3.1 2.9 1.4 8.5-.2 12.1 1.7-1 3.8-2.3 5.5-3.6.5-.3 1-.3 1.4.1 1.4 1.4-.3 12.9-4.5 20.7 1.5-.4 3.6-1.2 5.7-2.7.9-.6 1.7-.3 2.1.2 2.1 2.5-2.8 19-11.8 28.6-.3.3-.5.4-.8.4zm9-45.6z"
-			/>
-		</svg>
-	);
-};
