--- src/components/iconSet/ts/logos/IconSKF.tsx
@@ -1,35 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconSKF = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			viewBox="0 0 44 60"
-			xmlns="http://www.w3.org/2000/svg"
-			xmlnsXlink="http://www.w3.org/1999/xlink"
-		>
-			<defs>
-				<path id="a-1" d="M.222.237H31.91v12.948H.222z" />
-			</defs>
-			<g fill="none" fillRule="evenodd">
-				<path d="M-9 0h60v60H-9z" />
-				<g transform="translate(5.946 46.8)">
-					<mask id="b-1" fill="#fff">
-						<use xlinkHref="#a" />
-					</mask>
-					<path
-						d="M23.117 12.973V.464h8.792v2.114h-6.2v2.97h5.36v2.114h-5.36v5.31h-2.592zm-10.312 0V.464h2.448v6.64l2.877-3.198h3.02l-3.18 3.32 3.412 5.747h-2.644l-2.36-4.088-1.125 1.188v2.9h-2.448zM.222 8.92l2.52-.245c.143.821.464 1.433.911 1.817.465.384 1.09.577 1.877.577.822 0 1.447-.175 1.876-.524.411-.333.626-.734.626-1.206 0-.28-.09-.542-.268-.751-.18-.21-.483-.385-.93-.542-.286-.105-.983-.28-2.055-.541-1.376-.332-2.341-.752-2.895-1.24-.786-.7-1.18-1.538-1.18-2.534 0-.629.18-1.223.555-1.782.375-.559.893-.978 1.59-1.275C3.546.377 4.386.237 5.351.237c1.608 0 2.806.35 3.628 1.048.804.682 1.233 1.625 1.269 2.778l-2.592.087c-.107-.646-.34-1.117-.697-1.397-.375-.28-.911-.42-1.626-.42-.75 0-1.34.158-1.751.455a.924.924 0 00-.411.786c0 .297.125.541.375.75.34.28 1.126.56 2.377.84 1.25.297 2.18.593 2.788.908a3.78 3.78 0 011.411 1.275c.34.525.5 1.188.5 1.992 0 .716-.196 1.38-.607 2.01-.41.628-.983 1.082-1.733 1.397-.75.296-1.662.454-2.788.454-1.608 0-2.86-.367-3.717-1.1C.9 11.382.383 10.316.222 8.92z"
-						fill="#FFF"
-						mask="url(#b)"
-					/>
-				</g>
-				<path
-					d="M17.716 30.691c-.461.221-.99.542-1.537 1.151-.49.545-.712 1.017-1.033 1.643a15.53 15.53 0 01-2.024-1.17c.469-.946 1.13-1.926 1.973-2.745.76-.74 1.682-1.357 2.62-1.665v2.786zm6.664 6.636l.368-.06c7.078-1.14 12.795-6.261 14.654-13.276H44C41.847 34.28 32.812 42 22 42 11.18 42 2.142 34.268 0 23.967h4.608c1.852 6.985 7.648 12.09 14.64 13.278l.371.063V21.667H5.791l-3.333-4.815H19.62V0h4.761v16.852h17.163l-3.332 4.815H24.38v15.66zm1.945-9.16c.529.956 1.255 1.728 2.158 2.498.948.81 1.804 1.357 2.458 1.673-.69.5-1.367.89-2.133 1.272-.472-.256-1.007-.659-1.421-1.012a8.101 8.101 0 01-1.06-1.111l-.002-3.32zm7.645 1.375a16.29 16.29 0 01-1.614 1.678c-.746-.37-2.141-1.243-3.524-2.642-1.369-1.385-2.475-3.108-2.507-4.99h2.437c.237 1.361.994 2.611 1.985 3.647 1 1.045 2.321 1.981 3.223 2.307zm-.623-5.954c.219.466.536 1 1.138 1.554.538.495 1.005.72 1.624 1.045a15.822 15.822 0 01-1.156 2.048 10.065 10.065 0 01-2.714-1.997c-.731-.769-1.341-1.701-1.646-2.65h2.754zM30.52 2.804c.4-.418.869-.704 1.644-.897a18.11 18.11 0 011.894 1.331c-.207.074-1.47.23-2.85 2.402-.74 1.164-1.183 1.972-1.919 2.935-.762.998-1.77 1.838-2.998 2.062V8.092c.972-.227 1.45-.966 1.944-1.865.608-1.105 1.43-2.533 2.285-3.423zm7.596 4.623c.323.432.798 1.15 1.172 1.8a3.42 3.42 0 00-1.127.64c-1.344 1.133-1.863 2.407-2.775 3.643-.342.463-.978 1.103-1.433 1.376h-3.203c1.194-.389 1.986-1.419 2.935-2.701.939-1.268 1.89-2.854 2.636-3.554.597-.559.946-.863 1.795-1.204zm-2.69-3.061a21.792 21.792 0 011.56 1.63c-.7.199-1.801 1.075-2.284 1.692-1.496 1.91-3.165 4.483-4.097 5.311-.924.821-2.307 1.735-4.314 1.887V12.41c3.947-.481 4.899-4.07 6.96-6.502.592-.7 1.607-1.514 2.174-1.542zM5.771 10.607c-.413-.405-.696-.879-.886-1.663a18.33 18.33 0 011.316-1.916c.073.21.227 1.486 2.374 2.884 1.15.748 1.95 1.196 2.902 1.94.986.771 1.816 1.79 2.038 3.034L11 14.885c-.225-.983-.955-1.466-1.844-1.967-1.092-.614-2.504-1.447-3.384-2.31zm4.57-7.684a19.772 19.772 0 011.78-1.186c.057.185.236.658.633 1.14 1.12 1.36 2.379 1.885 3.6 2.808.458.346 1.091.988 1.36 1.45v3.24c-.384-1.209-1.402-2.01-2.67-2.97-1.253-.95-2.82-1.911-3.512-2.667-.552-.603-.854-.956-1.19-1.815zM7.318 5.645c.493-.534 1.065-1.101 1.612-1.578.196.707 1.062 1.82 1.672 2.31 1.887 1.512 4.431 3.201 5.25 4.144.812.935 1.715 2.334 1.865 4.365h-2.448c-.475-3.993-4.023-4.957-6.427-7.041-.691-.6-1.496-1.626-1.524-2.2zm5.872 17.943c-.945.534-1.709 1.27-2.47 2.182-.8.96-1.341 1.826-1.654 2.486A14.968 14.968 0 017.808 26.1c.254-.478.652-1.019 1-1.438a8.1 8.1 0 011.099-1.072l3.282-.001zm-1.36 7.733a16.276 16.276 0 01-1.658-1.633c.365-.754 1.229-2.166 2.611-3.565 1.37-1.384 3.073-2.503 4.934-2.535v2.465c-1.347.24-2.582 1.005-3.606 2.008-1.033 1.011-1.959 2.348-2.28 3.26z"
-					fill="#FFF"
-				/>
-			</g>
-		</svg>
-	);
-};
