--- src/components/iconSet/ts/logos/IconSKM.tsx
@@ -1,38 +1,0 @@
-import * as React from 'react';
-import { IconProps } from '../Icon';
-
-export const IconSKM = (props: IconProps) => {
-	return (
-		<svg
-			className={props.className ? props.className : null}
-			name={props.name}
-			viewBox="0 0 60 46"
-			xmlns="http://www.w3.org/2000/svg"
-		>
-			<g fill="none" fillRule="evenodd">
-				<path d="M0-7h60v60H0z" />
-				<path
-					d="M60 43.473C60 44.869 58.887 46 57.514 46H2.487C1.113 46 0 44.869 0 43.473V2.527C0 1.132 1.113 0 2.487 0h55.027C58.887 0 60 1.132 60 2.527v40.946"
-					fill="#fff"
-				/>
-				<path
-					d="M25.909 11.835s1.131-1.268 2.332-3.372c.91-1.595 3.122-3.05 3.194-3.051.257-.008 2.5 1.627 2.655 1.795.05.055-1.357.751-2.398 1.864-.784.838-1.255 1.545-2.62 3.749-.545.879-1.224 1.726-1.643 2.125-.418.399-1.52 1.29-1.52 1.29v-4.4"
-					fill="#CC1E1C"
-				/>
-				<path
-					d="M25.909 18.941v-2.305s1.624-1.16 2.312-1.904c.694-.752 1.51-2.005 2.024-2.856.514-.852 1.423-2.22 1.889-2.823.71-.919 2.447-2.288 2.569-2.288.211 0 2.054 2.088 2.115 2.254.024.066-1.496 1.12-2.055 1.82-.523.657-1.453 2.054-1.662 2.44-.363.668-1.073 1.87-1.48 2.388-.373.473-1.21 1.387-1.769 1.854-.367.307-1.662 1.42-1.662 1.42H25.91M30 18.941s1.473-1.144 2.039-1.67c.741-.689 1.451-1.715 2.023-2.619.512-.81 1.158-1.777 1.668-2.297.541-.552 2.008-1.516 2.162-1.532.155-.015 1.635 2.33 1.653 2.42.015.077-.958.797-1.622 1.64-.561.711-1.328 1.792-1.698 2.297-.372.505-.523.63-.834.965-.387.413-.897.78-.897.78L30 18.942M9.756 18.941s-1.119-1.123-2.975-2.313c-1.406-.903-2.688-3.098-2.69-3.17-.007-.255 1.435-2.481 1.583-2.633.049-.05.663 1.345 1.644 2.378.74.778 1.363 1.245 3.306 2.6.776.54 1.522 1.214 1.875 1.63.351.415 1.137 1.508 1.137 1.508h-3.88M17.727 18.941h-2.065s-1.04-1.813-1.706-2.58c-.673-.776-1.796-1.687-2.559-2.26-.763-.574-1.99-1.589-2.529-2.108-.823-.793-2.05-2.732-2.05-2.867 0-.236 1.87-2.294 2.02-2.361.059-.027 1.003 1.67 1.631 2.294.588.583 1.84 1.62 2.185 1.854.599.405 1.676 1.198 2.14 1.653.423.416 1.242 1.35 1.66 1.973.276.41 1.273 1.855 1.273 1.855v2.547"
-					fill="#CC1E1C"
-				/>
-				<path
-					d="M17.727 14.882s-1.153-1.461-1.683-2.023c-.694-.735-1.729-1.44-2.64-2.007-.816-.509-1.79-1.15-2.315-1.655-.556-.537-1.528-1.993-1.544-2.146-.015-.153 2.347-1.622 2.44-1.64.076-.015.802.95 1.651 1.61.718.556 1.806 1.318 2.315 1.685.51.368.637.518.973.828.417.383.788.889.788.889l.015 4.459M2.727 27.059h39.546v-6.765H2.727z"
-					fill="#CC1E1C"
-				/>
-				<path fill="#CC1E1C" d="M19.091 43.294h6.818V4.059h-6.818z" />
-				<path
-					d="M36.852 42.623L31.9 35.424l4.952-7.23v14.43zm12.93-15.564l-2.128 7.762-2.143-7.762H29.237l-4.691 8.928 4.69 7.307h14.232V35.15l2.275 8.144h3.822l2.423-8.144v8.144h6.65V27.06h-8.855zM18.626 27.727s-9.081-1.726-12.282.37c-2.351 1.54-2.253 3.804-2.253 5.16 0 1.913.99 3.69 3.1 4.078 3.88.714 5.308-.119 5.308.924 0 .93-.637.791-2.927.655-2.052-.122-5.118-.686-5.118-.686l-.005 4.28s5.04 1.133 9.253.679c3.081-.332 5.389-2.075 5.389-5.189 0-2.137-.433-4.601-3.693-4.712-3.96-.134-4.459.062-4.459-.816 0-.76.623-.648 3.68-.579 2.071.047 4.007.394 4.007.394v-4.558"
-					fill="#CC1E1C"
-				/>
-			</g>
-		</svg>
-	);
-};
