+++ src/components/svgSet/ts/SVGHelpers.tsx
@@ -1,0 +1,574 @@
+import * as React from 'react';
+
+export const ICON_KEYS = {
+	ENVELOPE: 'ENVELOPE',
+	LOCK: 'LOCK',
+	PERSON: 'PERSON',
+	PIN: 'PIN'
+};
+
+export const LOGO_KEYS = {
+	CARITAS_WHITE: 'CARITAS_WHITE',
+	IN_VIA: 'IN_VIA',
+	KREUZBUND: 'KREUZBUND',
+	RAPHAELSWERK: 'RAPHAELSWERK',
+	SKF: 'SKF',
+	SKM: 'SKM'
+};
+
+export const DEFAULT_SVG_VALUES = {
+	ICON_FILL: '#3f373f',
+	ICON_SIZE: '24',
+	ICON_VIEWBOX: '0 0 72 72',
+	LOGO_FILL: '#fffff',
+	LOGO_SIZE: '60'
+};
+
+//DEV-NOTE: add specific viewBox values for any SVGs that don't have the square ratio of the standard icons
+export const getSVGViewBox = (SVGName: string) => {
+	switch (SVGName) {
+		case LOGO_KEYS.SKF:
+			return '0 0 44 60';
+		case LOGO_KEYS.CARITAS_WHITE:
+			return '0 0 60 60';
+		case LOGO_KEYS.SKM:
+			return '0 0 60 46';
+		case LOGO_KEYS.IN_VIA:
+			return '0 0 54 60';
+		case LOGO_KEYS.KREUZBUND:
+			return '0 0 62 60';
+		case LOGO_KEYS.RAPHAELSWERK:
+			return '0 0 391 391';
+		default:
+			return DEFAULT_SVG_VALUES.ICON_VIEWBOX;
+	}
+};
+
+export const getSVGPath = (SVGName: string) => {
+	switch (SVGName) {
+		case ICON_KEYS.ENVELOPE:
+			return (
+				<path
+					d="M65.9963 28V55.4658C65.9963 56.9883 65.4719 58.2913 64.4232 59.3748C63.3745 60.4583 62.1134 61 60.6397 61H11.3565C9.88288 61 8.62171 60.4583 7.57303 59.3748C6.52434 58.2913 6 56.9883 6 55.4658V28C6.98244 29.13 8.10987 30.1334 9.38229 31.0102C17.4618 36.6833 23.0083 40.6614 26.0219 42.9445C27.2943 43.9131 28.3267 44.6685 29.1192 45.2109C29.9116 45.7533 30.9666 46.3066 32.284 46.871C33.6014 47.4353 34.8288 47.7181 35.9663 47.7194H36.0337C37.1724 47.7194 38.3999 47.4366 39.716 46.871C41.0322 46.3054 42.0871 45.752 42.8808 45.2109C43.6745 44.6698 44.707 43.9144 45.9781 42.9445C49.7729 40.1087 55.3307 36.1305 62.6515 31.0102C63.9239 30.1114 65.0401 29.108 66 28H65.9963ZM66 18.36C66 20.1235 65.4532 21.8083 64.3595 23.4143C63.2658 25.0204 61.904 26.3927 60.2741 27.5313C51.8821 33.3568 46.6593 36.9839 44.6057 38.4124C44.382 38.5687 43.9076 38.9093 43.1827 39.4342C42.4577 39.9591 41.8553 40.3835 41.3753 40.7072C40.8953 41.0309 40.3147 41.3933 39.6335 41.7945C38.9523 42.1957 38.3105 42.4969 37.708 42.6982C37.1056 42.8994 36.5475 43 36.0337 43H35.9663C35.4525 43 34.8944 42.8994 34.292 42.6982C33.6895 42.4969 33.0477 42.1957 32.3665 41.7945C31.6853 41.3933 31.1047 41.0309 30.6247 40.7072C30.1447 40.3835 29.5423 39.9591 28.8173 39.4342C28.0924 38.9093 27.618 38.5687 27.3943 38.4124C25.3632 36.9839 22.439 34.9473 18.6217 32.3026C14.8045 29.6579 12.5165 28.0675 11.7578 27.5313C10.3741 26.5939 9.06856 25.3047 7.84114 23.6637C6.61371 22.0226 6 20.4991 6 19.093C6 17.352 6.4631 15.9015 7.38929 14.7417C8.31548 13.5818 9.6379 13.0012 11.3565 13H60.6397C62.0909 13 63.3464 13.5243 64.4063 14.5729C65.4663 15.6215 65.9963 16.8826 65.9963 18.3562L66 18.36Z"
+					fill="#3F373F"
+				/>
+			);
+		case ICON_KEYS.LOCK:
+			return (
+				<>
+					<defs>
+						<path
+							id="lock-a"
+							d="M48,27.9898836 L48,24.9695747 C48,17.8182095 42.6167461,12 36,12 C29.3832539,12.0088288 24,17.8270383 24,24.9784036 L24,27.9979767 L18,28 L18,23.8453049 C18,14.0121777 26.0748809,6.01213966 36,6 C45.9251191,6 54,14.0000381 54,23.8331653 L54,27.9878603 L48,27.9898836 Z M10,28 L62,28 C63.1045695,28 64,28.8954305 64,30 L64,65 C64,66.1045695 63.1045695,67 62,67 L10,67 C8.8954305,67 8,66.1045695 8,65 L8,30 C8,28.8954305 8.8954305,28 10,28 Z M35,39 C33.8954305,39 33,39.8954305 33,41 L33,55 C33,56.1045695 33.8954305,57 35,57 L37,57 C38.1045695,57 39,56.1045695 39,55 L39,41 C39,39.8954305 38.1045695,39 37,39 L35,39 Z"
+						/>
+					</defs>
+					<use xlinkHref="#lock-a" />
+				</>
+			);
+		case ICON_KEYS.PERSON:
+			return (
+				<>
+					<defs>
+						<path
+							id="person-a"
+							d="M35.9230769,28 C33,28 30.3846154,26.9305556 28.2307692,24.7916667 C26.0769231,22.6527778 25,20.0555556 25,17 C25,13.9444444 26.0769231,11.5 28.0769231,9.20833333 C30.2307692,7.06944444 32.8461538,6 35.9230769,6 C39,6 41.6153846,7.06944444 43.7692308,9.20833333 C45.9230769,11.3472222 47,13.9444444 47,17 C47,20.0555556 45.9230769,22.6527778 43.7692308,24.7916667 C41.6153846,26.9305556 39,28 35.9230769,28 Z M27.5290807,32.8491306 C29.7354597,35.1432731 32.4146341,36.2903443 35.4090056,36.2903443 C38.5609756,36.2903443 41.2401501,35.1432731 43.4465291,32.8491306 C43.711905,32.5731988 43.9613216,32.2901553 44.194779,32 L46.7560976,32 C52.4136487,32 57,36.8021314 57,42.7258607 L57,58.8146517 C50.747518,63.6048839 43.373759,66 36,66 C28.626241,66 21.252482,63.6048839 15,58.8146517 L15,42.7258607 C15,36.8021314 19.5863513,32 25.2439024,32 L26.7808308,32 C27.0142881,32.2901553 27.2637048,32.5731988 27.5290807,32.8491306 Z"
+						/>
+					</defs>
+					<use xlinkHref="#person-a" />
+				</>
+			);
+		case ICON_KEYS.LOCK:
+			return (
+				<>
+					<defs>
+						<path
+							id="pin-a"
+							d="M46.6382979,25.4366197 C46.6382979,22.6368502 45.6410767,20.246767 43.6466344,18.26637 C41.6521921,16.2859731 39.245158,15.2957746 36.4255319,15.2957746 C33.6059059,15.2957746 31.1988717,16.2859731 29.2044294,18.26637 C27.2099871,20.246767 26.212766,22.6368502 26.212766,25.4366197 C26.212766,28.2363892 27.2099871,30.6264725 29.2044294,32.6068694 C31.1988717,34.5872663 33.6059059,35.5774648 36.4255319,35.5774648 C39.245158,35.5774648 41.6521921,34.5872663 43.6466344,32.6068694 C45.6410767,30.6264725 46.6382979,28.2363892 46.6382979,25.4366197 Z M55.9956251,26.0007276 C55.9956251,28.838703 55.5661526,31.169168 54.7072077,32.9921224 L40.4886799,63.2261861 C40.071603,64.0851623 39.4532794,64.7625721 38.6337088,65.2584157 C37.8141383,65.7542593 36.9355062,66.0014519 35.9978125,65.9999936 C35.0601189,65.9985352 34.1814867,65.7513426 33.3619162,65.2584157 C32.5423457,64.7654888 31.9371468,64.088079 31.5463196,63.2261861 L17.2884174,32.9921224 C16.4294725,31.169168 16,28.838703 16,26.0007276 C16,20.4793632 17.953407,15.7659323 21.8602209,11.8604347 C25.7670349,7.95493714 30.4802946,6.00145918 36,6.00000082 C41.5197054,5.99854245 46.2329651,7.95202041 50.1397791,11.8604347 C54.046593,15.768849 56,20.48228 56,26.0007276 L55.9956251,26.0007276 Z"
+						/>
+					</defs>
+					<use xlinkHref="#pin-a" />
+				</>
+			);
+		case LOGO_KEYS.CARITAS_WHITE:
+			return (
+				<>
+					<path
+						d="M52.293 56.3469V3.64317C52.293 2.02574 50.923 0.715149 49.2335 0.715149H10.7662C9.07644 0.715149 7.70679 2.02574 7.70679 3.64317V56.3469C7.70679 57.9638 9.07644 59.2748 10.7662 59.2748H49.2335C50.923 59.2748 52.293 57.964 52.293 56.3469Z"
+						fill="#CC1E1C"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M16.8363 52.1455C16.5304 51.0363 15.7585 50.3737 14.5939 50.3737C12.803 50.3737 12.0603 51.9867 12.0603 53.5427C12.0603 55.1133 12.7886 56.6828 14.5939 56.6828C15.8897 56.6828 16.705 55.9192 16.9089 54.6807H17.8408C17.5349 56.337 16.4867 57.4608 14.5939 57.4608C12.1914 57.4608 11.1285 55.7469 11.1285 53.5427C11.1285 51.4687 12.337 49.5955 14.5939 49.5955C16.3266 49.5955 17.4623 50.417 17.7244 52.1455H16.8363Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M23.4464 53.3411C22.7035 53.6438 21.9467 53.7015 21.1894 53.7587C19.8641 53.8596 19.0635 54.3355 19.0635 55.3437C19.0635 56.2509 19.8641 56.6828 20.6945 56.6828C22.0774 56.6828 23.4464 56.0205 23.4464 54.5079V53.3411ZM21.9756 52.9089C22.296 52.8801 22.8347 52.7937 23.1257 52.6639C23.4171 52.5341 23.4464 52.1455 23.4464 51.886C23.4464 51.0076 22.9223 50.3737 21.568 50.3737C20.3594 50.3737 19.5295 50.7625 19.3693 52.0591H18.4813C18.6125 50.302 19.937 49.5955 21.5827 49.5955C23.1113 49.5955 24.3198 50.1864 24.3198 51.9149V55.9771C24.3198 56.6544 24.4946 56.8127 25.2371 56.5531V57.2305C25.106 57.2739 24.8002 57.3747 24.5529 57.3747C24.4071 57.3747 24.2908 57.3601 24.1595 57.3311C23.5772 57.245 23.4752 56.7404 23.4608 56.2225C22.7329 57.0286 21.7428 57.4609 20.6506 57.4609C19.3259 57.4609 18.1317 56.7551 18.1317 55.3144C18.1317 54.061 19.0053 53.2979 20.7672 53.0674L21.9756 52.9089Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M26.6709 57.2596H25.7975V49.7976H26.6709V51.195H26.7001C27.1804 49.9994 27.8939 49.5955 29.1606 49.5955V50.5326C27.1953 50.4887 26.6709 51.9296 26.6709 53.6004V57.2596Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M29.834 57.2596H30.7078V49.7976H29.834V57.2596Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M34.8532 57.2878C34.6494 57.3311 34.2705 57.4609 33.8483 57.4609C32.9895 57.4609 32.5082 57.0865 32.5082 55.9771V50.5182H31.4455V49.7976H32.5082V47.723H33.382V49.7976H34.7946V50.5182H33.382V55.6317C33.382 56.3664 33.4113 56.6828 34.1393 56.6828C34.373 56.6828 34.6196 56.6398 34.8532 56.5822V57.2878Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M40.1689 53.3411C39.426 53.6438 38.6686 53.7015 37.9119 53.7587C36.5867 53.8596 35.7856 54.3355 35.7856 55.3437C35.7856 56.2509 36.5867 56.6828 37.417 56.6828C38.8 56.6828 40.1689 56.0205 40.1689 54.5079V53.3411ZM38.6981 52.9091C39.0187 52.8802 39.5573 52.7937 39.8486 52.6639C40.1396 52.5341 40.1689 52.1455 40.1689 51.886C40.1689 51.0077 39.6443 50.3737 38.2898 50.3737C37.0819 50.3737 36.252 50.7625 36.0919 52.0591H35.2033C35.335 50.302 36.6597 49.5955 38.3043 49.5955C39.8338 49.5955 41.0422 50.1864 41.0422 51.9149V55.9771C41.0422 56.6544 41.2172 56.8127 41.9596 56.5533V57.2305C41.8287 57.2739 41.5229 57.3747 41.2759 57.3747C41.1291 57.3747 41.0133 57.3601 40.8821 57.3311C40.2992 57.245 40.1978 56.7404 40.1833 56.2225C39.4549 57.0289 38.4657 57.4609 37.3728 57.4609C36.0481 57.4609 34.8542 56.7552 34.8542 55.3144C34.8542 54.061 35.7284 53.2979 37.4891 53.0674L38.6981 52.9091Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M42.8174 54.9255C43.0943 56.2079 43.7643 56.6828 45.0745 56.6828C46.3416 56.6828 46.9377 56.0205 46.9377 55.3002C46.9377 53.1397 42.1043 54.6669 42.1043 51.7283C42.1043 50.4889 43.0794 49.5957 44.7537 49.5957C46.3848 49.5957 47.4479 50.2582 47.6663 51.9442H46.778C46.618 50.7914 45.8318 50.3737 44.7393 50.3737C43.633 50.3737 43.0357 50.8919 43.0357 51.6561C43.0357 53.6871 47.8697 52.4196 47.8697 55.2858C47.8697 56.712 46.5743 57.4609 44.7979 57.4609C43.0218 57.4609 42.0025 56.337 41.9293 54.9255H42.8174Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M50.0154 20.0608H33.6822V3.63367H26.5762V20.0608H10.2425V27.207H26.5762V43.6346H33.6822V27.207H50.0154V20.0608Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M23.7587 19.1935H25.7179V16.8122C25.2007 16.135 24.6469 15.4464 24.0694 14.8196C22.8073 13.4208 21.1587 12.5398 19.6137 11.5036C18.5305 10.6879 17.5135 9.44111 16.7441 8.38745C16.008 9.04175 15.3194 9.749 14.6846 10.5036C14.8161 10.7302 15.0387 11.0893 15.415 11.6592C16.4346 13.203 17.965 14.0425 19.4848 15.0009C19.9226 15.234 20.3347 15.5968 20.7467 15.9076C21.9255 16.846 22.899 18.0841 23.7587 19.1935Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M22.1377 19.1935C21.7018 18.4049 20.9499 17.5039 20.3604 16.8662C18.5571 15.234 16.0846 14.6902 14.7454 12.4881C14.3691 11.8697 14.1524 11.5169 14.0277 11.3264C13.4123 12.1395 12.8556 13 12.366 13.9022C13.289 15.4614 14.4785 16.8517 16.1362 17.695C17.3984 18.3373 18.0675 18.9322 18.3267 19.1935H22.1377Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M43.2838 8.18558C42.6517 8.64494 42.0438 9.0465 41.482 9.58659C40.1424 10.7008 39.3679 12.2632 38.3648 13.6796C37.405 15.0364 35.9256 16.1719 34.5074 17.0232V19.1935H37.0149C38.1688 18.4079 39.4317 17.2781 40.297 16.2703C41.7134 14.3016 42.769 11.8924 44.8562 10.623C45.0684 10.4938 45.2444 10.3838 45.3918 10.2893C44.7412 9.53734 44.0364 8.83436 43.2838 8.18558Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M48.0498 33.0315C47.1482 31.8391 46.247 30.7255 44.9591 29.9225C44.0495 29.3928 43.0525 28.7018 42.1509 28.066H38.2744L38.6228 28.6271C39.8846 30.3623 41.7394 31.3471 43.4137 32.5384C44.6119 33.3337 45.6586 34.5186 46.4085 35.7047C47.0229 34.8685 47.5746 33.983 48.0572 33.0557L48.0498 33.0315Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M45.3198 36.0362C44.0505 33.998 41.7394 32.8239 39.9101 31.5024C38.5947 30.5542 37.6632 29.3354 36.8134 28.066H34.5074V30.1127C35.5155 31.5994 37.0205 33.359 38.4681 34.2224C40.3644 35.3825 42.2245 36.9273 43.7144 38.7036C44.4206 38.059 45.0801 37.3651 45.6903 36.6277C45.5906 36.4653 45.4669 36.2724 45.3198 36.0362Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M23.2462 28.066C21.6052 29.2323 20.1595 30.6653 19.0468 32.3312C17.965 34.2224 16.7542 35.7769 14.977 37.0461L14.9558 37.0821C15.5995 37.8162 16.2949 38.5028 17.0367 39.1371C17.1975 39.0308 17.3846 38.8975 17.6043 38.7298C19.2893 37.4461 20.618 35.9065 21.6227 34.197C22.6388 32.6009 24.18 31.1987 25.7179 30.0795V28.066H23.2462Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M16.9896 28.066C15.9761 29.0502 15.3812 30.0039 14.6938 31.1396C14.185 32.0202 13.328 32.8957 12.4345 33.4911C12.931 34.3906 13.4928 35.2478 14.1142 36.0567L15.1829 35.3108C15.6209 34.9998 16.0846 34.6113 16.3936 34.197C17.218 33.3156 17.8103 32.15 18.5316 31.1917C19.3301 30.0777 20.2831 29.197 21.2875 28.3682C21.3522 28.3155 21.5152 28.1755 21.6422 28.066H16.9896Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M25.7179 31.5534C24.2873 32.5778 22.9008 33.9483 22.0605 35.4405C21.2102 37.0203 19.9226 38.3933 18.3514 39.4295C18.1284 39.5765 17.9576 39.6901 17.8263 39.7788C18.6567 40.4199 19.5374 40.9984 20.4622 41.5055C20.6633 41.3541 20.9375 41.1494 21.3136 40.8806C23.2881 39.4652 24.3417 37.3372 25.7179 35.4074V31.5534Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M38.7771 35.9322C37.2067 35.08 35.5836 33.4338 34.5074 31.9782V36.2586C34.6285 36.4007 34.8244 36.6382 35.1456 37.0203C36.4558 38.5816 38.4936 39.1448 39.8072 40.6475C40.0219 40.8924 40.1853 41.0759 40.3099 41.2126C41.1631 40.7125 41.9759 40.1505 42.7424 39.5339C41.7394 37.9933 40.3392 36.7822 38.7771 35.9322Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M39.8751 5.80878C39.4842 6.04668 38.8088 6.43683 37.7215 7.46253C36.5387 8.57705 35.6888 10.6009 34.5074 11.8647V15.474C36.0085 14.4203 37.5356 12.9555 38.4681 11.426C39.3697 10.0012 40.5032 8.65398 41.9452 7.85125C42.178 7.72165 42.3515 7.61648 42.4814 7.52938C41.6603 6.89144 40.7894 6.31569 39.8751 5.80878Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M25.7179 10.8123C25.5297 10.5812 25.2296 10.2247 24.9969 10.0012C23.5674 8.6299 21.3839 7.61095 20.0108 6.01938C19.1492 6.52125 18.3281 7.08502 17.5536 7.7048C18.5837 9.10353 20.0773 10.6749 21.5966 11.4521C23.166 12.3644 24.5246 13.7221 25.7179 14.9903V10.8123Z"
+						fill="#FFFFFE"
+					/>
+					<path
+						fillRule="evenodd"
+						clipRule="evenodd"
+						d="M46.1483 11.2185C45.9915 11.3046 45.8081 11.4136 45.5898 11.5516C43.7871 12.6914 42.7046 14.6602 41.5462 16.3182C40.8732 17.333 39.8427 18.3362 38.882 19.1935H43.0539C44.0402 18.3095 44.9311 17.0365 45.4608 16.1886C46.0277 15.2823 46.9876 14.4421 47.7603 13.6651C47.2808 12.8093 46.7419 11.9918 46.1483 11.2185Z"
+						fill="#FFFFFE"
+					/>
+					<mask
+						id="mask0"
+						mask-type="alpha"
+						maskUnits="userSpaceOnUse"
+						x="7"
+						y="0"
+						width="46"
+						height="60"
+					>
+						<path d="M7 0H53V59.9593H7V0Z" fill="white" />
+					</mask>
+					<g mask="url(#mask0)">
+						<path
+							fillRule="evenodd"
+							clipRule="evenodd"
+							d="M29.835 48.2351H30.8162V46.9305H29.835V48.2351Z"
+							fill="#FFFFFE"
+						/>
+						<path
+							fillRule="evenodd"
+							clipRule="evenodd"
+							d="M49.0207 0H10.9788C8.78489 0 7 1.72764 7 3.85112V56.1139C7 58.2375 8.78489 59.965 10.9788 59.965H49.0207C51.2149 59.965 53 58.2376 53 56.1139V3.85132C53 1.72764 51.2149 0 49.0207 0ZM49.0207 0.947639C50.6917 0.947639 52.0467 2.24729 52.0467 3.85132V56.1139C52.0467 57.7178 50.6917 59.0175 49.0207 59.0175H10.9788C9.30782 59.0175 7.95328 57.7175 7.95328 56.1139V3.85112C7.95328 2.24729 9.30782 0.947639 10.9788 0.947639H49.0207Z"
+							fill="#FFFFFE"
+						/>
+					</g>
+				</>
+			);
+		case LOGO_KEYS.IN_VIA:
+			return (
+				<>
+					<defs>
+						<path id="a-2" d="M.045.178h14.353v15.409H.045z" />
+					</defs>
+					<g fill="none" fillRule="evenodd">
+						<path d="M-4 0h60v60H-4z" />
+						<g fill="#fff">
+							<path d="M6.872 57.967c.116-.36.105-.866.105-1.32.088-.044.011-.098.022-.163.055-.165-.022-.449.22-.47.24-.536.065-1.147.131-1.76-.164-.404.077-.82 0-1.224l.044-.109c-.021-.6.14-1.037.163-1.613v-.001l.001-.069c.143-.218.143-.546.23-.808 0-.267.096-.51.16-.75.034-.125.06-.249.06-.375.208-.558.154-1.236.307-1.815-.022-.458.175-.961.033-1.442-.241-.23-.625-.295-.9-.448-.23.142-.547-.033-.679.295-.187.077-.252.24-.428.339.055.164.022.371.033.546-.263.667-.132 1.388-.318 2.076A48.773 48.773 0 006 51.306v.001c.001.816.018 1.639.034 2.51.099.71.12 1.65.176 2.382.079.588.109 1.176.208 1.769.025.146.053.292.088.44.098.108.022.316.153.392v-.11c.022-.087.11-.119.187-.13 0 0 .024-.589.026-.592v-.001M15.278 49.071c-.146.498-.176 1.037-.289 1.548a1.088 1.088 0 00-.067.148h-.002c-.19.49-.091 1.088-.272 1.574-.06.045-.095.205-.225.171-.259-.388-.507-.844-.707-1.266-.07-.156-.138-.315-.206-.475-.247-.58-.488-1.175-.777-1.7l-.066-.119c.012-.433-.33-.661-.366-1.072-.212-.216-.224-.628-.448-.821a2.976 2.976 0 00-1.356-1.37c-.306.08-.46-.25-.742-.136-.118.194-.449.057-.531.33-.024.389-.177.73-.2 1.118-.107.206-.048.48-.142.685l-.059.856-.059.056c-.011.157-.024.313-.038.47-.049.566-.108 1.133-.156 1.701a22.663 22.663 0 00-.089 1.686c-.13.08.024.182-.023.297-.13.57 0 1.13 0 1.643.047.695.212 1.3.59 1.813.33.08.435.4.789.331.472-.148.13-.73.236-1.072-.177-.377-.271-.89-.142-1.289.012-.411.083-.913.048-1.335.2-.49.259-1.038.342-1.563.07-.16.108-.333.134-.51.035-.24.049-.487.09-.722.106-.103.035-.297.176-.353l.165.102c.17.326.33.653.517.974.035.06.071.119.108.178.047.183.342.274.307.491.719 1.14 1.497 2.042 2.381 3.046.177.4.802.707.767 1.198.082.296-.06.57.106.821.849.263 1.568 1.016 2.5 1.095.117-.182.471-.296.424-.604-.354-.48-.59-.81-.743-1.37-.177-.17-.047-.364-.141-.546.094-.526-.177-1.005-.083-1.53-.106-.034-.07-.182-.118-.273.189-.16 0-.4.059-.57-.33-.206.2-.469-.141-.64-.038-.43.028-.87.092-1.296v-.002c.017-.115.034-.23.049-.344.07-.195-.023-.446.083-.628.02-.238.053-.481.095-.726.076-.434.185-.869.329-1.27a.585.585 0 01.094-.628c-.13-.046-.188-.24-.342-.217.012-.08-.047-.183.06-.228.105-.662.483-1.233.613-1.883h-.083c-.106-.09-.024-.239-.024-.353-.118.069-.188-.034-.283-.092-.448.354-.483.8-.766 1.21-.471 1.015-1.096 1.973-1.332 3.092-.055.12-.1.243-.136.369M30.544 51.621c.074-.16.264-.317.032-.457.091-.407.263-.79.39-1.184-.024-.096-.001-.152.038-.193.099-.098.3-.1.179-.362.172-.333.355-.568.47-.925.343-.407.526-.888.905-1.357.263-.247.515-.592.664-.925.114-.21.355-.297.378-.555-.035-.099-.126-.049-.195-.062-.94.494-1.695 1.37-2.474 2.048-.15.258-.332.567-.573.74-.335.463-.678.923-1.015 1.397a21.633 21.633 0 00-1.024 1.563c-.075.09-.144.183-.21.276v.002c-.463.669-.754 1.456-1.325 1.942-.24-.197-.218-.604-.298-.924-.037-.34-.078-.68-.123-1.02a52.991 52.991 0 00-.278-1.838 80.55 80.55 0 00-.653-3.384c-.172-.297-.573-.654-1.008-.592-.034.086-.126.049-.195.061-.068.136-.32.21-.194.445.08.395.034.702.114 1.085.115.755.23 1.579.278 2.385.01.156.016.31.02.464.033.453.13.911.206 1.374.058.357.102.717.092 1.081.114.58.229 1.246.332 1.862.137.74.584 1.727.057 2.369-.022.136-.091.382.058.443.01.47.503.309.71.593.458.234.848.74 1.375.826.206-.037.378-.049.583 0 .07-.16.298-.185.298-.382.207-.222-.045-.53.115-.765.035-.715.527-1.16.71-1.813.539-.444.71-1.184 1.043-1.74.023-.086-.08-.086-.103-.148.32-.505.676-1.023.825-1.664-.115 0-.195-.087-.275-.148-.07-.136-.012-.284.046-.408a.206.206 0 01.025-.106v-.004" />
+							<g transform="translate(33.6 44.4)">
+								<mask id="b-2" fill="#fff">
+									<use xlinkHref="#a" />
+								</mask>
+								<path
+									d="M2.239 6.87a.86.86 0 01.037-.094c.062-.553-.04-1.095.1-1.635V5.14c.01-.04.021-.08.034-.12-.179-.267.011-.615-.056-.894.135-.57-.134-1.14-.078-1.697-.101-.592-.123-1.138-.19-1.708 0-.174.268-.337.056-.5h-.828C1.1.141.787.151.742.442.608.407.485.49.43.617c.112.348-.123.824 0 1.22l-.056.058C.384 2.592.205 3.266.24 3.963a7.93 7.93 0 00-.15 1.177l-.018.357a53.9 53.9 0 00-.026 1.376v.003c-.006 1.402.049 2.89.216 4.28.112.221.078.512.168.756-.045.255.09.51.134.778-.045.326.168.534.135.86.036.135.073.271.115.405.123.395.291.764.635.99.224.081.447 0 .503-.245-.012-.24-.025-.49-.04-.745v-.001c-.04-.658-.1-1.343-.228-1.938-.057-.522-.123-1.208-.023-1.731-.034-.535.168-1.07.135-1.615.056-.128.145-.303.111-.5.323-.378.191-.877.333-1.297V6.87M9.126 7.077c-.627.024-1.254.326-1.847.175.021-.127.063-.254.11-.383.102-.285.227-.58.204-.907.044-.232.48-.384.302-.698.011-.041.026-.082.042-.123.1-.245.279-.48.327-.76.134-.22.246-.44.403-.65.355.39.418.942.624 1.41.015.034.03.068.048.1.248.504.423 1.093.625 1.631.008.024.015.045.024.066-.235.22-.594.058-.862.14zm4.947 7.926c-.045-.325-.257-.535-.134-.895-.035-.05-.066-.1-.094-.153-.134-.253-.198-.535-.253-.823-.068-.489-.515-.895-.358-1.476l-.045-.081c.01-.5-.045-.825.045-1.29-.134-.558-.045-1.116-.18-1.674-.26-.559-.337-1.195-.624-1.739-.047-.09-.1-.176-.159-.26-.19-.395-.325-.86-.649-1.15a.977.977 0 01-.119-.321v-.002c-.033-.157-.048-.324-.094-.479-.403-.522-.671-1.173-1.13-1.673-.27-.371-.481-.836-.94-.976-.426-.244.022-.709-.023-.976A2.789 2.789 0 008.107.78c-.112.034-.212.162-.324.115-.806.849-1.187 2.115-1.769 3.15.1.243-.168.325-.168.523-.118.155-.174.396-.257.571h-.001l-.01.021c-.079.14-.18.279-.213.442-.078.36-.28.65-.426.976.065.111-.008.195-.048.294a.223.223 0 00-.007.182c-.359.419-.247 1.035-.537 1.476.022.279-.269.337-.157.64-.392.255.1.708-.325.929-.09.453-.268.755-.268 1.254-.19.477-.101 1.116-.057 1.639.128.363.315.862.687.963h.002c.02.006.04.01.061.013a.787.787 0 00.192-.013h.003a.496.496 0 00.376-.324c.023-.476.124-1.069.347-1.533.168-.222.023-.593.291-.745-.056-.266.101-.522.19-.731-.1-.337.28-.407.292-.698.022-.394.234-.743.515-.999.582.07 1.31-.116 1.902-.174.638 0 1.276-.302 1.992-.279.359.162.303.638.46.953.201.29.033.65.235.918.045.848.57 1.499.805 2.266.012.418.504.592.404 1.022.106.104.204.212.295.322v.002c.358.43.617.905.947 1.35.392-.047.48.383.862.278-.101-.197-.168-.406-.325-.58z"
+									mask="url(#b)"
+								/>
+							</g>
+							<path
+								d="M25.905 3.774c.553-.093.979.151 1.52.104.219.14.553-.023.748.14.254.047.623.012.887.14.242.046.461.162.737.116.104.104.323.023.45.14.207-.094.264.173.472.08.138.035.242.163.414.116.012.105.162.012.22.082.379.093.713.302 1.105.36-.012.046.046.07.08.058h.08c.093.046.208.14.335.14.127-.012.219.139.334.08.184.082.356.268.552.221.208.128.404.221.657.28l.864.452c.23-.012.345.232.575.22-.011.024.012.035.035.059.15.058.31.174.495.22.703.36 1.405.837 2.13 1.232 1.417.894 2.776 2.067 3.813 3.39-.012.175.23.233.218.396.288.418.542.72.691 1.196.116.22.173.418.3.616.034.127.115.243.172.36-.023.174.138.394.138.557.035.082.07.21.139.314-.058.22.126.383.104.557.08.256.206.534.115.836.115.175-.012.453.115.64l-.057 1.312c-.082.197-.082.441-.116.662.034.267-.173.36-.104.639-.103.046-.046.15-.057.232-.116.07-.023.279-.116.383-.057.093-.08.244-.103.36-.08.244-.184.523-.276.79-.047-.011-.07.035-.058.081-.092.28-.265.64-.357.976-.035.174-.173.314-.196.5-.07.057-.092.185-.138.278-.23.406-.484.778-.691 1.232-.161.011-.058.22-.196.22-.253.453-.725.778-1.025 1.08-1.174.709-2.383.825-3.754.894-.599.21-1.197-.08-1.739.256-.207-.07-.333.116-.506.058a13.25 13.25 0 00-3.145.918c-.725.279-1.335.545-1.969.94-.23.116-.449.28-.69.372-.415.314-.876.709-1.325.941-.161-.29-.046-.535-.057-.894l-.046-.047.046-.058.092-2.346c.08-.209.046-.464.138-.662-.092-.349.15-.546.104-.871l1.105-1.08c.657-.953 1.347-1.765 1.969-2.706.092-.268.346-.465.415-.779.253-.406.576-.871.795-1.254.333-.558.564-1.162.863-1.707-.046-.128.07-.21.104-.279.115-.314.218-.627.334-.941.046 0 .07-.046.057-.093V16.2c.104-.022.012-.174.104-.22.173-.43.334-.999.45-1.417-.093-.325.15-.51.103-.813l.057-.08.058-1.755c-.115-.244-.011-.557-.08-.813.011-.534-.208-1.069-.357-1.533-.173-.406-.438-.732-.61-1.115-.163-.14-.243-.325-.335-.5-.633-.638-1.267-1.23-2.073-1.707-.138 0-.253-.162-.414-.162-.127-.14-.38-.174-.53-.28-.138-.01-.3-.127-.495-.116-.104-.046-.242-.115-.38-.139l-.058-.046c-.08-.035-.276-.058-.415-.093-.195-.104-.437-.128-.667-.162-.07-.106-.22-.036-.334-.059-.093-.093-.323-.023-.438-.139-.288.011-.518-.14-.829-.082-.161-.127-.438.035-.576-.14-.564-.092-1.186-.022-1.716-.162-.219-.022-.472.059-.633-.092-.898.023-1.716-.128-2.58-.186-.184-.047-.426-.14-.518-.232-.115-.21.08-.28.219-.36.771-.128 1.624-.094 2.43-.14l.057-.058c.922.058 1.889-.012 2.764.174"
+								fillOpacity=".804"
+							/>
+							<path
+								d="M20.355 9.668c.189.058.377.115.566.16.213-.011.354.138.566.115.035.115.2.011.283.08.436.138.92.275 1.332.412.732.206 1.475.412 2.288.447.142.114.377-.012.507.103.401-.012.767.023 1.132.057.047.115.224-.012.236.137.047.86.224 1.752.283 2.61 0 .115.118.184 0 .298-.154.024-.342-.045-.46.058-.637-.058-1.238.126-1.804.057h-4.245v-.023c-.212-.034-.377-.068-.566.023-1.108-.034-2.063.057-3.172.023l-.082.057c-1.262-.114-2.382.103-3.656.034l-.059.046c-1.344-.114-2.735.126-4.068 0-.27-.091-.613-.034-.907-.046-.036-.08-.166-.171-.118-.309.2-.023.259-.24.424-.274.165-.23.425-.298.649-.47 1.202-.778 2.546-1.397 3.773-1.946.141-.057.365-.08.507-.218.342-.114.672-.217 1.014-.332.483-.263 1.108-.366 1.615-.55.236.035.436-.182.708-.136.165-.046.318-.138.507-.138.33-.126.731-.08 1.05-.252.223.069.377-.114.565-.046v-.034c.413-.057.72.126 1.132.057"
+								opacity=".8"
+							/>
+							<path
+								d="M.943 21.878c.069.22.149.438.16.646.104.184.184.45.207.669.138.16.092.415.253.565.126.461.39.865.632 1.257.3.496.529.992.931 1.418l.023.08 1.38 1.443c.16.069.286.3.459.369.437.45.965.818 1.494 1.142.184.172.437.253.644.437.149-.034.218.15.344.185l.023-.012c.173.139.368.231.552.346h.046c.345.174.666.392 1.011.531.127 0 .253.115.368.173.15.173.46.092.598.288.137.035.253.115.38.184.137 0 .217.14.367.116.16.093.356.104.494.242.253.047.506.104.724.243.126 0 .218.07.333.127.253.011.448.184.69.242.184-.035.275.173.46.126.333.173.666.185 1.01.335.185-.023.311.08.46.127.092.011.207-.046.265.057.253-.034.368.116.62.093.046.08.161-.023.184.092.104.046.265-.034.357.069.08.023.207-.046.253.058.333.023.631.138.953.196.138 0 .276 0 .368.08.242-.057.414.104.632.07.07.127.253-.011.322.115.16.023.356-.069.46.07.367-.035.666.057.977.126.483.035 1 .012 1.494.024.092.057.195.207.126.345-.563.98-1.862.969-2.827 1.372l-3.643 1.085c-1.782.012-3.483-.012-5.24-.058-1.197-.496-2.323-1.188-3.483-1.788-.977-.83-1.977-1.672-2.977-2.398-.23-.22-.506-.438-.747-.658-.15-.057-.253-.276-.414-.322-.333-.462-.827-.739-1.207-1.12-.24-.507-.413-1.037-.678-1.533-.15-.588-.494-1.096-.597-1.695l-.046-.058L0 22.65c.184-.46.15-1.06.517-1.395.07-.035.15-.011.207.012.16.173.127.415.218.611"
+								fillOpacity=".8"
+							/>
+							<path d="M36.551 8.16v.058c-.16 0-.264-.115-.39-.173-.483-.162-1.058-.335-1.46-.496.644.127 1.23.403 1.85.61zm-10.481.634v.392c-.299-.011-.678.024-.943-.138-.368-.058-.77.15-1.08 0l.448-.093c-.184 0-.207-.207-.413-.138.643-.115 1.321-.115 1.93-.104l.058.081zm.023.554c.046.322.207.75.115 1.084-.942-.092-1.908-.277-2.874-.392-.344-.115-.712-.173-1.045-.3v-.058a18.83 18.83 0 013.804-.334zm-9.677 1.96c-.31.116-.632.381-1 .312v-.058c.357-.011.632-.322 1-.254zm29.617 8.42h-.162c-.022-.197.012-.462.138-.612v-.75c.088.044.09.13.086.226-.004.077-.014.16.018.224.046.312.103.646-.08.911zm-19.47-2.157c-1.183.057-2.286-.012-3.504.057-.702.092-1.414-.034-2.127.139-1.609.115-3.24.334-4.827.599-.581.015-1.11.12-1.648.226-.323.06-.65.123-.995.166-1.574.208-3.149.577-4.723.854-.138.138-.357.08-.518.173-.459-.343-.996-.718-1.337-1.194-.256-.358-.401-.773-.317-1.274.05-.148.1-.297.154-.443.174-.458.398-.878.834-1.08 1.597-.057 3.264-.126 4.907-.219.334.012.701.104 1 .024 1.804-.047 3.85.126 5.677-.047 1.253.035 2.472-.092 3.701 0 .357.012.713-.046 1.07.047.183-.14.424.046.585-.047.678.093 1.391.035 2.011.104.067.384.039.8.031 1.216 0 0-.002.47.027.699zm1.851-1.88c.575-.092 1.173-.058 1.793-.139 1.804-.08 3.712 0 5.563.138.747-.01 1.528.116 2.263.105.885.115 1.701.138 2.564.276.93-.07 1.804.093 2.735.196.264.162.551.035.85.116.322.256 1.074-.113 1.082.49l-.001.063c.056.538.117 1.103.092 1.657-.016.346-.067.69-.173 1.019-.276.161-.552-.035-.804-.104-.104-.196.034-.6-.242-.669.1-.124.044-.188-.032-.25H44.1c-.064-.054-.144-.105-.16-.188-.005.067-.005.13-.002.19h.042l-.042.002c.019.286.125.524.107.858-1.518-.254-2.909-.577-4.437-.808-.07-.017-.141-.035-.212-.051-2.315-.544-4.918-.466-7.293-.78-1.218 0-2.516-.092-3.723-.137a20.008 20.008 0 01-.03-.803s.01-.792.064-1.181zm16.182 5.858l-.046-.046c.127-.184.334-.242.414-.473.057-.023.092.035.138.058l-.506.461zm-1.402.67v-.082c.299-.472.713-.807.943-1.326l.24.046c-.275.635-.689.97-1.183 1.361zm-.551 0c-1.046.703-2.057 1.326-3.31 1.499-.265.045-.495.138-.747.219-.805.277-1.69.138-2.54.334-1.598-.115-3.092.104-4.69-.093-.46-.08-.999.07-1.424-.08-.644-.115-1.242-.115-1.932-.104-.115-1.58.277-3.125.3-4.682l.057-.058c1.47-.058 2.954-.15 4.413.081 1.172-.057 2.275.196 3.448.23 1.31.047 2.62.255 3.919.416 1.207.196 2.38.473 3.55.738-.137.623-.597 1.072-1.045 1.5zm-16.331-.947c.103.404-.08.82-.058 1.246-.057.485-.08.98-.184 1.442-.701-.058-1.368.127-2.046-.058-.092.023-.23.081-.299-.023-.586-.058-1.195-.116-1.747-.161l-.045-.058c-1.23-.023-2.506-.127-3.644-.45-1.643-.162-3.114-.484-4.631-.934-.207-.15-.471-.173-.724-.196-1.035-.3-1.897-.68-2.862-1.165v-.058c1.206-.173 2.402-.472 3.586-.634 2.045-.357 4.01-.519 6.09-.715 1.138-.161 2.449-.15 3.61-.22 1.023-.126 2.115-.068 3.172-.092.046.715-.173 1.373-.219 2.076zm2.01-9.04c1.828.183 3.713.137 5.552.38 2.436.104 4.47-.047 6.815-.197l.908-.138c.793.577 1.597 1.35 2.23 2.053.138.161.333.38.39.577-1.276-.139-2.574-.37-3.919-.439-.23.162-.425-.08-.69 0-3.125-.242-6.297-.196-9.32-.357-.288-.15-.61.046-.908-.058l-1-.08c-.15-.173.16-.404-.023-.612.115-.357-.103-.76-.035-1.13zm17.164-.52c1.702-.534 3.428-.996 5.04-1.73.54-.139.954-.473 1.414-.75-.046-.057-.046-.08-.035-.161-.183.08-.379.15-.551.277l-.08-.116c-.196-.034-.346.058-.495.139-.15.058-.356.253-.471.058-.16.08-.356.207-.517.195.103-.184.402-.195.551-.357-.126-.057-.298 0-.413-.057.057-.185.356-.082.356-.3-.069-.058-.15-.023-.218 0-.092-.07-.242-.046-.196-.197-.344.139-.666.3-1.022.415-1.184.485-2.448.808-3.701 1.131-.678.115-1.333.311-2.011.473-.587-.335-1.115-.634-1.678-1.027-.449-.403-1.115-.403-1.575-.853-.598-.173-1.115-.52-1.713-.692-.183.07-.401-.115-.608-.173-.219-.012-.54-.323-.656-.046l-1.08-.415c-.47.046-.943-.289-1.436-.37.034.058-.035.15-.104.174l-1.793-.473v.08c-.24.105-.517-.057-.77-.08-.413.3-.988-.265-1.333.162l-.023-.024c-.172 0-.494-.138-.552.139-.068-.012-.16.023-.218-.024-.483.081-.92-.057-1.413.081v.081c.988-.081 2.022.357 3.01.334 1.863.381 3.644.935 5.322 1.742a42.616 42.616 0 013.23 1.684l-.058.058c-1.655.115-3.367.345-5.069.334-1.735.034-3.47-.139-5.137-.277-.448-.138-1.069 0-1.379-.358l-.161-1.695c-.058-.127-.15-.346-.034-.438l1.138.057c.103-.045.114-.103.103-.218.15-.081.38.023.47-.058-.39-.219-.895-.15-1.32-.277-.219-.057-.472 0-.633-.173-.31-1.718-.804-3.264-1.138-4.982-.023-.254-.172-.53-.08-.773.184.265.39.312.69.358.264.058.597.058.746-.185-.023-.288.185-.472.115-.784.334-.057-.068-.288.024-.438.195-.254-.219-.289-.254-.473-.206-.208-.55-.254-.827-.38-.333.057-.574-.277-.965-.197-.264.023-.39.312-.598.358.127.265-.23.058-.195.277-.276-.023-.08.253-.334.276-.011.116.058.15.08.254.012.058-.149.092-.057.162.472 1.995.391 4.186.943 6.17-.012.092.172.173.023.253-1.023-.035-2.092.138-3.115.277-.241.058-.54-.023-.747.139-.839.103-1.632.334-2.482.437-.575-.346-1.241-.495-1.816-.853v-.035c-.31-.103-.563-.3-.908-.38-.724-.519-1.563-.865-2.206-1.465-.07-.23-.207-.45-.277-.645-.333-.196-.321-.692-.746-.854-.242.023-.414-.161-.69-.057-.207-.047.046-.15-.023-.277-.069-.046-.092-.208-.195-.22-.093.289-.46.324-.667.416-.23-.023-.471.23-.609.415-.092.208-.264.23-.276.473-.126.23.253.288.092.496.138.126.16.404.299.553-.115.323.23.45.39.577 1.713.68 3.403 1.73 5.012 2.526v.057c-1.173.392-2.391.75-3.529 1.188-.832.318-1.552.623-2.35 1.013l-.097.048a24.735 24.735 0 00-4.747 3.01c-1.207.173-2.414.092-3.62.277-.207.196-.538.103-.807.152a.73.73 0 00-.044.01l.08.138c-.057.046-.149.023-.218.023v.092c.494.023.943-.104 1.437-.058.011.023.069.046.057.081-.23.104-.517.035-.781.058-.092.104.057.173.115.242.643.035 1.195-.058 1.77-.023-.115.08-.357.069-.472.115-.184-.011-.402-.011-.494.081.345.116.644.011.989-.035.356.046.632.035.965.116-.035.184-.15.334-.218.496-.066.122-.12.251-.163.385a2.717 2.717 0 00-.01 1.552c.013.056.028.112.046.169.25.76.95 1.291 1.39 1.884-.804.357-1.666.611-2.482.968-.045.104.104.197-.023.278-.138.01-.115.184-.115.241-.069.035-.172.047-.195.14-.138.264-.494.322-.77.449v.08c.242-.057.61-.242.804 0 .53-.23 1.207-.357 1.679-.691.149-.173.459-.104.609-.277.241-.012.436-.208.666-.162-.103.07-.241.185-.39.185-.265.219-.61.265-.885.392-.138.115-.437.07-.334.358h.138c.563-.231 1.126-.427 1.69-.635.505-.138 1.195-.588 1.655-.023 1.218.646 2.517 1.072 3.861 1.43 1.552.6 3.276.773 4.965 1.026.126.07.322-.022.437.081.862-.011 1.609.265 2.459.277 1.908.323 3.93.507 5.93.692.012.507-.114 1.026-.195 1.523a34.583 34.583 0 00-.494 2.906c-.104.184.08.392-.058.553.024.219-.206.404-.114.588-.058.024.023.15-.081.139.012.173.012.346-.103.472.057.15-.115.231-.115.323-.012.024-.035.047-.058.035v.796c.127.103-.011.311.115.415-.08.38.265.508.242.865.172.346.655.438.862.692l.023-.035c.172.058.413.15.551.035.253 0 .402-.104.61-.22.355-.172.678-.426.988-.691-.047-.288.356-.265.195-.553.322-.185-.057-.485.115-.669.069-.127 0-.208-.08-.278.01-.218.103-.449-.115-.576-.22 0-.185-.242-.277-.37-.08-.322-.378-.068-.528-.299.035-.22-.035-.484-.184-.634-.146-.283-.017-.488-.163-.757a.144.144 0 00-.01-.016c.058-.346.093-.761.253-1.003.207-.508.127-1.13.357-1.627.057-.472.115-.945.16-1.418.081-.184.207.012.334-.057 2.31.08 4.77.057 7.034.23.551-.07 1.069-.127 1.597-.058 2.08-.265 4.195-.426 6.125-1.026 1.736-.657 3.31-1.407 4.586-2.964.701.3 1.483.508 2.173.888.482.265 1.068.462 1.413.888.15.323.506.461.85.554.138-.266.471-.035.667-.312-.034-.023-.08-.057-.058-.103a.29.29 0 01.277-.08c.379-.186-.22-.44.218-.59l-.058-.057c.012-.3-.356-.323-.47-.553-.472-.139-.851-.508-1.345-.6-1.012-.554-2.184-.807-3.253-1.28.066-.6.148-1.175.13-1.772a5.378 5.378 0 00-.026-.385c.01-.45-.23-.842-.196-1.269.908.104 1.759.335 2.655.438.575.139 1.16.289 1.793.3-.07-.138-.288-.276-.471-.323.092-.369.46 0 .689-.115.046-.092-.08-.115-.058-.196.185-.115.369.104.61.035.046-.104.103-.023.16-.035.139-.104.368-.023.587-.08v-.081l.003-.007c.014-.028.046-.028.077-.028.092-.092-.045-.115-.115-.138-.172-.15-.47-.139-.712-.22-1.07-.23-2.172-.541-3.31-.691-.804-.138-1.632-.277-2.482-.358l-.151-.285c-.657-1.185-1.542-2.236-2.309-3.198.25-.083.5-.163.752-.242zm-21.38 0c.735.152 1.481.28 2.24.38.091.15.022.37.08.554.15.427.057.9.08 1.35l-.16.057c-1.23-.092-2.345-.046-3.54-.138-1.782.265-3.414-.046-5.207.115-.333.058-.758-.011-1.16.046-1.288.138-2.552.023-3.862.173-1.115-.035-2.15.058-3.23.104-.022-.15.161-.023.196-.162.16.046.333-.16.494-.253 1.184-.588 2.39-.946 3.586-1.488.483-.242 1.023-.404 1.494-.611v-.115c-.506.069-1.011.288-1.494.507-1.069.254-2.057.773-3.034 1.188.597-.346 1.333-.773 2.068-1.05.573-.136 1.033-.518 1.593-.657l.021-.005c.058-.014.096-.021.156-.03.966-.438 1.988-.726 3.034-.968.46-.139.897-.346 1.38-.415.448-.047.758.276 1.183.322 1.332.43 2.688.806 4.082 1.095z" />
+						</g>
+					</g>
+				</>
+			);
+		case LOGO_KEYS.KREUZBUND:
+			return (
+				<g fill="none" fillRule="evenodd">
+					<path d="M1 1h60v60H1z" />
+					<path
+						d="M1.985 32.605L1 16.144s9.6-2.315 10.832-2.304c1.232.01 7.221-.659 7.221-.659l.985-11.852L40.556 1l1.474 11.524 18.709 3.621-.985 16.132L42.03 30.63l-2.298 15.803s-6.771-1.52-9.847-1.646c-3.075-.125-11.817.988-11.817.988V32.277l-16.083.329z"
+						stroke="#FFFFFE"
+						strokeWidth=".558"
+					/>
+					<path
+						d="M4.367 56.946l1.647 1.783c.331.367.48.541.48.673 0 .148-.193.218-.593.218v.227h2.962v-.227c-.505-.026-.906-.35-1.603-1.119l-1.9-2.106 1.142-1.205c.549-.568 1.028-1.014 1.855-1.058v-.227H5.831v.227c.41 0 .592.07.592.271 0 .157-.104.323-.374.62l-1.682 1.8V54.85c0-.62.166-.69.837-.717v-.227H2.38v.227c.636.026.801.096.801.717v4.054c0 .62-.165.69-.801.717v.227h2.823v-.227c-.671-.026-.837-.096-.837-.717v-1.957M11.107 57.226h.383l2.03 2.622h1.664v-.228c-.33.027-.487-.104-.696-.358l-1.69-2.158c.853-.219 1.463-.69 1.463-1.643 0-.979-.697-1.556-2.605-1.556H9.103v.227c.654.027.82.097.82.717v4.055c0 .62-.166.69-.82.716v.228h2.84v-.228c-.67-.026-.836-.096-.836-.716v-1.678zm0-2.657c0-.227.096-.314.505-.314.715 0 1.43.384 1.43 1.267 0 .97-.767 1.337-1.935 1.354V54.57zM20.531 59.847l.4-1.59h-.226c-.505 1.048-1.132 1.24-2.195 1.24-1.046 0-1.15-.122-1.15-.48v-2.054h1.01c.767 0 1.115.175 1.238.874h.226V55.74h-.226c-.114.62-.314.786-1.238.786h-1.01v-1.878c0-.245.06-.306.33-.306h.898c1.159 0 1.385.21 1.595.874h.243l-.034-1.311h-4.95v.227c.567.026.733.096.733.717v4.054c0 .62-.166.69-.732.717v.227h5.088M23.275 54.849c0-.62.165-.69.767-.717v-.227h-2.719v.227c.602.027.767.097.767.717v2.971c0 1.503.802 2.167 2.405 2.167 1.69 0 2.282-.909 2.282-2.324v-2.648c0-.577.157-.813.837-.883v-.227h-2.056v.227c.627.07.784.306.784.883v2.656c0 1.093-.506 1.792-1.525 1.792-1.09 0-1.542-.524-1.542-1.538v-3.076M32.437 59.847l.375-1.608h-.262c-.418.918-.827 1.171-1.838 1.171H28.96l3.78-5.226v-.28h-4.73l-.21 1.521h.262c.21-.682.54-1.083 1.577-1.083h1.63l-3.782 5.225v.28h4.949M36.174 59.847c1.568 0 2.317-.682 2.317-1.643 0-.909-.862-1.302-1.629-1.45v-.018c.723-.122 1.316-.612 1.316-1.363 0-.961-.906-1.468-2.335-1.468h-2.614v.227c.636.026.802.096.802.717v4.054c0 .62-.166.69-.802.717v.227h2.945zm-.958-2.849c1.35-.026 2.064.376 2.056 1.206-.009.656-.514 1.294-1.438 1.294-.496 0-.618-.14-.618-.376v-2.124zm0-2.342c0-.288.165-.402.47-.402.871 0 1.307.49 1.307 1.136 0 .857-.505 1.259-1.777 1.259v-1.993zM40.887 54.849c0-.62.166-.69.767-.717v-.227h-2.719v.227c.602.027.767.097.767.717v2.971c0 1.503.802 2.167 2.405 2.167 1.69 0 2.282-.909 2.282-2.324v-2.648c0-.577.157-.813.837-.883v-.227H43.17v.227c.627.07.784.306.784.883v2.656c0 1.093-.506 1.792-1.525 1.792-1.089 0-1.542-.524-1.542-1.538v-3.076M50.781 57.995l-3.529-4.09h-1.611v.227c.366.088.6.254.801.603v4.003c0 .577-.157.813-.801.882v.228h2.09v-.228c-.644-.07-.8-.305-.8-.882v-3.434l4.042 4.683h.296v-4.972c0-.577.157-.813.801-.882v-.228h-2.09v.227c.644.07.801.306.801.883v2.98M55.377 59.847c1.865 0 3.38-.996 3.38-2.971s-1.515-2.971-3.38-2.971h-2.962v.227c.583.026.749.096.749.717v4.054c0 .62-.166.69-.75.717v.227h2.963zm-1.028-5.05c0-.403.165-.543.697-.543 1.62 0 2.492 1.101 2.492 2.622 0 1.52-.871 2.622-2.492 2.622-.532 0-.697-.14-.697-.542v-4.16zM21.103 44.121c-.048.14-.076.295-.076.478 0 .146.125.25.166.375h-.499v-.853h.409zm4.159 0c.175.06.333.106.547.106.193 0 .341-.042.482-.106h-1.03zm8.3 0c-.153.033-.26.122-.438.17-.005-.047-.037-.065-.037-.11a.09.09 0 01.02-.06h.455zm3.49 0c-.1.045-.207-.004-.304.06-.146.09-.1.27-.146.44-.44-.147-.636-.44-1.099-.44-.251 0-.585.022-.585.274 0 .556.69.68 1.244.68h.476c.357 0 .559.11.916.11-.046.138-.11.22-.11.367 0 .046.036.064.036.11.165-.004.257-.036.421-.036.147 0 .22.073.366.073.055 0 .087-.041.128-.073.12.087.165.179.257.293h-.66c-.09-.041-.15-.11-.255-.11-.078 0-.106.074-.183.074-.586 0-.88-.33-1.465-.33h-.403c-.292 0-.439-.184-.732-.184h-.33c-.315 0-.489-.147-.805-.147h-.402c-.124 0-.17-.11-.293-.11h-.874a.388.388 0 00-.005-.404c.435 0 .668-.161 1.099-.22.224-.028.549.008.549-.221a.224.224 0 00-.132-.207l3.29.001zm-10.537.133c.325 0 .463.293.787.293a.596.596 0 00.476-.22c.083.096.183.16.183.294 0 .078-.06.11-.073.183-.06-.013-.082-.073-.146-.073h-.732c-.083.1-.16.183-.293.183-.106 0-.151-.073-.256-.073-.106 0-.152.073-.257.073h-1.537c-.485 0-.76.106-1.245.147-.014-.068-.028-.114-.037-.183.5-.046.71-.44 1.208-.44h.403c.124 0 .174.1.293.11.366.027.581.036.952.036.133 0 .183-.16.183-.294.037 0 .055-.036.091-.036zm3.863.33v.367h-.623a.188.188 0 00-.073-.073v-.11c.28 0 .458-.042.696-.184zm3.917 0c.16 0 .33.096.33.257 0 .152-.197.184-.348.184-.129 0-.202-.055-.312-.11.064-.166.151-.33.33-.33zm-12.996.22c.055.143.146.211.146.367 0 .092-.055.134-.073.22-.027 0-.041.038-.073.038-.115 0-.147-.138-.256-.147-.426-.046-.678.096-1.099.009v-.124c.12-.028.188.005.312.005.178 0 .251-.138.42-.184.243-.064.408-.055.623-.184z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.446 43.192a.996.996 0 01-.079.055v-.055h.079zm1.176 0c-.198.362-.41.617-.519.948h-.408v-.278c.332-.043.54-.064.872-.064l.077-.607h-.022zm5.455 0c.04.16.072.33.108.494-.208.085-.73-.176-.959-.176-.81 0-1.154.287-1.965.287l.222-.606h2.593zm9.899 0c.122.018.246.038.358.071.247.075.343.3.604.3.114 0 .183-.047.275-.113a.19.19 0 00.073.075v.225c-.256.014-.412.052-.66 0-.53-.112-.754-.524-1.299-.524-.137 0-.21.075-.348.075-.091 0-.128-.075-.22-.075h-.695c-.096 0-.142.075-.238.075-.123 0-.187-.075-.311-.075-.146 0-.224.042-.366.075a.812.812 0 00-.044-.11h2.87zm-10.685.947c.162-.073.315-.176.515-.278-.333-.255-.54-.425-.956-.425-.395 0-.478.425-.79.637l.202.067h1.029zm6.816 0c.043-.054.146-.065.2-.127.082-.093.073-.262.201-.262.316 0 .453.262.769.262.197 0 .302-.084.494-.112.54-.084.86-.075 1.41-.075.356 0 .54.187.896.187v.113a.215.215 0 01-.026.014h-3.29a.169.169 0 00-.07-.014.584.584 0 00-.129.014h-.455zm.639-.8a.257.257 0 01-.146 0 .743.743 0 01.146 0zm-8.768 0c.105 0 .151.074.256.074.225 0 .343-.075.568-.075.174 0 .274.019.44.075-.047.075-.092.15-.184.15h-1.281c-.041-.052-.078-.09-.11-.15.119-.028.187-.075.311-.075z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.895 42.262c-.073.405-.141.712-.453.949l-.076-.001v-.754a7.63 7.63 0 00.086-.195l.443.001zm1.403 0c.048.06.097.124.145.194h-.582c-.106-.05-.19-.126-.276-.194h.713zm7.516 0c.096.122.16.3.33.3.165 0 .202-.225.367-.225.169 0 .233.188.402.188.064 0 .092-.043.147-.075v.3c-.11-.076-.179-.15-.312-.15-.37 0-.453.486-.823.486h-.476c-.22 0-.339.056-.55.113v-.188h.294c.123-.089.292-.121.292-.28 0-.249-.279-.295-.512-.356v-.113h.84zm4.109 0c.19.199.3.532.61.532.32 0 .462-.26.634-.531l.659-.001-.067.046c.05.01.096.004.146 0a.117.117 0 01-.03-.047l.909.001c-.013.269-.513.218-.513.495 0 .35.346.396.704.452h-2.87c-.157-.322-.496-.452-.872-.452h.183c.114 0 .22-.089.22-.206 0-.22-.277-.185-.467-.289h.754zm-12.278.95l.007-.054-.029.054h.022zm2.838-.002l.084-.229c.166-.297.415-.51.768-.51h1.08c.402 0 .559.337.661.738h-2.593zm.86-.753c.115 0 .18.057.275.113.01.028.019.047.037.075-.041.01-.064.037-.11.037-.165 0-.256-.084-.403-.15.07-.042.12-.075.202-.075zm9.794.225c.082 0 .123.061.183.112-.05.07-.105.108-.146.188h-.184c0-.061-.036-.09-.036-.15 0-.094.091-.15.183-.15z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.752 40.404c.02.083.038.169.038.278 0 .166.27.166.333.291.166-.125.374-.25.374-.479a.705.705 0 00-.006-.09h-.739zm4.55 0a.822.822 0 00.382.097.822.822 0 00.381-.097h.189c.165.37.4.555.635.889h-.582c-.229-.104-.353-.334-.623-.334-.125 0-.208.084-.333.084-.083.458-.145.792-.54 1.042v-.792c.156-.332.238-.554.316-.888l.175-.001zm4.982 0c.082.164.134.355.167.57l-.79.082a7.592 7.592 0 01-.039-.652h.662zm6.663 0c-.287.1-.524.144-.785.36-.183.152-.256.313-.476.405-.352.151-.879.073-.879.459 0 .142.096.275.238.275.311 0 .476-.147.787-.147.385 0 .531.367.916.367 0 .051.036.079.036.129v.01h-.909a.184.184 0 01-.006-.028l-.043.028h-.659l.043-.065c-.256-.239-.416-.477-.769-.477h-.366c-.238 0-.366-.11-.604-.11-.165 0-.27.068-.42.004v-.262c.182-.06.246-.183.402-.293.27-.189.659-.216.659-.551a.341.341 0 00-.016-.104h2.85zm1.044 0c-.01.023-.022.046-.034.07a.379.379 0 00.003-.069h.031zm-17.539 1.858c.112-.257.179-.465.246-.768.208.083.312.167.54.167.229 0 .333-.084.54-.167.134.336.322.523.52.768h-.713c-.1-.077-.202-.143-.347-.143-.124 0-.207.083-.332.083l-.01.06h-.444zm8.521.004v-.113c.252.019.389 0 .64 0 .09 0 .15.049.2.114h-.84zm4.196-.004a.359.359 0 01-.046-.028c.197-.019.298-.11.494-.11.13 0 .225.057.306.139h-.754zm1.986-1.68c.201 0 .348.11.421.293-.362.18-.673.028-1.025.22-.105.055-.133.184-.256.184-.078 0-.11-.055-.183-.073.073-.12.21-.11.293-.22.096-.13.054-.368.22-.368.068 0 .095.073.164.073.146 0 .22-.11.366-.11zm1.647.293c.357.018.307.51.604.698.252.16.733-.028.733.275 0 .11-.096.147-.147.239-.201-.262-.42-.326-.732-.44-.28-.102-.732.045-.732-.258 0-.202.087-.33.004-.514h.27zm-11 .477h.22c.008.042.036.065.036.11 0 .102-.11.138-.11.24 0 .233.32.21.55.238v.11a.858.858 0 01-.44.11c-.284 0-.453-.027-.732-.073-.129-.019-.202-.078-.33-.11a.275.275 0 00-.037-.074c.312-.225.476-.431.842-.55z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.698 39.475c-.012.193.01.398.01.636 0 .123.023.216.044.311h.74c-.026-.201-.13-.338-.203-.524l.207-.423h-.798zm5.126 0c.092.077.173.177.307.23.01.028.019.047.037.075-.243.098-.398.15-.659.15h-.366v-.225c.086-.066.142-.159.207-.23h.474zm10.863 0c.067.279.395.417.702.417.274 0 .42-.094.677-.187v.412c0 .122-.03.207-.077.306h-.03c-.008-.154-.108-.306-.26-.306-.228 0-.342.13-.548.224a2.71 2.71 0 01-.204.082h-2.85c-.053-.171-.236-.267-.424-.306-.233-.052-.398.089-.623.004v-.229c.284.052.458-.112.751-.112.27 0 .275.365.53.45.385.13.637.111 1.044.111.293 0 .714.038.714-.262 0-.22-.306-.178-.512-.224-.376-.09-.641-.07-.952-.3.082-.033.184-.047.26-.08h1.802zm-11.559.95l.016-.073c.06.025.11.05.159.072h-.175zm.938 0c.048-.024.099-.048.159-.073l.03.073h-.19zm3.557-.002c0-.17.01-.345.04-.567.301.115.496.309.622.566h-.662zm6.833-.943c-.042.09-.101.131-.183.187a.272.272 0 00-.074-.037v-.15c.101-.005.156-.005.257 0zm.164.487a.62.62 0 01.385.112v.08c-.275.07-.462.07-.732-.005.05-.14.196-.187.347-.187z"
+						fill="#fff"
+					/>
+					<path
+						d="M25.626 38.546c.112.174.249.343.392.533l.54-.085a5.357 5.357 0 01-.01-.447l-.781-.001a.25.25 0 01-.039.016.04.04 0 00-.003-.016h-.099zm4.867 0c.174.074.429.078.639.102.183.024.288-.004.476-.004-.05.15-.096.234-.147.379h-.805c-.247 0-.375-.112-.623-.112-.2 0-.329 0-.512.074a.19.19 0 00-.073-.074v-.365h1.046zm5.954 0a.178.178 0 01.008.054c0 .206-.257.262-.458.262-.325 0-.494-.164-.787-.3a.715.715 0 00-.037-.015l-.706-.001.084.016c-.293.455-.737.422-1.208.675.032.023.05.042.073.075.21 0 .32-.075.53-.075.147 0 .22.075.367.075.178 0 .242-.188.421-.188.128 0 .256.094.256.225 0 .076-.044.117-.106.145h1.803a.46.46 0 01-.013-.108c0-.309.366-.383.366-.693a.47.47 0 00-.022-.147h-.571zm-14.951.948l.084-.172c-.27-.149-.54-.638-.707-.383a1.08 1.08 0 00-.176.554h.799zm3.854.002c.056-.063.118-.11.213-.11a.38.38 0 01.26.109h-.473z"
+						fill="#fff"
+					/>
+					<path
+						d="M2.936 31.113c-.23.127-.438.189-.718.318v.588c0 .137.064.293.202.293.215 0 .315-.147.53-.147.344 0 .54.106.88.074.1-.01.15-.074.255-.074.069 0 .096.074.165.074.471 0 .956-.409 1.208-.01.055-.215-.018-.353-.018-.578a.77.77 0 01-.403.11c-.22 0-.512-.018-.512-.238 0-.115.11-.16.11-.275a.293.293 0 00-.035-.135H2.936zm42.584.002c-.03.009-.062.02-.095.033a.212.212 0 00-.024-.034h.12zm1.87-.001a1.757 1.757 0 00-.026.071h-.805c-.144 0-.247-.038-.358-.07l1.19-.001zm11.658-.001c.156.07.314.126.507.216v.587c0 .138-.064.294-.201.294-.215 0-.316-.147-.53-.147-.344 0-.541.106-.88.074-.1-.01-.15-.074-.256-.074-.068 0-.096.074-.165.074-.47 0-.956-.409-1.208-.01-.055-.215.019-.353.019-.578a.77.77 0 00.402.11c.22 0 .513-.018.513-.238 0-.115-.11-.161-.11-.276 0-.01 0-.021.002-.031l1.907-.001zm-33.324 7.433c-.009-.02-.033-.031-.033-.057 0-.05.046-.074.073-.11.014.055.028.091.037.147a.332.332 0 01-.035.02h.781c.002-.322.006-.598-.155-.936-.083-.166-.374-.166-.374-.375 0-.292.208-.458.208-.75 0-.27-.416-.438-.54-.667 0 .521.02.813.02 1.334 0 .25-.353.292-.353.542 0 .346.11.604.273.852h.098zm3.723 0v-.157h.476c.055-.068.11-.11.11-.202 0-.362-.384-.66-.732-.569 0-.105-.005-.179-.073-.257.279-.036.43-.146.714-.146.356 0 .32.624.677.624h.879c.064 0 .086-.06.146-.074.01 0-.11.024-.146 0l.183.478c-.197-.156-.353-.22-.605-.22-.32 0-.787-.083-.787.238 0 .15.083.234.204.285h-1.046zm5.02 0a4.697 4.697 0 00-.978-.094c-.092 0-.133.046-.22.074a.518.518 0 00.147-.368c.366 0 .563-.133.915-.22.188-.046.44-.06.44-.257 0-.29-.445-.12-.696-.257-.17-.092-.243-.257-.44-.257v-.698h.11c.288.097.421.23.696.367.22.11.412.083.586.258.169.17.123.514.366.514h.33c.26 0 .274.344.512.44.317.131.69.218.783.499h-.572c-.038-.125-.212-.168-.358-.168-.16 0-.242.073-.402.073-.312 0-.403-.367-.714-.367-.183 0-.302.087-.421.22v.147c.238.061.407.01.622.095h-.706zm24.136-7.217h.11l.073.073c-.078.115-.151.22-.293.22-.105 0-.165-.064-.256-.11.082-.133.224-.114.366-.183zM25.436 33.64c0 .229.083.354.083.583 0 .167.27.167.333.292.166-.125.374-.25.374-.48 0-.25-.125-.395-.208-.604l.29-.583c-.27-.146-.54-.625-.706-.375-.249.375-.166.708-.166 1.167zm8.837-.706c.229.041.353.062.582.083.041.209.083.313.083.521 0 .271-.166.417-.208.688l-.457.041v-1.333zm.387-1.716h.403v.11a.964.964 0 00-.256.514c-.032 0-.042.037-.073.037-.27 0-.32-.335-.586-.368v-.183c.196-.037.343-.01.512-.11zm.405 3.746v.076-.076zm-23.947 1.834v.074-.074zm9.58-.243c.208.083.311.167.54.167.229 0 .332-.084.54-.167.166.417.416.604.665.959h-.582c-.228-.105-.353-.334-.623-.334-.125 0-.208.083-.332.083-.084.459-.146.792-.54 1.042v-.791a4.04 4.04 0 00.332-.96zM3.17 31.43h-.11l-.073.074c.078.114.15.22.293.22.105 0 .164-.064.256-.11-.082-.133-.224-.115-.366-.184z"
+						fill="#fff"
+					/>
+					<path
+						d="M2.279 30.184a.336.336 0 00.25.111c.11 0 .165-.075.275-.075.229 0 .362.117.586.075 0 .033.036.042.036.075 0 .08-.073.108-.073.187 0 .099.11.127.11.225 0 .16-.201.136-.33.225a2.332 2.332 0 01-.197.124H4.6c-.043-.087-.112-.159-.112-.274 0-.23.17-.314.256-.524-.26 0-.398-.099-.659-.113-.131-.006-.299.007-.406-.036h-1.4zm13.299 0c.006.037.01.076.01.12 0 .149-.115.21-.183.336h-.44c-.324 0-.526-.06-.805-.224v-.113h.769c.126 0 .239-.044.327-.12l.322.001zm10.348 0c.163.208.253.478.301.798l-.79.085c-.03-.359-.05-.591-.033-.883h.522zm8.702 0c0 .003-.002.006-.003.008h-.146a.4.4 0 01-.003-.007h.152zm11.9 0c-.186.012-.334.066-.334.25 0 .37.485.351.842.394.183.023.288-.005.476-.005-.041.124-.08.203-.12.308h-1.19a.931.931 0 00-.264-.041c-.166 0-.282 0-.42.042h-.115a.186.186 0 00-.05-.042v-.525h.475c.055-.07.11-.112.11-.206 0-.06-.01-.119-.03-.175h.62zm.97 0l.087.232c-.197-.16-.352-.225-.604-.225-.065 0-.136-.004-.208-.006l.725-.001zm11.12 0a.558.558 0 01-.233.007c0 .033-.036.042-.036.075 0 .08.073.107.073.187 0 .098-.11.126-.11.225 0 .159.201.136.33.224.146.104.276.171.408.23h-1.908c.017-.146.145-.222.145-.38 0-.229-.17-.313-.256-.524.118 0 .21-.02.301-.043l1.285-.001zm.7 0a.394.394 0 01-.073.008.233.233 0 01-.056-.007h.13zm.202.419c.014.056.028.093.037.15-.087.032-.138.065-.22.112-.06-.042-.092-.07-.146-.112.114-.075.2-.113.33-.15zm-57.264.104c-.014.056-.027.094-.037.15.087.033.138.066.22.112.06-.042.092-.07.147-.112a1.109 1.109 0 00-.33-.15z"
+						fill="#fff"
+					/>
+					<path
+						d="M4.167 29.255l.028.01c.133.052.33.038.33.188 0 .182-.261.15-.44.15-.43 0-.672-.094-1.098-.15-.032.098-.11.154-.11.262 0 .065.046.093.073.15l-.073.074c-.174-.177-.174-.486-.42-.486-.166 0-.349.074-.349.243 0 .185.056.386.17.507H3.68c-.064-.025-.106-.07-.106-.151 0-.08.041-.117.073-.187h.55c.123 0 .168-.113.292-.113h.403c-.052-.19-.08-.34-.095-.496l-.629-.001zm21.654 0c.04.071.09.137.114.21h-.499v-.21h.385zm21.911 0a.314.314 0 000 .001zm9.558 0a.113.113 0 00-.042.093c0 .183.261.15.44.15.43 0 .672-.094 1.098-.15.032.099.11.155.11.262 0 .066-.046.094-.074.15l.074.075c.174-.178.174-.487.42-.487.165 0 .349.075.349.244 0 .264-.114.56-.349.61h-.13c-.075-.018-.128-.067-.217-.067-.14 0-.243.043-.354.068H57.33c.11-.028.215-.06.358-.068.21-.01.512.028.512-.188 0-.08-.041-.117-.073-.187h-.55c-.123 0-.169-.112-.292-.112h-.403c.04-.148.066-.271.083-.393h.325zm-42.035.948a.572.572 0 00.185-.33c.063.121.116.213.137.33h-.322zm10.148 0c.006-.106.017-.22.033-.35.21.08.37.199.489.35h-.522zm9.072 0c-.036-.094-.07-.15-.07-.256 0-.294.352-.374.64-.374.137 0 .311.047.311.187 0 .34-.595.139-.73.443h-.15zm11.432 0c-.096-.284-.412-.484-.703-.405 0-.108-.005-.183-.073-.262.279-.038.43-.15.714-.15.357 0 .32.636.677.636h.879c.064 0 .087-.06.146-.075.01 0-.11.025-.146 0l.096.256h-.726a2.682 2.682 0 00-.245 0h-.619z"
+						fill="#fff"
+					/>
+					<path
+						d="M2.147 28.325l-.075.06v.261c.068.01.11.038.183.038.137 0 .146-.197.146-.337a.05.05 0 00-.005-.022h-.249zm.504 0c-.019.087-.03.18-.03.284 0 .164.192.183.33.262.43.252.76.226 1.216.401h.629a4.205 4.205 0 01-.015-.401c-.06.112-.128.225-.256.225h-.513c-.293 0-.293-.412-.293-.712 0-.022.004-.041.011-.06l-1.079.001zm12.79 0v.18c0 .07.073.098.073.168 0 .094-.064.126-.11.206-.032-.145.005-.248-.073-.375-.059-.098-.189-.12-.267-.18l.377.001zm10.53 0a1.558 1.558 0 00-.202.775c0 .064.023.12.053.173h-.385v-.938l.076-.01h.458zm9.056 0c.011.01.023.02.037.03v.13h-.842v-.16h.805zm11.71 0a.561.561 0 00.371.142c.184 0 .31-.058.444-.142h.179v.48c0 .05.073.06.073.111 0 .144-.067.219-.073.357-.015-.33-.436-.394-.77-.394-.123 0-.187.075-.31.075-.096 0-.142-.075-.238-.075h-.44c-.096 0-.141.075-.238.075-.15 0-.206-.14-.347-.187-.16-.052-.33-.033-.403-.188.226-.007.485-.071.562-.254h1.19zm12.404 0c.007.056.011.115.011.18 0 .163-.192.182-.329.262-.44.257-.773.224-1.245.412-.095.036-.222.04-.287.093h-.326c.022-.155.027-.308.027-.505.06.112.129.224.257.224h.512c.279 0 .292-.372.293-.666h1.087zm.56 0v.217c-.068.01-.11.037-.182.037-.112 0-.139-.13-.145-.253l.327-.001zm-23.393.137v.08c-.082.084-.132.145-.22.225-.086-.164-.228-.178-.365-.3.22-.051.375-.094.585-.005z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.972 27.396c.01.138.018.285.027.445l.11.112h.475c.083-.145.115-.257.183-.412a.272.272 0 00.074-.037l.073.075c-.083.302-.207.505-.263.765h1.08c.034-.087.142-.128.244-.128h.367c.178 0 .265-.127.439-.15v-.187c-.247.103-.39.224-.66.224-.393 0-.95.066-.95-.337 0-.098.054-.15.109-.225.082.043.142.075.238.075.2 0 .327-.116.463-.22H1.972zm13.462 0c.004.048.006.098.006.154v.794h-.376a.141.141 0 01-.063-.12c0-.3.403-.356.403-.656 0-.064-.01-.12-.03-.172h.06zm10.43 0c-.144.182-.318.318-.513.487a3.086 3.086 0 01-.04-.488l.553.001zm10.59 0v.154c-.11-.075-.178-.15-.311-.15-.37 0-.453.487-.824.487h-.476c-.22 0-.338.056-.549.112v-.187h.293c.124-.09.293-.122.293-.281a.25.25 0 00-.036-.134l1.61-.001zm9.955 0c.174.147.442.234.442.49 0 .164-.155.216-.256.338.05.043.096.083.142.12h-1.19a.305.305 0 00.023-.12c0-.305-.375-.281-.622-.45-.14-.094-.172-.227-.2-.379l1.662.001zm13.385 0l-.02.34-.11.113h-.475c-.082-.145-.115-.257-.183-.412a.274.274 0 01-.073-.037l-.074.075c.094.34.24.554.281.868h-1.086V28.3c0-.127-.133-.188-.256-.188h-.366c-.179 0-.266-.126-.44-.15v-.187c.247.103.39.225.66.225.393 0 .951.065.951-.337 0-.098-.055-.15-.11-.225-.082.042-.142.075-.238.075a.551.551 0 01-.326-.116h1.865zm-57.399.95c-.009-.02-.028-.034-.031-.055h-.146c-.028.02-.05.037-.071.054h.248zm23.117-.002c.29-.036.49-.054.796-.054l.084-.66c-.146.278-.302.49-.422.714h-.458zm8.709.001v-.084h.732c.023.034.045.06.073.085h-.805zm13.33 0l.178-.121v.12h-.178zm11.822-.001a1.679 1.679 0 01-.002-.083c0-.033.032-.047.037-.075h.146c.06.042.092.07.146.113v.045h-.327zM46.21 28.11c.123 0 .201.056.274.15-.105.019-.164.075-.274.075-.083 0-.124-.052-.202-.075v-.112c.074-.014.124-.038.202-.038z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.91 26.467c.016.313.04.605.062.948h2.01c.107-.081.22-.154.378-.154.165 0 .274.009.421.074 0-.191.064-.294.073-.486-.068.032-.105.075-.183.075h-.476c-.279-.225-.805-.085-.805-.45v-.007h-.58a.184.184 0 01-.152.082.544.544 0 01-.308-.082h-.44zm33.918 0c-.11.079-.223.165-.362.24h-.146a.63.63 0 00-.096-.24h.604zm.927 0l-.005.016a.343.343 0 00-.004-.016h.01zm10.502 0c-.13.025-.258.056-.258.184 0 .21.251.28.457.28.092 0 .124-.07.202-.112h.146v.3c-.17-.047-.265-.122-.44-.15-.265-.042-.425-.037-.695-.037-.174 0-.403.028-.403.206 0 .122.062.206.144.276h-1.66c-.015-.082-.029-.169-.058-.258.343 0 .842.07.842-.28 0-.136-.142-.207-.274-.207-.092 0-.128.075-.22.075-.165 0-.256-.075-.421-.075a.747.747 0 00-.029-.203l2.667.001zm12.601 0c-.017.306-.04.599-.062.947H57.93c-.157-.11-.291-.258-.515-.258a.911.911 0 00-.421.075c0-.192-.064-.295-.074-.487.069.033.105.075.183.075h.476c.251-.202.701-.109.79-.353l1.49.001zm-44.483.948c-.045-.122-.14-.214-.263-.296a.86.86 0 01.22-.187c.05.19.089.317.103.483h-.06zm9.936 0v-.193c.145-.021.228-.128.374-.128.124 0 .207.085.332.085a1.492 1.492 0 01-.152.235h-.553zm9.535 0c-.082-.134-.294-.173-.477-.221v-.225c.252.019.389 0 .64 0 .261 0 .27.412.532.412.164 0 .2-.225.366-.225.169 0 .233.188.402.188.064 0 .092-.043.147-.075v.146h-1.61zm23.612-.633l.073.15a.307.307 0 01-.146 0 .756.756 0 01.073-.15zm-21.815-.188a1.84 1.84 0 00-.114.263c.005-.038 0-.07 0-.113 0-.075.071-.093.114-.15zm-33.325.292l-.073.15a.31.31 0 00.146 0 .756.756 0 00-.073-.15z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.895 25.538a1.052 1.052 0 00-.006.112c0 .317.008.583.021.835h.44a.213.213 0 01-.095-.18c0-.14.137-.224.274-.224.165 0 .312.131.312.3a.189.189 0 01-.032.104h.58c.007-.222.296-.217.513-.254.385-.066.563-.3.952-.3v-.304c-.325.042-.512 0-.842-.033-.002-.023-.011-.04-.02-.056H1.895zm23.76 0a.155.155 0 01.037.101c0 .164-.192.188-.33.263h-.146v-.337a.705.705 0 00.036-.026l.404-.001zm11.324 0a.195.195 0 00-.011.064v.375c0 .214-.14.313-.214.509h-.009c-.032-.143-.144-.247-.29-.247-.267 0-.44.113-.626.246h-.604c-.164-.243-.485-.328-.783-.508.316 0 .494.009.805-.038.17-.023.247-.15.421-.15.275 0 .385.262.66.262.27 0 .603-.06.603-.337a.252.252 0 00-.058-.175l.106-.001zm9.433 0c.019.031.036.064.074.064.098 0 .172-.03.244-.064h.448c.076.089.13.214.26.214h.44v.187c-.257.023-.435.089-.586.3l.073.075v.15c-.034.008-.07.014-.107.021H44.59c-.042-.154-.117-.28-.117-.471 0-.084.082-.098.146-.15h.183c.005.033.037.056.037.094 0 .07-.046.108-.073.168.137.103.233.188.402.188.174 0 .367-.085.367-.263 0-.238-.32-.309-.55-.262a.989.989 0 00-.097-.251h1.524zm13.473 0v.008c0 .36-.01.656-.027.94h-1.49a.287.287 0 00.016-.098c0-.229-.292-.224-.512-.262-.384-.065-.563-.3-.952-.3v-.288h2.965zm-.64.439c.137 0 .274.084.274.224 0 .192-.215.262-.402.262a.186.186 0 01-.183-.187c0-.168.146-.3.31-.3zm-13.051-.263c.082 0 .119.061.183.113v.224c-.06.01-.1.038-.165.038-.11 0-.146-.09-.238-.15.046-.112.1-.225.22-.225z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.946 24.608a.256.256 0 00-.057.162c0 .126.073.192.073.318 0 .19-.05.305-.067.469h2.097a.115.115 0 01-.017-.057c0-.243.307-.225.513-.337a.53.53 0 00-.146-.187c-.247.173-.43.262-.733.262-.137 0-.292-.066-.292-.206 0-.234.31-.178.512-.28l.27-.144h-.826l-.066.068a1.662 1.662 0 00-.142-.068H1.946zm32.781 0a.256.256 0 00.049 0h.618v.001c0 .127.114.225.238.225.128 0 .2-.075.33-.075.123 0 .187.075.31.075.275 0 .696-.393.696-.112v.412c0 .098.073.145.073.243 0 .075-.042.12-.062.18h-.106c-.16-.178-.614-.066-.93-.161-.554-.169-.847-.407-1.428-.45.004-.136.039-.215.066-.338h.146zm12.304 0c.004.044.005.085-.033.114-.183.14-.549.004-.549.243 0 .234.325.216.403.43a1.05 1.05 0 00-.476.113c.015.012.025.03.036.048h-1.525c-.06-.117-.122-.225-.122-.385 0-.164.096-.337.256-.337h.623c.153 0 .295-.09.346-.226h1.041zm12.847 0a.262.262 0 01.007.058c0 .126-.074.191-.074.318 0 .226.072.348.074.572h-2.966v-.015c.325.042.513 0 .842-.033.005-.047.037-.065.037-.112 0-.244-.307-.225-.513-.337a.53.53 0 01.147-.188c.247.174.43.263.732.263.137 0 .293-.066.293-.206 0-.234-.311-.178-.513-.281l-.072-.038 2.006-.001zm-34.627.95c.058-.044.103-.087.184-.087.084 0 .172.029.22.087h-.404zm21.477 0c.087-.041.17-.087.288-.087.071 0 .12.038.161.087h-.449z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.779 23.68v.453c.073.099.133.188.256.188h.806c.206 0 .33.126.439.3l-.007.006H4.1c.43-.23.735-.405 1.194-.718v-.15a1.031 1.031 0 01-.177-.08l-3.337.001zm23.965 0c.136.041.267.087.441.087.085 0 .141-.04.203-.087h.043c.043.184.074.37-.008.574-.013.042.005-.127-.036-.113-.156.052-.22-.187-.385-.187-.293 0-.402.3-.695.3-.179 0-.311-.173-.311-.356 0-.099.055-.171.135-.218h.613zm9.172 0c.063.104.132.176.22.274h.403v-.274h.666a.343.343 0 00-.007.068c0 .201.243.244.44.244.1 0 .182-.085.182-.188 0-.05-.015-.09-.039-.125l.253.001c.004.044.006.092.006.143 0 .126-.05.192-.073.318h-.55c-.2 0-.31-.112-.512-.112-.301 0-.511.29-.512.597h-.618c.114-.008.215-.085.215-.204 0-.22-.096-.468-.311-.468-.046 0-.05.056-.092.075.039-.126.07-.239.034-.35l.295.001zm10.887 0c.237.113.474.187.792.274.252.07.622-.005.622.262 0 .136-.142.173-.183.3-.012.036-.007.074-.004.11h-1.04a.318.318 0 00.02-.11c0-.061-.037-.09-.037-.15-.261-.023-.399-.112-.66-.112h-.512c-.206 0-.348-.117-.44-.3.162-.117.435-.097.472-.274h.97zm12.764 0v.087c-.06 0-.087.037-.147.037-.064 0-.087-.056-.146-.075v-.049H56.48v.124c.526.36.85.538 1.392.824l2.006-.001c-.03-.133-.164-.224-.305-.224-.398 0-.659 0-1.006.188l-.074-.075c.11-.173.234-.3.44-.3h.805c.124 0 .183-.089.256-.187v-.349h-1.427zm-55.503.948c-.282-.12-.523-.12-.865-.12-.1 0-.195.046-.254.12h1.119zm31.516 0a.905.905 0 00.007-.037c.051.013.089.032.14.037h-.147zm22.086-.45c-.005.048-.02.072-.038.115a.235.235 0 01.038-.114zm-53.46-.475v.168c.06 0 .088.038.147.038.064 0 .087-.057.146-.075v-.131c-.114.004-.178 0-.292 0zm1.906.58c.005.048.019.072.038.114a.236.236 0 00-.038-.114z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.879 22.751c.13.128.313.165.522.165.202 0 .311-.112.513-.112.178 0 .297.084.403.224h.512c.114 0 .174.08.256.15v.225c-.04.01-.064.037-.11.037-.324 0-.49-.15-.805-.224-.15-.033-.252-.019-.403 0-.384.047-.6.13-.988.15v.333h3.337c-.155-.086-.278-.184-.482-.184-.155 0-.228-.107-.329-.224.087-.108.233-.099.293-.225.048-.1.013-.224.036-.315H1.879zm56.766 0a.599.599 0 00-.188.173h-.513c-.114 0-.173.08-.256.15v.225c.041.009.064.037.11.037.325 0 .49-.15.805-.225.151-.032.252-.018.403 0 .384.047.6.132.989.15v.438h-1.428v-.082c-.115 0-.179.005-.293 0v.082H56.48v-.026c.256-.084.384-.262.659-.262.156 0 .229-.108.33-.225-.087-.107-.234-.098-.293-.224-.031-.066-.028-.141-.028-.211h1.497zm1.095 0a.928.928 0 01-.368.062c-.106 0-.187-.031-.268-.061h.636zm-34.608.95a.498.498 0 01.249-.065c.14 0 .254.03.364.064h-.613zm1.257-.002l.035-.026v-.45c-.069-.037-.105-.074-.183-.074-.421 0-.64.187-1.062.187-.142 0-.233-.051-.366-.005v-.182c.229-.155.385-.262.66-.262.37 0 .576.056.946.075-.121.26-.044.495.013.736h-.043zm8.232.001a.34.34 0 00-.028-.064h.288c.011.023.023.044.035.064h-.295zm.919-.001v-.063c-.298-.22-.522-.468-.879-.375.174-.21.385-.262.66-.262.09 0 .127.075.219.075h.659c.206 0 .306.15.512.15.115 0 .21-.047.256-.15 0 .248.05.407.068.624h-.253c-.068-.099-.206-.152-.254-.287h-.22c-.022.116-.083.185-.103.287h-.665zm9.294 0a.215.215 0 00.005-.045c0-.257-.334-.28-.586-.28v-.15h.293c.27.145.517.014.769.187.177.123.333.213.489.287h-.97z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.656 21.821l.013.14c.014.182.183.299.366.299.142 0 .206-.113.348-.113.124 0 .178.075.275.15-.188.323-.577.131-.88.337a.552.552 0 00.1.135h2.756a.14.14 0 01.074-.097c.15-.08.293.018.44-.075.187-.122.25-.337.475-.337.091 0 .151.065.183.15 0-.225.096-.342.146-.562-.068.032-.105.075-.183.075a.805.805 0 01-.396-.103l-1.944.001a.153.153 0 01.034.102c0 .121-.1.224-.22.224-.17 0-.222-.167-.29-.326H1.656zm23.756 0a.197.197 0 00-.013.072c0 .127.147.14.256.187.252.108.385.263.66.263.045 0 .064-.038.11-.038v.262c-.088-.056-.152-.112-.257-.112h-.55c-.104 0-.15-.075-.255-.075h-.403c-.265 0-.402-.288-.403-.559h.855zm11.702 0c.037.083.074.164.074.278 0 .098-.055.15-.074.244a.272.272 0 00-.073.037c-.091-.22-.256-.374-.494-.374-.453 0-.59.524-1.043.524-.302 0-.481-.094-.77-.005v-.482c.074.019.124.037.202.037.11 0 .238-.037.238-.15a.174.174 0 00-.037-.11l1.977.001zm19.806 0v.147c-.064.024-.096.075-.165.075-.091 0-.123-.07-.2-.112.018-.043.033-.075.046-.11h-.758c.052.173.125.288.125.484.032-.084.092-.15.183-.15.225 0 .289.216.476.337.147.094.289-.004.44.075.077.041.082.117.082.201h1.497a.443.443 0 01.214-.05c.096 0 .17.024.244.05h.636a.518.518 0 00.255-.238c-.302-.206-.69-.014-.878-.337.096-.075.15-.15.274-.15.142 0 .206.112.348.112.183 0 .352-.117.366-.3l.003-.035-1.33.001c-.053.12-.114.222-.247.222a.225.225 0 01-.22-.222H56.92zm.953.56l.076.037v.187h-.114c.01-.084.02-.14.038-.225zm-52.726-.533h-.292v.225c.064.023.096.074.164.074.092 0 .124-.07.202-.112a1.31 1.31 0 01-.074-.187zm-1.242.636l-.075.038v.187h.113a1.858 1.858 0 00-.038-.225z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.59 20.892c.006.312.032.59.066.948h1.297a1.331 1.331 0 00-.04-.086l.074-.075h.11c.11.068.264.081.332.162l1.944-.001a.484.484 0 01-.263-.423c0-.164.17-.262.33-.262.307 0 .718.192.622.487.083-.136.078-.253.183-.375v-.262c-.082.033-.128.075-.22.075a.767.767 0 01-.52-.188H2.816a.285.285 0 01.024.113c0 .225-.147.45-.367.45h-.366c-.04-.085-.11-.127-.11-.225 0-.142.064-.256.162-.338h-.57zm58.116 0a.417.417 0 01.069.233c0 .099-.069.14-.11.225H59.3c-.22 0-.366-.225-.366-.45v-.008h-2.814a.85.85 0 01-.371.084c-.092 0-.137-.043-.22-.075v.262c.105.122.101.239.183.374-.096-.294.316-.486.623-.486.16 0 .33.098.33.262 0 .332-.335.524-.66.524-.078 0-.114-.042-.183-.075.006.028.013.054.02.078h.76l.026-.078h.292v.078h1.392v-.003c0-.178.22-.173.366-.262h.11l.073.075c-.031.06-.056.127-.083.19h1.33c.035-.369.063-.648.074-.948h-.475zm-35.15.948v-.003c0-.206.165-.374.366-.374h.403c.115 0 .17.08.256.15-.054.093-.138.141-.17.226h-.855zm10.58 0c-.065-.085-.207-.115-.329-.115.028-.127.064-.225.005-.337h.251c.495.3.833.45 1.41.45.233 0 .522.008.567-.226.013.09.043.16.073.227h-1.977zm22.753-.752c.142 0 .238.136.238.28 0 .141-.155.207-.293.207-.105 0-.15-.075-.256-.075-.142 0-.206.112-.348.112-.146 0-.215-.098-.348-.15l.037-.337c.137.132.252.188.44.188.224 0 .306-.225.53-.225zm-54.006.104c-.142 0-.238.136-.238.281 0 .14.156.206.293.206.105 0 .151-.075.257-.075.141 0 .205.113.347.113.147 0 .215-.099.348-.15l-.037-.337c-.137.13-.251.187-.439.187-.224 0-.306-.225-.53-.225z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.634 19.963c.02.053.035.11.035.18 0 .085-.073.122-.073.207-.009.21-.01.39-.007.561h.571a.698.698 0 01.443-.15c.098 0 .177.063.215.15h2.686a.43.43 0 01-.137-.318c0-.126.146-.154.146-.28 0-.162-.213-.192-.22-.35H1.635zm58.477 0a.46.46 0 00-.006.076c0 .085.073.122.073.206.011.253.011.464.004.666h-.475c-.106-.164-.321-.254-.536-.254-.134 0-.233.117-.237.254H56.12c.166-.083.287-.226.287-.422 0-.127-.146-.155-.146-.281 0-.119.115-.166.179-.245h3.671zm-1.873.02h.366c.005.056.037.09.037.15 0 .098-.106.15-.202.15-.123 0-.174-.084-.274-.15l.073-.15zm-54.701.105H3.17c-.005.056-.037.088-.037.15 0 .098.105.15.202.15.123 0 .173-.085.274-.15l-.073-.15z"
+						fill="#fff"
+					/>
+					<path
+						d="M3.127 19.034c-.107.573-1.568-.003-1.568.66 0 .117.043.198.075.287h3.66v-.006c0-.248.292-.253.512-.356.22-.103.275-.341.513-.374.251-.033.407.004.659.037.024-.1.057-.17.092-.248h-.124c-.151.068-.32.061-.518.061-.108 0-.181-.039-.27-.06l-.47-.001a.962.962 0 00-.212.061c-.251.103-.274.45-.549.45-.201 0-.33-.207-.33-.413 0-.037.013-.07.033-.098H3.127zm53.246 0c.184.133.23.406.473.406.2 0 .328-.202.33-.407l1.512.001c.264.393 1.526-.06 1.526.556 0 .16-.08.251-.103.391h-3.672a.172.172 0 00.041-.11c0-.248-.293-.253-.512-.356-.22-.103-.275-.342-.513-.375-.251-.032-.407.005-.659.038a.982.982 0 00-.047-.144h1.624zm1.663.219c.183 0 .421.037.421.225 0 .238-.316.262-.549.262-.21 0-.325-.099-.512-.188v-.112c.233-.108.38-.187.64-.187zm1.263.3h.403v.224c-.069.038-.12.075-.201.075-.129 0-.193-.075-.312-.112.037-.07.074-.117.11-.188zm-55.562-.196c-.183 0-.42.038-.42.225 0 .239.315.262.548.262.21 0 .325-.098.513-.187v-.113c-.234-.107-.38-.187-.64-.187zm-1.263.3h-.403v.225c.069.037.12.074.202.074.128 0 .192-.074.311-.112-.037-.07-.073-.117-.11-.187z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.451 18.105c.02.22.21.323.401.447.142.094.179.262.348.262.229 0 .284-.3.512-.3.243 0 .422.22.422.469a.385.385 0 01-.007.07H4.63a.236.236 0 01.188-.09h.402c.124-.172.366-.196.366-.411 0-.2-.18-.262-.255-.447h-.809c-.085.015-.14.072-.235.072a.247.247 0 01-.173-.072h-.358a.78.78 0 01.073.11c-.146.08-.233.187-.403.187a.259.259 0 01-.256-.262c0-.013.003-.024.006-.035h-.723a.357.357 0 01-.216.072.266.266 0 01-.188-.071l-.598-.001zm4.857 0c-.228.03-.336.28-.392.522.302 0 .471-.075.769-.038.206.024.393.07.475.263a3.03 3.03 0 00.33-.712 2.482 2.482 0 01-.494.037c-.22 0-.353-.057-.551-.071l-.137-.001zm7.27 0c-.228.142-.423.268-.736.268-.113 0-.09-.15-.119-.267l.856-.001zm33.087 0a6.212 6.212 0 01-.839.043h-.402c-.147 0-.22.112-.366.112h-.476c-.293 0-.458-.112-.75-.112-.225 0-.344.075-.568.075-.098 0-.152-.058-.22-.119l3.62.001zm9 0c.098.102.156.261.192.418-.302 0-.471-.075-.769-.038-.205.023-.393.07-.475.262a2.9 2.9 0 01-.306-.642h1.358zm2.929 0a.257.257 0 01-.247.193c-.17 0-.256-.108-.403-.187l.004-.007-1.565.001c-.083.113-.196.185-.196.343 0 .215.242.238.366.412h.403c.11 0 .22.074.22.187v.005h1.512a.303.303 0 01-.048-.174c0-.248.178-.468.42-.468.23 0 .284.3.513.3.17 0 .206-.169.348-.262.157-.103.314-.19.376-.343h-1.703zm-52.436.949a.494.494 0 00-.114-.014c-.14 0-.247-.004-.357.014h.47zm.912-.001c.018-.038.036-.079.054-.126a.577.577 0 01-.178.126h.124zm47.68 0c-.03-.074-.066-.14-.1-.23.197.2.416.186.695.186.151 0 .234-.075.385-.075.224 0 .361-.009.567.075a.365.365 0 01.076.043H54.75zm2.81-.83c.092 0 .128.112.128.206 0 .126-.077.243-.201.243-.1 0-.201-.065-.201-.168 0-.155.123-.281.274-.281zm-53.347.104c-.091 0-.128.113-.128.206 0 .127.078.244.201.244.101 0 .202-.066.202-.169a.279.279 0 00-.275-.28z"
+						fill="#fff"
+					/>
+					<path
+						d="M2.47 17.524c.065.092.114.19.114.328a.33.33 0 01-.131.264h.723c.01-.026.027-.047.03-.077h.403c.064 0 .11.032.147.077h.358a.214.214 0 01-.065-.152c0-.206.131-.354.31-.44h-1.89zm3.025 0c.134.12.216.328.42.328.143 0 .147-.174.257-.262a.675.675 0 01.095-.065l-.772-.001zm2.11 0a.583.583 0 01.141.14c.042-.028.069-.051.11-.074v-.066h-.251zm5.17 0c.322.287.698.377 1.128.26v.15a3.155 3.155 0 00-.324.183h-.856a.24.24 0 00-.028-.07c-.132-.206-.389-.117-.549-.3-.058-.067-.064-.143-.067-.223h.696zm31.78 0c.111.207.404.298.668.298.124 0 .202-.117.202-.243a.268.268 0 00-.006-.055h2.527c-.394.373-.806.529-1.28.593h-3.621a.562.562 0 00-.037-.033c.246-.42.64-.442 1.053-.56h.494zm9.505 0a.668.668 0 00-.03.037l-.053-.037h.083zm2.124 0c-.089.109-.173.223-.326.223-.123 0-.142-.13-.217-.223h.543zm3.057 0a.497.497 0 00-.051.223c0 .192.16.337.348.337.251 0 .32-.276.384-.524h.146c.151.112.257.22.257.412a.38.38 0 01-.027.145h-1.703a.26.26 0 00.01-.07c0-.047-.033-.07-.038-.112h-.402c-.108 0-.162.09-.216.18l-1.565.001a.478.478 0 00.06-.107h.769c.11 0 .165.075.274.075.124 0 .238-.098.238-.224a.443.443 0 00-.153-.336h1.669zm-57.192.592c-.106-.097-.152-.28-.197-.452h-.146c-.151.113-.256.22-.256.412 0 .014 0 .027.002.04h.597zm3.282.001v-.002h-.77a.234.234 0 00-.038.003h.808zm1.114 0a1.892 1.892 0 00-.09-.002.353.353 0 00-.047.003h.137zm47.863.001a6.346 6.346 0 01-.024-.07c.187.033.302.038.494.038.252 0 .389-.076.64-.076a.33.33 0 01.248.108h-1.358z"
+						fill="#fff"
+					/>
+					<path
+						d="M2.151 16.944c.023.063.04.135.067.208.062.163.17.268.252.384h1.888a.914.914 0 01.386-.084h.55a.29.29 0 01.2.084h.773c.124-.07.254-.084.418-.084h.622c.119 0 .215.027.298.084h.251v-.121c-.17-.024-.16-.234-.256-.375a.322.322 0 00-.102-.096h-2.54c-.08.057-.138.134-.25.134h-.732c-.045-.053-.099-.091-.142-.134h-.412a.22.22 0 01-.179.096.502.502 0 01-.294-.095l-.798-.001zm10.632 0a.798.798 0 00.315.067.79.79 0 00.3-.067H14a.505.505 0 01.123.142c-.357 0-.526.224-.879.262-.215.023-.398-.042-.549.112.026.027.053.053.08.077h-.696c-.003-.076-.003-.153-.043-.227-.114-.21-.361-.187-.604-.187-.082 0-.119.075-.201.075-.092 0-.138-.038-.22-.075a.613.613 0 01.12-.18l1.652.001zm35.021 0c.002.165.066.325.22.366v.15a2.625 2.625 0 01-.078.077h-2.527c-.014-.067-.05-.115-.068-.19h.403c.186 0 .187-.245.27-.403h1.78zm1.686 0a.34.34 0 00-.039.142h-.695c-.097 0-.14-.074-.197-.142h.93zm8.335 0a.696.696 0 00-.027.03h-.732a.224.224 0 01-.112-.03h-2.785c-.09.14-.085.343-.251.366v.188c.022.012.04.024.058.037l.08.001a.493.493 0 01.41-.188h.623c.206 0 .357.023.513.15a.261.261 0 01.039.039h.542c.08-.097.161-.19.297-.19h.55c.198 0 .406.065.542.189h1.669c.08-.164.235-.279.315-.488.013-.035.024-.07.034-.104h-1.765zm-13.764.592c.149-.043.3-.098.448-.188 0 .071.017.134.046.189h-.494z"
+						fill="#fff"
+					/>
+					<path
+						d="M1.334 16.247c-.01.163-.022.317.116.395.187.108.361 0 .549.112.084.05.123.12.152.203h.798c-.09-.059-.175-.137-.291-.203-.12-.065-.293-.065-.293-.206 0-.037.036-.056.036-.093.147-.02.234-.038.385-.038.306 0 .677.098.677.412a.225.225 0 01-.041.13l.412-.001a.22.22 0 01-.078-.166c0-.188.142-.272.293-.375.071-.048.121-.113.171-.17H1.334zm10.958 0a.613.613 0 01-.183.066c-.384.07-.613.075-1.006.075-.151 0-.238-.01-.385-.038a.8.8 0 01.064-.103h1.51zm3.621 0c0 .12.069.22.004.328-.105.173-.293.15-.495.15-.334 0-.416-.412-.75-.412-.252 0-.389.075-.64.075a.81.81 0 01-.44-.141h2.321zm30.49 0c-.338.14-.67.138-.906.478v-.478h.906zm2.6 0c-.021.073-.077.123-.138.178h-.256v-.178h-1.354a.494.494 0 01-.147.328c-.275.267-.586.145-.952.262a.228.228 0 00-.132.12h1.78v-.008c-.188-.037-.44-.028-.44-.224 0-.103.119-.15.22-.15.15 0 .256.145.256.3.238-.047.412-.113.622 0 .04.02.07.051.096.083h.931c.032-.066.072-.125.072-.214 0-.195.002-.34-.039-.497h-.52zm8.643 0c.023.023.049.046.079.066.15.103.292.187.292.374 0 .132-.105.183-.193.271h1.766c.038-.127.065-.238.185-.308.187-.112.361-.005.549-.112.108-.061.124-.169.12-.291h-2.798zm-52.69.71a.308.308 0 01.117-.053c.307-.066.609.01.769-.262-.183-.171-.309-.343-.52-.396l1.164.001c-.109.103-.278.17-.278.339 0 .379.554.21.915.28.14.028.271.03.374.092h-2.54zm6.175 0c.241-.244.65-.195 1.051-.195.25 0 .41.119.6.196h-1.65zm2.266.002c.103-.04.204-.084.34-.084.103 0 .19.03.263.085h-.603zm40.77-.001l.006-.009c.11-.159.284-.15.476-.187.361-.07.915.099.915-.28 0-.108-.069-.175-.147-.235h.82c-.103.078-.194.184-.307.29.16.272.462.197.769.263.112.023.17.11.254.157h-2.786zm4.82-.645c.151 0 .238.019.384.037 0 .038.037.056.037.094 0 .14-.174.14-.293.206-.224.126-.33.3-.586.3a.225.225 0 01-.22-.225c0-.314.371-.412.678-.412zm-42.742.374h.622c-.045.136-.15.225-.293.225-.233 0-.361-.084-.585-.15.1-.023.15-.075.256-.075z"
+						fill="#fff"
+					/>
+					<path
+						d="M5.564 15.318c-.55.056-1.029.163-1.698.294-.682.131-1.071.206-1.758.3-.306.042-.768-.094-.768.224 0 .043-.003.087-.006.13H4.22c.062-.071.125-.13.232-.13.146 0 .22.113.366.113h.366a.57.57 0 01.14.016h1.162a.25.25 0 00.089-.185c0-.192-.293-.201-.293-.393 0-.173.33-.126.33-.3 0-.028-.01-.05-.027-.068l-1.021-.001zm1.194 0v.107c.041.01.064.037.11.037.216 0 .36-.068.523-.144h-.633zm10.46 0c.041.048.082.1.127.152-.096.117-.197.187-.348.187-.256 0-.43-.121-.567-.337l.002-.002h-.723c-.052.033-.109.068-.158.115.21.056.476.093.476.318 0 .3-.403.281-.696.281-.183 0-.402-.056-.402-.243 0-.16.178-.183.256-.319a1.55 1.55 0 00-.33-.037c-.366 0-.567.15-.933.15-.403 0-.604-.225-1.007-.225-.334 0-.42.337-.732.45-.532.195-1.097.039-1.401.457h1.51c.206-.112.337-.308.605-.308.295 0 .481.177.695.308h2.321c0-.015.002-.03.004-.046.032-.187.197-.3.384-.3.238 0 .348.188.586.188.334 0 .714-.145.714-.487a.427.427 0 00-.13-.302h-.253zm29.897 0c-.022.15-.01.345-.08.49-.087.182-.238.224-.403.336-.075.052-.152.09-.228.122h-.907v-.197c0-.07-.073-.098-.073-.168 0-.09.115-.084.183-.131.173-.121.278-.263.315-.452h1.193zm.818 0a.517.517 0 01-.098 0h.098zm7.083 0v.003c-.041.01-.064.038-.11.038a.891.891 0 01-.27-.04h.38zm1.915 0c.297.052.608.117.977.19.682.13 1.071.206 1.757.3.307.041.77-.094.77.224 0 .078.009.159.01.234h-2.799c-.104-.106-.167-.234-.324-.234-.146 0-.22.112-.366.112h-.366a.555.555 0 00-.351.122h-.82c-.103-.08-.22-.15-.22-.29 0-.192.293-.202.293-.393 0-.161-.283-.133-.325-.265h1.764zm-9.675.948c.011-.117-.001-.245-.001-.384.389-.019.6-.187.988-.187.311 0 .476.126.77.225.168.056.365-.014.438.15.033.069.056.133.072.195h-.52a.234.234 0 00.01-.065c0-.088-.097-.13-.184-.13-.091 0-.137.037-.22.074v.122h-1.353zm-9.556-.87c-.037.27-.05.435-.073.71-.16-.257-.302-.383-.513-.598.156.037.243.075.403.075.1 0 .16-.09.183-.188z"
+						fill="#fff"
+					/>
+					<path
+						d="M14.606 14.388c.03.037.066.07.066.126 0 .14-.1.262-.238.262-.233 0-.334-.187-.567-.187h-.403c-.187 0-.288.112-.476.112-.389 0-.695-.004-.951-.3v-.013h2.569zm2.484 0a.18.18 0 01.072.089c.078.234-.037.44.11.636.064.088.138.154.2.223h-.254c-.095-.106-.193-.185-.349-.185-.186 0-.3.064-.436.184h-.724c.093-.06.171-.117.171-.24 0-.234-.237-.272-.402-.431-.09-.087-.162-.186-.233-.276h1.845zm9.448 0c-.363.16-.57.576-1.03.576h-.439c-.306 0-.43-.258-.732-.3-.164-.023-.27.023-.439.037.055-.107.06-.262.183-.262.133 0 .197.113.33.113.228 0 .361-.005.585-.038.101-.014.212-.05.264-.126h1.278zm10.092 0c.058.09.197.126.32.126.188 0 .31-.069.456-.125l.22-.001v.276c-.068-.042-.118-.075-.2-.075-.11 0-.165.075-.275.075h-.33c-.227 0-.347-.18-.516-.277l.325.001zm11.472 0c-.003.05-.005.102-.005.164-.183 0-.274.112-.457.112-.124 0-.188-.075-.312-.075-.096 0-.22.033-.22.131 0 .127.079.244.202.244.115 0 .179-.075.293-.075.288 0 .393.224.64.374-.12.033-.198.066-.313.073h-.095c-.204-.025-.306-.223-.525-.223-.135 0-.176.095-.195.222h-1.193a.856.856 0 00.015-.166c0-.191-.078-.294-.146-.468.22 0 .476-.075.476-.3a.48.48 0 000-.012l1.835-.001zm-41.517.949a.168.168 0 00-.12-.044c-.339 0-.628.016-.901.044h1.02zm.806-.001c.085-.04.175-.083.283-.118.15-.052.27.014.402-.075-.087 0-.133-.038-.22-.038-.434 0-.736-.018-1.098.225v.006h.633zm47.245 0c-.177-.054-.321-.152-.535-.223-.151-.051-.27.014-.403-.075.087 0 .133-.037.22-.037.435 0 .736-.019 1.098.225v.11h-.38zm.531 0a.121.121 0 01-.005-.035c0-.075.074-.113.147-.113a8.6 8.6 0 011.622.147h-1.764zm-38.774-.897c.101 0 .138.094.22.15-.1.07-.146.187-.274.187-.101 0-.152-.065-.239-.112v-.112c.11-.052.17-.113.293-.113zm.367.226a.232.232 0 01.113.038c-.042-.01-.07-.024-.113-.038z"
+						fill="#fff"
+					/>
+					<path
+						d="M26.244 13.459c-.127.026-.223.079-.223.212 0 .174.215.225.384.225.23 0 .531-.028.531-.262a.268.268 0 00-.07-.176l1.213.001.065.1v.112c-.375.015-.466.37-.695.675h-.623a.706.706 0 00-.288.06h-1.279a.173.173 0 00.03-.098c0-.23-.343-.117-.513-.262-.229-.197-.352-.45-.659-.45v-.137h2.127zm7.795 0v.1h-.257c-.036-.028-.068-.056-.11-.075a.759.759 0 00.015-.025h-1.228c.131.059.182.212.335.212h.805c.238 0 .298.244.513.338.48.21.791.337 1.318.337h.659c.084 0 .153.024.215.06h.326a.177.177 0 01-.029-.098c0-.182.234-.182.403-.225.188-.046.28-.187.476-.187.096 0 .133.075.22.113.045-.122.086-.188.146-.3-.252.117-.416.187-.696.187-.247 0-.549-.084-.549-.337 0-.038.01-.072.027-.1h-.574c-.128.34-.586.475-.99.475-.156 0-.234-.07-.366-.15.047-.12.105-.23.173-.325h-.832zm4.02 0c-.033.056-.071.102-.14.102-.106 0-.183-.05-.257-.102h.398zm-26.022.948v-.249c.073.033.123.057.201.075l2.142-.187h.549l2.013-.187c0 .033.037.042.037.075 0 .117-.147.145-.147.262 0 .13.157.144.258.211h-1.845c-.123-.156-.245-.286-.463-.286-.11 0-.22.075-.22.187 0 .042.02.072.044.1l-2.57-.001zm25.369.001c.066-.025.136-.049.22-.062v.062h-.22zm8.86-.001c-.01-.252-.297-.361-.549-.361h-.293c-.16 0-.242.112-.402.112-.316 0-.472-.187-.788-.187-.196 0-.494-.07-.494.131 0 .047.06.052.073.094-.169-.108-.237-.225-.366-.375v-.187c.101-.014.17-.037.275-.037 1.29 0 1.995.271 3.276.449.44.06.696.178 1.135.112a2.162 2.162 0 00-.032.248h-1.835zm-28.628-.623a.548.548 0 00-.073.187.453.453 0 01-.147-.15c.087 0 .138-.01.22-.037z"
+						fill="#fff"
+					/>
+					<path
+						d="M26.022 12.529c-.068.093-.135.188-.257.188-.22 0-.348-.037-.568-.037-.208 0-.448-.036-.59-.151h1.415zm1.8 0c.088.023.182.04.286.076a.419.419 0 01-.366.187c-.225 0-.289-.262-.513-.262h-.476l-.027-.001h1.096zm7.99 0a.228.228 0 01-.126.038.3.3 0 01-.15-.038h-3.251c.367.293 1.131-.037 1.131.45 0 .127-.133.187-.256.187-.147 0-.22-.112-.366-.112-.339 0-.604.066-.806.337.151.08.275.024.44.075a.24.24 0 01.03.012h1.229c.043-.078.078-.162.168-.162.083 0 .12.061.183.113v.05h.833c.21-.294.516-.462.906-.462.156 0 .312.103.312.262 0 .073-.013.139-.035.199h.574c.047-.08.147-.124.248-.124.1 0 .137.112.238.112.105 0 .15-.112.256-.112.122 0 .205.064.291.124h.398c.026-.04.048-.086.08-.124-.275.056-.66-.047-.66-.337 0-.169.147-.244.147-.412a.228.228 0 00-.012-.075l-1.802-.001zm-11.694.95v-.2h.402c.21 0 .302-.187.513-.187.105 0 .15.074.256.074h.366c.174 0 .247-.15.421-.15.261 0 .417.174.531.413-.114.022-.25.025-.363.049h-2.126zm2.748 0c-.038-.05-.083-.096-.113-.163.266-.037.43-.004.696-.037.522-.07.81-.201 1.318-.337v.224a.21.21 0 00-.073.113c-.243-.056-.435.023-.622.187l.007.012h-1.213zm9.992-.874h.33c.004.047.036.08.036.13 0 .146-.17.207-.311.207-.115 0-.179-.056-.275-.113.041-.112.124-.159.22-.224z"
+						fill="#fff"
+					/>
+					<path
+						d="M27.175 11.601c.126.268.095.527.31.76a.64.64 0 00.338.185l-1.096.001c-.178-.016-.262-.186-.45-.186-.12 0-.187.092-.255.185l-1.414.001a.323.323 0 01-.124-.261c0-.337.622-.244.622-.58a.172.172 0 00-.033-.105h2.102zm5.544 0c.011.007.024.01.039.01a.213.213 0 00.065-.01h3.978a.187.187 0 00-.09.16c0 .281.435.108.66.263.086.056.031.163.073.262.044.11.137.165.17.262l-1.802-.001a.201.201 0 00.094-.168c0-.159-.174-.243-.33-.243-.128 0-.187.117-.256.225.077.07.139.143.217.186h-3.252a.492.492 0 01-.04-.037c.22-.051.44-.159.44-.393 0-.168-.152-.234-.294-.318.05-.059.073-.141.11-.198h.218zm6.07 0c-.232.194-.37.46-.705.46-.137 0-.192-.15-.33-.15-.068 0-.104.042-.164.075h-.146a.527.527 0 00-.188-.385h1.533zm-.833.01h.183v.15a1.18 1.18 0 01-.183 0v-.15zm-4.576.113c.206 0 .256.215.44.3.173.08.439.028.439.224 0 .15-.165.225-.312.225-.196 0-.196-.28-.384-.337-.174-.052-.44.019-.44-.169 0-.14.12-.243.257-.243zm-8.054.412c.073 0 .096.065.146.112a.42.42 0 00-.11.113h-.146a.209.209 0 00-.073-.113c.064-.047.1-.112.183-.112zm3.516.187v.114-.114z"
+						fill="#fff"
+					/>
+					<path
+						d="M26.029 10.672a.275.275 0 01-.08.06c-.248.112-.66-.038-.66.243 0 .14.12.243.256.243a.637.637 0 00.55-.262h.402c0 .173-.096.346.037.45.183.14.462-.085.586.112.021.034.04.067.055.101h-2.102c-.107-.147-.443-.143-.443-.363 0-.174.183-.239.183-.412a.43.43 0 00-.033-.172h1.249zm1.347 0v.056c-.142-.014-.224.005-.366.005a.523.523 0 01-.24-.06h.606zm6.004 0c.185.096.34.239.586.284.297.056.48.038.787.038.286 0 .648-.048.675-.322h2.32a.152.152 0 00-.011.06c0 .21.215.336.42.336.211 0 .422-.14.422-.355l-.001-.042h2.344v.135c.077.024.1.113.183.113.11 0 .22-.075.22-.188 0-.025-.01-.041-.019-.059h.439c.013.098.019.203.019.322 0 .206-.129.412-.33.412-.293 0-.526.06-.732-.15-.156-.16-.252-.337-.476-.337-.215 0-.165.318-.33.449-.27.215-.562-.014-.878.112a.845.845 0 00-.23.14h-1.532a.325.325 0 00-.197-.065c-.09 0-.188.02-.257.065h-3.979c.065-.02.11-.065.19-.065h.477c.188 0 .403-.07.403-.262 0-.463-.61-.487-1.062-.487.022-.044.042-.093.067-.135h.482zm-.878.95c.025-.04.058-.066.11-.066.048 0 .07.044.107.065h-.217zm3.898-.89c.147 0 .22.112.311.224-.155.127-.315.098-.512.075v-.187c.064-.061.11-.113.201-.113z"
+						fill="#fff"
+					/>
+					<path
+						d="M19.768 9.743c-.118.08-.264.206-.264.09 0-.033.003-.063.006-.09h.258zm5.116 0a.361.361 0 01-.144.034.487.487 0 01-.185-.034h.329zm1.108 0c.103.134.245.22.431.22h.403c.115 0 .16.113.275.113.154 0 .19-.164.198-.333h.012c.019.177.064.314.064.52-.146-.056-.206-.187-.366-.187-.137 0-.256.103-.256.243 0 .258.375.239.622.206v.165h-.606c-.157-.077-.283-.202-.492-.202-.124 0-.169.118-.249.202H24.78c-.11-.253-.467-.325-.773-.427v-.187h.623c.274 0 .384.262.659.262.274 0 .38-.23.549-.45v-.145h.154zm10.45 0a.736.736 0 01-.134.07v.08c.137.033.224-.004.366-.004.215 0 .44.117.44.337 0 .191-.23.196-.403.262-.124.047-.174.187-.311.187-.179 0-.311-.154-.311-.337 0-.173.114-.258.182-.412-.123-.083-.29-.11-.39-.184H33.04a.582.582 0 01-.063.184c.055-.033.082-.075.146-.075.39 0 .664.033.952.3.091.084.091.262.22.262.288 0 .407-.262.695-.262.256 0 .44.243.44.505l-.002.035h2.32c.039-.094.173-.118.28-.166.138-.06.143-.262.294-.262.188 0 .248.229.255.427h2.344v-.052h.366c.002.021.01.037.018.053h.44c-.035-.269-.12-.484-.275-.765-.03-.058-.04-.137-.072-.184h-.211a.251.251 0 01-.12.034.548.548 0 01-.195-.033h-4.43zm-3.544.95c.03-.053.07-.093.134-.093a.72.72 0 01.347.092h-.481zm6.723-.617c.193 0 .385.121.385.318 0 .192-.234.244-.421.244-.165 0-.348-.075-.348-.244 0-.197.192-.318.384-.318z"
+						fill="#fff"
+					/>
+					<path
+						d="M22.653 8.814v.044c-.302.014-.448.188-.75.188-.158 0-.229-.13-.329-.232H19.67c0 .127-.007.257-.019.419-.016.216-.115.336-.14.528h.257a.53.53 0 01.066-.041c.307-.155.55-.028.879-.113.462-.116.677-.411 1.153-.411.224 0 .32.159.494.3.256.21.508.186.842.186h.513c.45 0 .541-.59.822-.868h-.567c-.017.2-.201.344-.402.344-.223 0-.454-.124-.474-.344h-.441zm3.172 0c-.02.039-.06.068-.06.12 0 .295.067.618.227.827h-.154V9.72c-.307-.028-.48 0-.787 0-.067 0-.116.022-.167.042h-.328c-.084-.035-.146-.097-.146-.192 0-.183.133-.337.311-.337.243 0 .362.187.605.187.105 0 .22-.06.22-.168a.537.537 0 00-.231-.438h.51zm1.48 0v.046a.542.542 0 00-.01-.046h.01zm8.63 0c.027.03.044.07.044.12 0 .173-.179.262-.348.262-.156 0-.348-.047-.348-.206 0-.063.017-.126.047-.176h-1.563c.14.094.235.235.235.419 0 .239-.215.365-.44.412-.196.037-.315-.014-.512-.038a.952.952 0 01-.01.154h2.842a.203.203 0 01-.086-.172c0-.365.412-.506.769-.506.105 0 .11.155.11.262a.483.483 0 01-.233.417h4.43a.638.638 0 01-.28-.23c.215-.285.805-.084.805-.449 0-.047-.027-.07-.037-.112h-.622c-.023-.047-.073-.075-.073-.131 0-.01 0-.019.002-.027h-.36c.118.026.211.068.211.195 0 .192-.196.3-.384.3-.261 0-.343-.244-.568-.375-.106-.061-.242-.08-.347-.119h-1.246a.684.684 0 01.202.27h.732c.115 0 .16-.113.275-.113.124 0 .201.117.201.243 0 .342-.343.543-.677.543-.522 0-.705-.561-1.226-.561-.124 0-.188.074-.312.074-.178 0-.347-.117-.347-.3a.23.23 0 01.052-.157h-.938zm-8.633.947c.002-.052.002-.105.002-.154 0 .056.004.107.01.155h-.012zm13.885.002c.04-.02.08-.043.137-.043.034 0 .057.017.074.042h-.211zM34.33 9.27c.069.01.115.02.183.038a.346.346 0 01-.183.15.118.118 0 00-.073-.075.429.429 0 01.073-.113z"
+						fill="#fff"
+					/>
+					<path
+						d="M22.087 8.232c.179.154.295.413.566.413v.18h.44a.484.484 0 01.475-.517c.242 0 .403.24.403.487 0 .01 0 .02-.002.03h.567a.452.452 0 01.332-.142c.162 0 .323.051.447.143h.51a.103.103 0 00.013-.05c0-.051-.032-.079-.037-.13h-.403c-.105 0-.15-.113-.256-.113h-.988c-.032-.056-.074-.084-.074-.15 0-.057.016-.108.042-.15h-1.181a.34.34 0 01-.252.113.566.566 0 01-.336-.114h-.266zm13.332.001a.248.248 0 01-.051 0h.051zm1.621 0v.15c-.073 0-.11.038-.183.038-.1 0-.146-.085-.22-.15a.414.414 0 00.034-.04l-.84.001.001.001c-.128.103-.27.061-.402.15-.124.084-.142.262-.293.262-.238 0-.348-.187-.586-.187h-.293c-.105 0-.155.066-.256.075-.28.028-.449-.005-.732-.005v.155c.177 0 .359.048.497.142h1.563a.2.2 0 01.173-.105c.152 0 .346.011.432.105h.939c.135-.156.503-.105.77-.105.13 0 .24.038.329.106h1.246c-.095-.035-.165-.086-.165-.2 0-.13.165-.168.293-.168.279 0 .398.201.659.3.092.035.204.046.3.067h.36c.04-.281.731-.034.731-.348 0-.108-.075-.17-.129-.245h-1.72a.742.742 0 01-.238.039.41.41 0 01-.176-.038h-.526c-.046.188-.342.188-.56.188-.192 0-.443-.005-.458-.189h-.55zm-17.37.592a7.302 7.302 0 00-.02-.517h.513c.063 0 .119-.03.158-.076h.5a.322.322 0 00-.109.245c0 .29.426.168.696.243.069.02.12.06.166.105h-1.905zm7.622 0c-.021-.08-.057-.142-.1-.217.069-.07.12-.112.183-.187-.05.153-.07.253-.072.404h-.01zm-.667-.292c.091 0 .15.01.238.038-.1.112-.197.187-.348.187-.05 0-.078-.033-.128-.038 0-.037-.037-.056-.037-.093 0-.117.16-.094.275-.094z"
+						fill="#fff"
+					/>
+					<path
+						d="M22.889 7.652c.1.067.167.162.167.293a.443.443 0 01-.115.299h1.181c.075-.122.24-.186.398-.186h.952c.036-.148 0-.25 0-.406h-.82c-.107.17-.328.219-.553.219-.179 0-.4-.047-.42-.218h-.79zm12.522 0a.13.13 0 01.02.069c0 .089-.079.15-.166.15a.166.166 0 01-.164-.169c0-.018.004-.034.01-.05h-.48c-.123.231-.59.138-.922.219.298.182.531.187.879.187h.476c.137 0 .178.164.303.185h.05c.11-.016.15-.148.27-.148.08 0 .1.083.144.148h.84c.054-.07.095-.148.187-.148.082 0 .114.075.183.113v.036h.55a.21.21 0 01-.001-.018c0-.248.307-.318.55-.318.214 0 .475.06.475.28 0 .02-.002.039-.006.055h.525c-.085-.037-.157-.098-.263-.148.06-.239.353-.173.586-.224.196-.046.297-.189.478-.22l-2.236.001c-.24.115-.477.219-.805.219a.582.582 0 01-.467-.219H35.41zm5.038 0c.027.048.058.08.106.106-.06.244-.306.304-.549.337-.186.026-.31.102-.459.15l1.721-.001a.244.244 0 01-.054-.149v-.443h-.765zm-20.127.592a.261.261 0 00.062-.167c0-.283-.506-.212-.606-.425h.54c.044.19.294.256.505.256.316 0 .49-.112.805-.112.332 0 .493.283.725.447h-.265a.41.41 0 00-.276-.11c-.07 0-.097.075-.165.075-.156 0-.229-.113-.385-.113-.16 0-.332.048-.44.148h-.5zm15.73-.486h.22v.225a.204.204 0 00-.11.075c-.068-.07-.119-.113-.183-.187.028-.038.055-.07.073-.113zm-2.928.075c.141 0 .224.023.366.038-.11.065-.183.084-.293.15a.55.55 0 00-.073-.188zm-6.883.225c-.005.047-.018.071-.037.114h-.11a.492.492 0 01.147-.114z"
+						fill="#fff"
+					/>
+					<path
+						d="M21.058 6.955c-.127.096-.257.131-.418.218.11.154.27.15.457.15.142 0 .206-.113.348-.113h.55c.109.14.187.234.219.412.073-.033.119-.075.201-.075.165 0 .346.036.474.12h.79a.212.212 0 010-.026c0-.098.082-.14.146-.206h.402c.188 0 .298.051.476.112a.412.412 0 01-.05.119h.82V7.66c0-.313.178-.597.443-.704H25c.02.014.043.024.07.03v.262c-.046 0-.065.037-.11.037-.106 0-.156-.098-.257-.112a2.659 2.659 0 00-.695-.005v-.182a.986.986 0 01.035-.031h-.523a.922.922 0 01-.263.03.472.472 0 01-.166-.03h-2.033zm3.651.001a.484.484 0 00.11 0h-.11zm2.364 0c.1.187.456.155.705.254-.361.145-.526.412-.915.412-.15 0-.33-.051-.33-.206 0-.084.074-.121.074-.206 0-.128-.087-.21-.202-.254h.668zm13.122 0c.057.066.104.122.104.217 0 .196-.21.239-.293.412-.293-.07-.732.047-.732-.262 0-.18.126-.332.293-.367h-6.053a.409.409 0 01-.024.03c.046.08.069.187.165.187.146 0 .22-.112.366-.112.306 0 .64.154.64.468 0 .055-.01.1-.03.137h.48a.153.153 0 01.136-.081c.066 0 .133.03.164.081h1.016a.417.417 0 01-.082-.25c0-.3.33-.43.623-.43.265 0 .412.112.677.112.146 0 .22-.075.366-.075.201 0 .316.117.458.262v.113c-.311.043-.541.16-.77.269h2.237a.376.376 0 01.07-.007h.44a.195.195 0 00.003.006h.765V7.51c0-.098.11-.126.11-.225 0-.107-.11-.154-.11-.262v-.068h-1.019zm-20.417.71a.191.191 0 01-.017-.081c0-.08.074-.108.074-.187V7.06c0-.035 0-.07.002-.107l.474.001c.008.124.22.127.22.255 0 .174-.22.22-.22.394 0 .022.002.043.007.063h-.54zm15.853-.643c.16 0 .385.005.385.169 0 .168-.183.243-.348.243-.16 0-.348-.06-.348-.225 0-.14.174-.187.311-.187zm-12.374.187c.124 0 .151.136.238.225v.15c-.073.019-.123.037-.201.037-.128 0-.275-.056-.275-.187 0-.126.115-.225.238-.225zm4.595 0h.146a.633.633 0 01-.11.114c-.018-.042-.032-.066-.036-.114z"
+						fill="#fff"
+					/>
+					<path
+						d="M21.901 6.375l.02.023h.403c.155 0 .21.15.366.15.159 0 .327-.06.414-.174h-1.203zm1.721 0c-.03.009-.06.016-.09.023v.112c.068.099.146.155.146.281 0 .097-.068.149-.158.177l.523-.001a.3.3 0 01.203-.082c.179 0 .32.067.463.082h.106a.414.414 0 00.145-.045.174.174 0 00.04.045h.916a.628.628 0 01.233-.045c.086 0 .177.014.256.045h.668a.248.248 0 01-.027-.12c0-.135.147-.182.147-.318a.19.19 0 00-.08-.155h-.606c.017.033.027.071.027.118 0 .126-.179.098-.293.13-.37.113-.563.3-.952.3-.339 0-.485-.23-.805-.337a.686.686 0 01.108-.21h-.97zm10.076 0a.296.296 0 01-.117.024.227.227 0 01-.105-.024h.222zm1.182 0a.414.414 0 01.037.06c.097.188.101.45.312.45.239 0 .317-.252.34-.51h.822c-.022.03-.046.054-.046.098 0 .365.49.29.806.45.137-.24.549-.17.549-.45a.286.286 0 00-.017-.098h.697c.173.24.697.173 1.077.173.268 0 .462-.03.665-.174h1.092v.593h-1.019a.179.179 0 00-.006-.007h-.549a.35.35 0 00-.073.007h-6.053c.078-.113.094-.27.232-.27.064 0 .082.076.146.076.124 0 .22-.118.22-.244a.317.317 0 00-.04-.154h.808zm-15.043.592c.012-.233.072-.466.29-.494.282-.04.688.074.808-.098h.416c.054.1.228.102.35.173.202.112.298.3.531.3.147 0 .22-.075.366-.075.208 0 .327.133.493.194h-2.033a.735.735 0 00.094-.082.466.466 0 00-.42-.225c-.202 0-.422.094-.422.3v.007h-.473zm18.43-.457c.091 0 .165.075.165.169a.166.166 0 01-.165.168c-.1 0-.201-.065-.201-.168s.1-.169.201-.169zm2.178.15h.257c0 .047.036.066.036.113 0 .117-.142.15-.256.15-.078 0-.115-.043-.183-.076a.933.933 0 01.146-.187z"
+						fill="#fff"
+					/>
+					<path
+						d="M22.221 5.794a.794.794 0 01-.044.054h-.659a.767.767 0 01-.044-.054h-1.566a.86.86 0 000 .035c0 .094.072.169.164.169.15 0 .233-.075.384-.075.238 0 .513.093.513.337a.217.217 0 01-.035.127h.417a.145.145 0 01-.016-.071c0-.108.114-.169.22-.169.163 0 .242.116.346.24h1.203a.295.295 0 00.061-.183c0-.14-.073-.216-.073-.356 0-.02.002-.037.005-.054h-.876zm1.651 0c.06.062.099.145.099.241 0 .193-.166.296-.349.351h.97c.213-.275.612-.313 1.008-.313.11 0 .215.042.238.15.248 0 .574-.014.669.163h.607c-.136-.109-.397-.134-.397-.332 0-.103.064-.206.164-.206.48 0 .746.187 1.227.187.14 0 .189-.124.255-.242h-4.491zm9.959 0a.516.516 0 00-.012.11c0 .131.023.23.11.318-.087.06-.15.128-.23.164h-.223c-.088-.044-.144-.138-.243-.201-.276-.179-.473-.247-.739-.391h1.337zm3.377 0a.34.34 0 01-.186.054.726.726 0 01-.28-.055h-1.584c.048.058.099.116.162.167.096.08.256.094.256.225 0 .064-.002.133-.008.201h.823a.167.167 0 00.027-.051h.66c.155 0 .228-.113.384-.113.103 0 .186.07.22.165h.698a.336.336 0 01-.058-.202c0-.173.16-.3.33-.3.288 0 .411.216.695.262.33.057.705-.121.842.188a1.225 1.225 0 01-.068.051h1.093v-.388c-.558 0-.874.056-1.428 0-.149-.015-.303-.083-.37-.205h-2.208zm-3.137.592c-.02-.04-.046-.078-.07-.126.074-.014.124-.038.202-.038.268 0 .532-.031.677.164h-.809z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.676 5.097v.016c.069.037.146.065.146.15 0 .154-.141.262-.292.262-.183 0-.293-.188-.293-.375 0-.022.005-.039.015-.053h-.245a.41.41 0 01.01.09c0 .246-.1.385-.11.62h1.567c-.167-.22-.162-.457-.29-.71h-.508zm3.526 0c.038.045.062.1.062.166 0 .243-.339.177-.513.337a.367.367 0 00-.22-.075c.014.056.028.093.037.15a.452.452 0 01.304.132h4.49a.415.415 0 01.149-.17c.247-.145.453-.047.732-.112 0-.277.297-.412.567-.412.526 0 .773.374 1.3.374.082 0 .119-.075.201-.075.078 0 .105.075.183.075.197 0 .325.024.494-.075.108-.061.077-.189.082-.315h-4.607c.105.03.205.075.205.184 0 .178-.233.174-.402.206-.348.066-.522.262-.879.262-.284 0-.494-.271-.476-.561a4.061 4.061 0 01-.44-.09h-1.269zm11.811 0a.285.285 0 01.039.147c0 .178-.21.243-.384.243-.202 0-.421-.112-.421-.318 0-.025.003-.049.01-.071h-1.454a.24.24 0 01.016.09c0 .21-.215.337-.421.337-.083 0-.12-.075-.201-.075-.156 0-.216.122-.367.15-.173.032-.347-.005-.439.15l.103.058h1.337c.029-.133.11-.246.244-.246h.696c.184 0 .278.118.387.246h1.584c-.122-.053-.214-.146-.214-.283 0-.183.206-.211.33-.337.029-.03.056-.06.081-.09h-.926zm2.77 0c.01.029.015.059.015.09 0 .104-.083.15-.147.225h-.476c-.141 0-.206-.112-.347-.112-.308 0-.396.367-.62.508h2.208a.279.279 0 01-.033-.133c0-.155.147-.235.293-.263.563-.112.925.192 1.465.005v-.192a.865.865 0 01-.028-.128h-2.33zm-15.686.71c.017-.085.068-.147.068-.245 0-.206-.256-.262-.457-.262-.275 0-.323.287-.487.508h.876z"
+						fill="#fff"
+					/>
+					<path
+						d="M21.501 4.516c-.13.291-.48.458-.825.458v.135h.509l-.033-.06c.156-.159.27-.3.494-.3.266 0 .421-.018.677.038.39.085.586.3.989.3.22 0 .33-.112.549-.112a.46.46 0 01.341.134h1.27c-.262-.066-.493-.134-.824-.134-.123 0-.188.074-.311.074-.133 0-.197-.102-.293-.187.069-.134.148-.25.24-.346h-.982c-.045.146-.242.196-.411.196-.21 0-.364-.091-.51-.195h-.88zm4.137 0c.194.149.368.27.383.533.105.01.17.038.275.038.169 0 .233-.187.402-.187.275 0 .417.117.677.187.029.007.058.014.088.023l4.607-.001a.348.348 0 01.028-.134c.12-.253.366-.3.64-.3.18 0 .257.122.422.187.23.09.566.058.643.247h1.454c.036-.125.165-.21.3-.21.18 0 .375.07.456.21h.926c.14-.165.243-.322.467-.322.083 0 .133.052.184.113h.878c.138 0 .27.084.314.209h2.33c-.02-.126-.035-.247-.154-.247-.165 0-.202.225-.366.225-.193 0-.211-.3-.403-.3-.142 0-.206.113-.348.113-.288 0-.471-.136-.64-.375l.006-.009h-.73a.27.27 0 01-.39 0H35.52a.313.313 0 01-.073.01.906.906 0 01-.13-.01h-.806a1.136 1.136 0 01-.309.047.589.589 0 01-.255-.047h-.74a.189.189 0 01-.158.084c-.064 0-.105-.037-.142-.084h-3.775v.047c.096.028.22.06.22.168 0 .131-.184.131-.312.131-.206 0-.297-.145-.494-.187-.165-.038-.343.103-.44-.038-.023-.036-.023-.078-.017-.121h-.798a.621.621 0 01-.32.084c-.148 0-.345.012-.414-.083h-.92zm15.1 0a.66.66 0 01-.33.086c-.061 0-.095-.038-.124-.086h.454zm-20.73.593c-.033-.148-.137-.242-.137-.415 0-.075.036-.128.076-.178h.456a.364.364 0 00-.02.121c0 .15.046.23.074.375-.073.03-.168.042-.204.096h-.246zm15.861-.584c.202 0 .312.01.513 0v.187c-.215.014-.334.01-.55 0-.004-.046-.018-.07-.036-.112a.284.284 0 00.073-.075zm-5.436.15c.078 0 .105.07.165.112a.396.396 0 01-.11.15h-.22c0-.033-.036-.042-.036-.075 0-.107.096-.187.2-.187zm3.972.225h.183a.12.12 0 00.037.076c-.101.01-.16.01-.257 0a.288.288 0 00.037-.076z"
+						fill="#fff"
+					/>
+					<path
+						d="M22.73 3.936l.106.076c.197.15.476.192.476.45a.22.22 0 01-.01.066h.982a.92.92 0 01.693-.29c.174 0 .252.116.421.149-.036-.061-.077-.098-.146-.112 0-.047-.037-.066-.037-.113 0-.15.165-.225.312-.225.123 0 .173.085.274.15V4.2a.869.869 0 00-.293.224c.044.038.087.072.13.105l.92-.001a.141.141 0 01-.025-.085c0-.178.033-.384.129-.508h-1.648a.78.78 0 01-.348.077H24.3c-.26 0-.361.263-.622.263-.206 0-.28-.175-.383-.34h-.565zm4.378 0c.056.053.112.114.194.151h.11a.691.691 0 01.055-.152h-.359zm3.452 0v.02c0 .36-.434.468-.787.468-.169 0-.251-.112-.42-.112-.097 0-.138.066-.22.112v.104h3.775c-.026-.033-.05-.07-.078-.104.069-.07.12-.15.22-.15.096 0 .183.07.183.17 0 .03-.009.06-.024.084h.739c-.055-.03-.093-.073-.093-.141 0-.122.1-.225.22-.225h.512c.17 0 .261-.11.367-.227h-1.322a.431.431 0 01-.18.04.33.33 0 01-.157-.04h-1.23a.413.413 0 01-.242.077.295.295 0 01-.199-.077H30.56zm7.213 0v.001c0 .155-.165.183-.293.263-.078-.117-.151-.225-.293-.225-.32 0-.467.262-.787.262-.129 0-.202-.075-.33-.075-.253 0-.33.308-.549.366h2.566a.205.205 0 01-.058-.141c0-.112.128-.15.238-.15.137 0 .215.089.31.187-.035.037-.067.073-.101.104h.731c.12-.153.198-.328.396-.328.22 0 .361.037.55.15.066.04.094.117.13.178h.455c.107-.062.183-.16.183-.29 0-.066-.041-.1-.073-.15h-.842c-.092 0-.128-.076-.22-.076h-.293c-.11 0-.165.075-.274.075-.22 0-.46-.043-.632-.15l-.815-.001zm-17.827.592c.05-.063.107-.122.107-.216v-.376h.78c-.005.082.014.173.025.264-.2.054-.395.15-.456.327l-.456.001zm2.434 0c-.19-.135-.368-.29-.643-.29-.073 0-.114.023-.183.037 0 .092-.02.177-.054.253h.88zm5.71 0c.011-.08.043-.161-.02-.216-.11-.094-.233-.06-.366-.112h-.183a.47.47 0 01-.23.329l.799-.001zm7.228.001c-.142-.02-.252-.067-.42-.067-.154 0-.268.036-.387.068h.807z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.693 3.239c.211.124.426.027.679.113.201.066.297.225.512.225h.403c.105 0 .17-.061.256-.112.165.168.353.14.55.262.09.055.15.138.203.222h-.566c-.321-.214-.612-.26-1.029-.26-.302 0-.521-.098-.769.075a.23.23 0 00-.097.186h-.781v-.148c.224-.043.549.009.549-.225 0-.183-.265-.183-.44-.15 0-.033-.036-.042-.036-.075 0-.046.02-.08.043-.114h.523zm6.598 0a.557.557 0 00.341.113c.43 0 .677-.004 1.098-.075a.514.514 0 00.13-.039h3.058l-.003.002c-.183.14-.402.21-.402.45a.35.35 0 00.111.26H30.56c-.007-.124-.076-.196-.146-.298h-.512c-.201 0-.293.15-.494.15-.357 0-.54-.188-.897-.188h-.476a.613.613 0 00-.567.335h-.36c-.062-.06-.125-.11-.226-.11a.267.267 0 00-.22.11h-1.647c.117-.06.2-.157.2-.297 0-.224-.189-.307-.377-.413h2.453zm6.466 0a.435.435 0 01-.124.711h1.322c.034-.038.069-.076.109-.11.357-.32.6-.6 1.08-.6.21 0 .265.262.476.262.265 0 .375-.225.64-.225.156 0 .165.192.22.337.06.159.29.164.293.335h.815a.445.445 0 01-.23-.39c0-.132.078-.193.11-.32a.206.206 0 00-.008-.001h-4.703zm7.055 0c0 .263-.348.263-.605.263-.2 0-.398.028-.494-.15-.019-.035-.032-.075-.045-.112l1.144-.001zm-7.517.71c-.218-.115-.305-.447-.592-.447-.303 0-.423.303-.638.448h1.23zm-9.946-.71c.201 0 .334.024.512.113-.27.024-.43.033-.695 0 .064-.042.1-.112.183-.112zm3.02 0c.133 0 .311.015.311.15 0 .174-.16.3-.33.3-.136 0-.292-.065-.292-.206 0-.154.16-.243.311-.243zm-.897.226h.22a.12.12 0 00.036.076.61.61 0 01-.292 0 .29.29 0 00.036-.076z"
+						fill="#fff"
+					/>
+					<path
+						d="M24.191 2.658c.004.028.036.038.036.07 0 .102-.119.149-.22.149-.123 0-.169-.112-.292-.112h-.403a.605.605 0 00-.11.224c-.123-.065-.206-.112-.348-.112-.21 0-.302.187-.512.187-.137 0-.206-.107-.275-.225v-.15l.087-.03h2.037zm1.988 0a.21.21 0 00-.012.07c0 .172.16.299.33.299.16 0 .229-.113.366-.188h.183c.032.171.12.318.245.412h-2.453a2.924 2.924 0 01-.062-.037.285.285 0 00.037-.075h.403c.33 0 .805.075.805-.262 0-.142-.157-.175-.297-.218h.455zm4.412 0c.096.05.187.107.317.107.092 0 .129-.075.22-.075.16 0 .275.154.275.318 0 .117-.106.206-.22.206-.128 0-.183-.112-.311-.112-.224 0-.343.112-.568.112-.201 0-.439-.075-.439-.28a.32.32 0 01.16-.275h-.275a.245.245 0 01.005.05c0 .163-.215.163-.366.205-.232.063-.334.248-.528.336h3.057c.12-.094.034-.266.107-.41a.458.458 0 01.152-.181H30.59zm2.068 0c.173.163.23.519.501.519h.403c.075 0 .141.028.194.074h4.703a1.61 1.61 0 00-.211-.037c-.17.042-.243-.187-.421-.187-.106 0-.151.075-.257.075-.105 0-.2-.08-.2-.188 0-.133.078-.209.19-.255h-1.183c-.093.057-.213.068-.344.068a.516.516 0 01-.263-.07l-.87.001c.03.048.053.097.053.163 0 .28-.348.356-.622.356-.417 0-.678-.127-.989-.412a.388.388 0 01.083-.108l-.767.001zm7.103 0a.193.193 0 00-.012.07c0 .172.16.299.33.299h.585c.06.074.146.121.146.223h-1.143c-.027-.078-.055-.148-.138-.148-.16 0-.252.023-.403.075a.273.273 0 00-.073-.038c.064-.08.11-.154.11-.262a.408.408 0 00-.064-.22l.662.001zM20.17 3.25c.031-.045.067-.087.067-.148 0-.117-.087-.169-.146-.263v-.112c.032 0 .041-.037.073-.037.147 0 .22.06.366.075 0 .177-.027.341.11.45a.512.512 0 00.053.035h-.523zm15.388-.448c.065 0 .087.066.129.112v.15c-.046.01-.06.075-.11.075a.186.186 0 01-.183-.187c0-.09.077-.15.164-.15z"
+						fill="#fff"
+					/>
+					<path
+						d="M21.589 2.077c-.201.19-.38.362-.693.362-.302 0-.421-.206-.696-.337a.52.52 0 00.03-.025h1.359zm3.18.001a6.534 6.534 0 00-.286.025.861.861 0 01-.008-.024h.294zm5.9 0c-.089.067-.162.136-.291.136h-1.392c-.258 0-.445-.058-.637-.136h-1.238c-.073.078-.179.139-.248.249.307 0 .462.187.769.187.187 0 .27-.15.457-.15.316 0 .472.187.787.187.17 0 .252-.112.421-.112.19 0 .416.06.453.231h.275a.57.57 0 01.298-.081c.107 0 .189.038.268.08h1.586a.503.503 0 01.287-.08c.081 0 .143.032.195.08h.767a.178.178 0 01.1-.043c.508-.06.805-.023 1.318-.037.02.029.04.055.057.081h.87c-.075-.042-.146-.098-.232-.156.138-.196.379-.264.473-.436h-.63c-.093.092-.248.136-.392.136-.186 0-.378-.046-.528-.137h-.454c.085.043.14.11.14.231 0 .16-.21.169-.366.169-.412 0-.61-.329-.976-.4h-.242c-.305.081-.406.512-.759.512-.137 0-.293-.066-.293-.206 0-.143.158-.192.206-.306h-1.05zm7.169 0c-.113.115-.413.054-.614.136-.101.043-.129.15-.238.15-.124 0-.188-.075-.312-.075s-.123.16-.183.262a.335.335 0 01-.113.119h1.183c.124-.051.29-.066.431-.081.394-.047.627-.024 1.025-.038.032.041.06.079.083.118h.662c.036-.093.14-.155.244-.155h.622c.087 0 .183-.042.183-.131 0-.119-.01-.213-.031-.306h-.246c-.229.316-.533.512-.95.512-.205 0-.233-.257-.42-.337-.159-.069-.377-.07-.469-.174h-.857zm-15.684.592c.291-.11.48-.23.81-.23.334 0 .49.224.824.224h.402v.005h-2.036zm3.57 0c-.113-.035-.216-.078-.216-.193 0-.113.11-.188.22-.188.26 0 .43.028.659.15-.075.085-.176.135-.209.23h-.454z"
+						fill="#fff"
+					/>
+					<path
+						d="M20.23 2.099c.065-.06.112-.125.209-.125s.141.075.238.075h.402c.032-.056.073-.085.073-.15 0-.314-.416-.253-.695-.375-.188-.08-.229-.252-.293-.45.151-.079.265-.074.44-.074.21 0 .32.075.53.075.234 0 .37-.075.604-.075.645 0 1.002.15 1.648.15 1.871.01 2.928.014 4.796-.075.086 0 .128-.075.22-.075h9.371c.018.056.032.089.037.15.032 0 .041.037.073.037.092 0 .105-.107.146-.187h2.71v.225c-.138.112-.284.065-.44.15-.155.084-.15.337-.33.337-.228 0-.301-.263-.53-.263-.256 0-.357.192-.568.338-.086.06-.22.074-.22.187a.18.18 0 00.045.125h-.858a.168.168 0 00.045-.125c0-.187-.11-.375-.293-.375-.252 0-.37.188-.622.188-.28 0-.38-.3-.66-.3-.155 0-.329.084-.329.243 0 .085.073.122.073.206a.33.33 0 01-.04.163h-.63a.272.272 0 00.085-.2c0-.398-.49-.487-.879-.487-.178 0-.403.042-.403.225a.53.53 0 00.278.462h-.455c-.183-.093-.504-.071-.701-.238-.105-.088-.041-.257-.147-.337-.242-.192-.48-.112-.787-.112-.316 0-.75-.023-.75.3 0 .13.11.187.11.318a.176.176 0 01-.014.069h-1.05a.616.616 0 01.148-.088c.128-.047.293-.065.293-.206 0-.234-.325-.164-.549-.206-.229-.042-.334-.187-.567-.187-.275 0-.421.15-.696.15-.352 0-.545-.15-.897-.15-.366 0-.879-.037-.879.337 0 .14.202.108.33.15.194.066.347.14.498.2h-1.239a.281.281 0 00.082-.2c0-.234-.265-.337-.494-.337-.096 0-.142.075-.238.075h-.403c-.274 0-.426.206-.549.45-.292 0-.496-.003-.74.012h-.293c-.076-.225-.106-.5-.34-.5-.206 0-.33.005-.53.038-.101.014-.17.06-.257.112h-.33l-.073-.075c.028-.056.074-.084.074-.15 0-.093-.092-.15-.184-.15-.16 0-.045.295-.183.375-.242.14-.494-.08-.732.075a2.34 2.34 0 00-.332.274H20.23zm12.576 0a.776.776 0 00-.14-.012.393.393 0 00-.102.013h.242zm7.728 0c.061-.084.117-.176.168-.275.033.104.06.189.078.275h-.246z"
+						fill="#fff"
+					/>
+					<path
+						d="M30.135 6.22c1.487 0 3.13.693 3.13 2.185 0 .413-.183 1.377-.183 1.377-.041.31-.073.485-.146.79-.09.371-.268.898-.293.936-.137.438-1.282 1.45-1.282 1.872 0 .551.586.845 1.135.845h.366c.49 0 1.771.656 2.16.77.98.295 1.543.552 2.563.552.179 0 .147-.262.147-.441v-.771c0-.33.695-1.198.951-1.469.453-.472.806-.674 1.428-.88.197-.065.302-.185.513-.185h.549c.357 0 .503.257.805.441.66.404 1.02.716 1.501 1.322.44.55.504 1.358 1.208 1.358.234 0 .357-.105.586-.147.352-.064.613.046.915-.146.33-.211.43-.588.824-.588.188 0 .275.216.275.404v.477c0 1.125-1.094 1.556-2.16 1.91-.77.257-1.19.486-1.977.66-.375.083-.879.088-.879.478v.514c.627.418 1.13.477 1.886.477 1.006-.082 1.578-.114 2.58-.22 1.277-.138 2.046-.33 3.149-.991.412-.248.879-.381.879-.863 0-.822-.39-1.35-1.062-1.818-.558-.39-1.501-.215-1.501-.9 0-.22.014-.353.037-.568.141-1.143.21-1.795.512-2.901.092-.33.055-.574.256-.845.188-.247.463-.188.77-.257.48-.11.75-.22 1.244-.22 1.016 0 1.583.317 2.49.771.75.376 1.464.753 1.464 1.597 0 .478-.193.73-.403 1.157-.499 1-1.025 1.56-1.025 2.68v.808c0 .711.512.991.842 1.616.224.426.445 1.173.476 1.23.348.628.659 1.423.769 1.634.137.257.444 1.556.586 2.092.27 1.024.545 1.589.658 2.644 0 0 .147 1.024.147 1.359v.863c0 .472-.019.865-.128 1.325l-.092.382c-.19.8-.924 4.13-1.025 4.736-.15.91-.356 1.404-.512 2.314-.318 1.85-.508 2.836-.952 4.626-.449 1.804-.59 2.84-.952 4.663a44.312 44.312 0 01-.513 2.313c-.141.583-.169.927-.329 1.506-.073.27-.192.55-.476.55-.078 0-.119-.036-.183-.073v-4.48c0-.904-.037-1.413-.11-2.313-.032-.399-.11-.624-.11-1.028v-6.609c0-.303.11-1.322.11-1.322h-.22c-.622.39-.553 1.038-.695 1.763-.256 1.294-.44 2.033-.55 3.341-.15 1.772-.086 2.79-.329 4.553-.251 1.836-.585 2.85-.952 4.663-.517-.239-.832-.606-.878-1.175-.096-1.157-.196-3.553-.22-3.929l-.11-1.689s-.08-1.262-.11-1.652c0 0-.082-1.088-.11-1.432l-.219-2.717v-2.02c0-.807.33-4.185.33-4.185.03-.379.146-1.684.146-1.763 0-.463.15-2.331.293-3.047.027-.142.11-1.065.11-1.065.028-.274-.339-.29-.586-.404-.504-.234-.838-.27-1.355-.477a9.155 9.155 0 00-1.244-.404c-.307-.078-.467-.257-.788-.257-.173 0-.2.23-.2.404 0 .124.255 2.152.438 3.047 0 0 .218 1.092.293 1.432 0 0 .293 1.143.293 1.322 0 .064.32 1.946.366 2.35.133 1.138.11 1.795.22 2.938.027.284.087.446.11.734l.11 1.395c.091 1.07.114 1.685.183 2.754-.453.179-.76.142-1.245.184-.71.064-1.121.04-1.83.11-.188.018-.376-.014-.477.147-.22.348-.096.651-.146 1.064a39.35 39.35 0 00-.22 2.13s-.105 1.11-.11 1.469c-.036 2.033-.004 3.18-.004 5.213-.417 0-.622-.21-1.02-.33-.522-.156-.86-.073-1.41-.073-.252 0-.229-.34-.275-.588-.073-.385-.732-3.02-.732-3.103 0-.091-.36-2.767-.366-2.845 0 0-.11-1.35-.11-1.432 0-.445-.014-.698.037-1.138.009-.088-.037-.881-.037-1.304 0-.294-.357-.307-.622-.422a20.35 20.35 0 01-1.83-.881c-.38-.202-.806-.335-.806-.771 0-.74.513-2.607.513-2.607l.659-3.048s.306-1.85.439-2.35c.128-.486.366-1.946.366-1.946.054-.288.293-3.47.293-3.561V21.31c0-.293-.37-.349-.66-.367-.622-.041-.983-.023-1.61-.037-.046.078-.073.726-.073.937 0 .092-.073.642-.073.936 0 .106-.11 1.248-.11 1.248-.076.86-.175 1.185-.257 2.167-.171 2.058.05 3.25-.256 5.287-.009.087-.114 1.414-.183 2.166-.096 1.042-.201 1.625-.33 2.662.019.042-.073.753-.073.753-.01.105-.146 1.359-.146 1.359-.058.54-.22 1.5-.22 1.597 0 .266-.732 4.686-.732 4.755 0 .463-.769 2.97-1.318 4.002h-.44c-.15-.29-.104-.514-.109-.844-.078-5.297.023-8.294-.11-13.585 0-.097-.068-.161-.146-.22-.028.605-.165.94-.22 1.541a46.595 46.595 0 00-.183 2.644c-.17 3.933-.183 6.164-.366 10.097a.924.924 0 01-.11.44c-.42-.224-.503-.6-.622-1.064-.224-.845-.412-1.318-.623-2.167-.315-1.266-.485-1.978-.695-3.268-.302-1.817-1.062-5.842-1.062-6.609v-8.15c0-1.933-.014-3.03.11-4.958l.073-1.028c-.055-.046-.128-.326-.146-.367-.838 0-1.304-.147-2.142-.147-.119 0-.165.138-.165.257.21 1.4.385 2.194.55 3.598.173 1.469.33 2.288.33 3.782v.661l.21 2.644v3.745c0 1.262-.092 1.978-.21 3.231-.079.799-.317 1.216-.55 1.983-.398 1.308-1.004 3.803-1.135 4.039-.405.73-1.043 3.107-1.172 3.892-.036.23-.022.55-.256.55-.613 0-.805-.743-.805-1.358.11-1.097.142-1.735.293-2.827.009-.087.1-5.191.036-8.372-.009-.537-.21-.872-.585-1.248-.133.082-.22.174-.22.33v1.322c0 .344-.11 1.359-.11 1.359-.111 1.376-.348 2.148-.44 3.525-.164 2.469-.136 3.873-.219 6.352-.041.009-.064.036-.11.036-1.345 0-1.345-1.735-1.647-3.047-.179-.785-.16-1.258-.293-2.056-.096-.583-.125-.921-.22-1.506 0 0-.242-1.822-.439-2.717-.398-1.804-.462-2.86-.879-4.663-.302-1.317-.491-2.072-.659-3.415 0 0-.068-.5-.073-.587-.082-1.441.01-2.268-.073-3.709 0-.046-.027-.67 0-1.028-.247-.197-.471-.174-.769-.294h-.22l-.279 10.207v5.288c0 .81-.077 1.257-.16 2.056-.155 1.45-.311 2.267-.549 3.708-.183 1.106-.188 1.822-.805 2.754-.188-.115-.128-.335-.183-.55-.124-.473-.151-.763-.183-1.249-.11-1.63-.101-2.556-.22-4.186-.133-1.84-.366-2.891-.989-4.626-.357-1-.402-1.69-1.098-2.497-.393.395-.242.85-.366 1.395-.215.937-.27 1.483-.476 2.424-.192.885-.146 1.418-.293 2.313-.096.596-.22 1.91-.22 1.91-.03.26-.11 1.1-.11 1.1a28.863 28.863 0 01-.438 2.975c-.179.927-.142 1.638-.843 2.276-.082-1.629-.087-2.56-.22-4.186-.16-1.95-.356-3.047-.548-4.993-.142-1.441-.225-2.263-.33-3.709-.064-.9-.07-1.41-.11-2.313 0 0-.08-2.13-.11-2.46l-.11-1.248s-.075-1.021-.11-1.322c0 0-.17-1.347-.219-1.91 0 0-.131-1.868-.22-2.57 0 0-.219-1.202-.219-1.744 0-.151-.147-3.63-.147-3.726 0-.63.197-.983.476-1.543.151-.307 1.062-1.858 1.575-2.46.393-.463.878-.693.878-1.303 0-.11-.073-.606-.073-.606-.142-1.171-1.172-1.611-1.172-2.79v-.441c0-.473.344-.68.696-.991.837-.74 1.446-1.286 2.563-1.286 1.016 0 2.196.376 2.196 1.396v3.01c0 .35.11 1.286-.073 1.396-.252.151-.457.11-.732.22-.53.211-.879.601-.879 1.175 0 .372.165.574.366.881.504.758.897 1.12 1.538 1.763.444.445.764.66 1.318.954.49.262.76.551 1.318.551h.366c.307 0 .302-.385.44-.66.137-.276.402-.386.402-.698 0-.368-.512-.303-.879-.33-.631-.051-1.464-.065-1.464-.698 0-.193.146-.285.146-.478.353-.032.563-.004.916-.037.85-.082 1.313-.266 2.16-.403.92-.152 1.867-.57 1.867-1.506 0-.546-.33-.794-.623-1.248-.196-.308-.438-.43-.512-.79-.052-.252-.008-.406.036-.66.14-.805.503-1.225 1.099-1.781.436-.408.726-.644 1.3-.808.453-.13.874-.074 1.208-.074.448 0 .709-.013 1.153.037.833.101 1.441.707 1.5 1.542.037.542-.027.854.037 1.395.005.051-.073.446-.073.661v.478c0 .6-.444 1.138-1.043 1.138-.234 0-.371-.073-.604-.073-.151 0-.257.142-.257.293 0 .198.06.441.257.441.782 0 1.194-.367 1.977-.367h.622c.265 0 .403-.115.659-.184.718-.197 1.149-.294 1.83-.587.394-.17.6-.308.952-.551.188-.129.366-.211.403-.44.114-.67.11-1.066.11-1.745 0-.555-.55-.679-.879-1.12-.49-.652-.659-1.18-.659-2 0-.556 0-.905.22-1.414.128-.299.865-1.3 1.281-1.506.444-.22.764-.22 1.263-.22z"
+						stroke="#CC1E1C"
+						strokeWidth=".837"
+					/>
+					<path
+						d="M30.135 6.22c1.487 0 3.13.693 3.13 2.185 0 .413-.183 1.377-.183 1.377-.041.31-.073.485-.146.79-.09.371-.268.898-.293.936-.137.438-1.282 1.45-1.282 1.872 0 .551.586.845 1.135.845h.366c.49 0 1.771.656 2.16.77.98.295 1.543.552 2.563.552.179 0 .147-.262.147-.441v-.771c0-.33.695-1.198.951-1.469.453-.472.806-.674 1.428-.88.197-.065.302-.185.513-.185h.549c.357 0 .503.257.805.441.66.404 1.02.716 1.501 1.322.44.55.504 1.358 1.208 1.358.234 0 .357-.105.586-.147.352-.064.613.046.915-.146.33-.211.43-.588.824-.588.188 0 .275.216.275.404v.477c0 1.125-1.094 1.556-2.16 1.91-.77.257-1.19.486-1.977.66-.375.083-.879.088-.879.478v.514c.627.418 1.13.477 1.886.477 1.006-.082 1.578-.114 2.58-.22 1.277-.138 2.046-.33 3.149-.991.412-.248.879-.381.879-.863 0-.822-.39-1.35-1.062-1.818-.558-.39-1.501-.215-1.501-.9 0-.22.014-.353.037-.568.141-1.143.21-1.795.512-2.901.092-.33.055-.574.256-.845.188-.247.463-.188.77-.257.48-.11.75-.22 1.244-.22 1.016 0 1.583.317 2.49.771.75.376 1.464.753 1.464 1.597 0 .478-.193.73-.403 1.157-.499 1-1.025 1.56-1.025 2.68v.808c0 .711.512.991.842 1.616.224.426.445 1.173.476 1.23.348.628.659 1.423.769 1.634.137.257.444 1.556.586 2.092.27 1.024.545 1.589.658 2.644 0 0 .147 1.024.147 1.359v.863c0 .472-.019.865-.128 1.325l-.092.382c-.19.8-.924 4.13-1.025 4.736-.15.91-.356 1.404-.512 2.314-.318 1.85-.508 2.836-.952 4.626-.449 1.804-.59 2.84-.952 4.663a44.312 44.312 0 01-.513 2.313c-.141.583-.169.927-.329 1.506-.073.27-.192.55-.476.55-.078 0-.119-.036-.183-.073v-4.48c0-.904-.037-1.413-.11-2.313-.032-.399-.11-.624-.11-1.028v-6.609c0-.303.11-1.322.11-1.322h-.22c-.622.39-.553 1.038-.695 1.763-.256 1.294-.44 2.033-.55 3.341-.15 1.772-.086 2.79-.329 4.553-.251 1.836-.585 2.85-.952 4.663-.517-.239-.832-.606-.878-1.175-.096-1.157-.196-3.553-.22-3.929l-.11-1.689s-.08-1.262-.11-1.652c0 0-.082-1.088-.11-1.432l-.219-2.717v-2.02c0-.807.33-4.185.33-4.185.03-.379.146-1.684.146-1.763 0-.463.15-2.331.293-3.047.027-.142.11-1.065.11-1.065.028-.274-.339-.29-.586-.404-.504-.234-.838-.27-1.355-.477a9.155 9.155 0 00-1.244-.404c-.307-.078-.467-.257-.788-.257-.173 0-.2.23-.2.404 0 .124.255 2.152.438 3.047 0 0 .218 1.092.293 1.432 0 0 .293 1.143.293 1.322 0 .064.32 1.946.366 2.35.133 1.138.11 1.795.22 2.938.027.284.087.446.11.734l.11 1.395c.091 1.07.114 1.685.183 2.754-.453.179-.76.142-1.245.184-.71.064-1.121.04-1.83.11-.188.018-.376-.014-.477.147-.22.348-.096.651-.146 1.064a39.35 39.35 0 00-.22 2.13s-.105 1.11-.11 1.469c-.036 2.033-.004 3.18-.004 5.213-.417 0-.622-.21-1.02-.33-.522-.156-.86-.073-1.41-.073-.252 0-.229-.34-.275-.588-.073-.385-.732-3.02-.732-3.103 0-.091-.36-2.767-.366-2.845 0 0-.11-1.35-.11-1.432 0-.445-.014-.698.037-1.138.009-.088-.037-.881-.037-1.304 0-.294-.357-.307-.622-.422a20.35 20.35 0 01-1.83-.881c-.38-.202-.806-.335-.806-.771 0-.74.513-2.607.513-2.607l.659-3.048s.306-1.85.439-2.35c.128-.486.366-1.946.366-1.946.054-.289.293-3.47.293-3.561V21.31c0-.294-.37-.349-.66-.367-.622-.041-.983-.023-1.61-.037-.046.078-.073.726-.073.937 0 .092-.073.642-.073.936 0 .106-.11 1.248-.11 1.248-.076.86-.175 1.185-.257 2.167-.171 2.058.05 3.25-.256 5.287-.009.087-.114 1.414-.183 2.166-.096 1.042-.201 1.625-.33 2.662.019.042-.073.753-.073.753-.01.105-.146 1.359-.146 1.359-.058.54-.22 1.5-.22 1.597 0 .266-.732 4.686-.732 4.755 0 .463-.769 2.97-1.318 4.002h-.44c-.15-.29-.104-.514-.109-.844-.078-5.297.023-8.294-.11-13.585 0-.097-.068-.161-.146-.22-.028.605-.165.94-.22 1.541a46.595 46.595 0 00-.183 2.644c-.17 3.933-.183 6.164-.366 10.097a.924.924 0 01-.11.44c-.42-.224-.503-.6-.622-1.064-.224-.845-.412-1.318-.623-2.167-.315-1.266-.485-1.978-.695-3.268-.302-1.817-1.062-5.842-1.062-6.609v-8.15c0-1.933-.014-3.03.11-4.958l.073-1.028c-.055-.046-.128-.326-.146-.367-.838 0-1.304-.147-2.142-.147-.119 0-.165.138-.165.257.21 1.4.385 2.194.55 3.598.173 1.469.33 2.288.33 3.782v.661l.21 2.644v3.745c0 1.262-.092 1.978-.21 3.231-.079.799-.317 1.216-.55 1.983-.398 1.308-1.004 3.803-1.135 4.039-.405.73-1.043 3.107-1.172 3.892-.036.23-.022.55-.256.55-.613 0-.805-.743-.805-1.358.11-1.097.142-1.735.293-2.827.009-.087.1-5.191.036-8.372-.009-.537-.21-.872-.585-1.248-.133.082-.22.174-.22.33v1.322c0 .344-.11 1.359-.11 1.359-.111 1.375-.348 2.148-.44 3.525-.164 2.469-.136 3.873-.219 6.352-.041.009-.064.036-.11.036-1.345 0-1.345-1.735-1.647-3.047-.179-.785-.16-1.258-.293-2.056-.096-.583-.125-.921-.22-1.506 0 0-.242-1.822-.439-2.717-.398-1.804-.462-2.86-.879-4.663-.302-1.317-.491-2.072-.659-3.415 0 0-.068-.5-.073-.587-.082-1.441.01-2.268-.073-3.709 0-.046-.027-.67 0-1.028-.247-.197-.471-.174-.769-.294h-.22l-.279 10.207v5.288c0 .81-.077 1.257-.16 2.056-.155 1.45-.311 2.267-.549 3.708-.183 1.106-.188 1.822-.805 2.754-.188-.115-.128-.335-.183-.55-.124-.473-.151-.763-.183-1.249-.11-1.63-.101-2.556-.22-4.186-.133-1.84-.366-2.891-.989-4.626-.357-1-.402-1.69-1.098-2.497-.393.395-.242.85-.366 1.395-.215.937-.27 1.483-.476 2.424-.192.885-.146 1.418-.293 2.313-.096.596-.22 1.91-.22 1.91-.03.26-.11 1.1-.11 1.1a28.855 28.855 0 01-.438 2.975c-.179.927-.142 1.638-.843 2.276-.082-1.629-.087-2.56-.22-4.186-.16-1.95-.356-3.047-.548-4.993-.142-1.441-.225-2.263-.33-3.709-.064-.9-.07-1.41-.11-2.313 0 0-.08-2.13-.11-2.46l-.11-1.248s-.075-1.021-.11-1.322c0 0-.17-1.347-.219-1.91 0 0-.131-1.868-.22-2.57 0 0-.219-1.202-.219-1.744 0-.151-.147-3.63-.147-3.726 0-.63.197-.983.476-1.543.151-.307 1.062-1.858 1.575-2.46.393-.463.878-.693.878-1.303 0-.11-.073-.606-.073-.606-.142-1.171-1.172-1.611-1.172-2.79v-.441c0-.473.344-.68.696-.991.837-.74 1.446-1.286 2.563-1.286 1.016 0 2.196.376 2.196 1.396v3.01c0 .35.11 1.286-.073 1.396-.252.151-.457.11-.732.22-.53.211-.879.601-.879 1.175 0 .372.165.574.366.881.504.758.897 1.12 1.538 1.763.444.445.764.66 1.318.954.49.262.76.551 1.318.551h.366c.307 0 .302-.385.44-.66.137-.276.402-.386.402-.698 0-.368-.512-.303-.879-.33-.631-.051-1.464-.065-1.464-.698 0-.193.146-.285.146-.478.353-.032.563-.004.916-.037.85-.082 1.313-.266 2.16-.403.92-.152 1.867-.57 1.867-1.506 0-.546-.33-.794-.623-1.248-.196-.308-.438-.43-.512-.79-.052-.252-.008-.406.036-.66.14-.805.503-1.225 1.099-1.781.436-.408.726-.644 1.3-.808.453-.13.874-.074 1.208-.074.448 0 .709-.013 1.153.037.833.101 1.441.707 1.5 1.542.037.542-.027.854.037 1.395.005.051-.073.446-.073.661v.478c0 .6-.444 1.138-1.043 1.138-.234 0-.371-.073-.604-.073-.151 0-.257.142-.257.293 0 .198.06.441.257.441.782 0 1.194-.367 1.977-.367h.622c.265 0 .403-.115.659-.184.718-.197 1.149-.294 1.83-.587.394-.17.6-.308.952-.551.188-.129.366-.211.403-.44.114-.67.11-1.066.11-1.745 0-.555-.55-.679-.879-1.12-.49-.652-.659-1.18-.659-2 0-.556 0-.905.22-1.414.128-.299.865-1.3 1.281-1.506.444-.22.764-.22 1.263-.22"
+						fill="#fff"
+					/>
+				</g>
+			);
+		case LOGO_KEYS.RAPHAELSWERK:
+			return (
+				<>
+					<g clip-path="url(#clip0)">
+						<path
+							d="M385.4 380.3H390.9V374.5H385.4V380.3ZM386 344.2H380.8L370 375.7H369.9L359.3 344.2H353.8L366.7 380.3H372.8L386 344.2ZM348.4 380.3H353.9V374.5H348.4V380.3ZM324.7 364.5C325 361.1 326.9 357.2 331.4 357.2C335.4 357.2 337.6 360.5 337.6 364.5H324.7ZM340.7 374.3C339.6 375.3 336 376.9 333 376.9C328.6 376.9 324.8 373.3 324.8 368.1H342.7C342.7 359.6 340 353.2 331.7 353.2C324.6 353.2 320 358.7 320 366.5C320 375.1 324.1 380.8 333 380.8C337.1 380.8 339.1 379.8 340.9 379.1V374.3H340.7ZM286.4 380.3H291V366.8L302.2 380.3H308.8L295.7 365.6L307.3 353.9H301L291 364.6V341.5H286.4V380.3ZM267.8 380.3H272.4V368.3C272.4 361.7 275.1 357.5 279.2 357.5C280 357.5 280.9 357.6 281.8 358V353.7C280.8 353.5 280.1 353.3 278.8 353.3C276.1 353.3 273.5 355.3 272.3 358H272.2V353.9H267.9V380.3H267.8ZM244.2 364.5C244.5 361.1 246.4 357.2 250.9 357.2C254.9 357.2 257.1 360.5 257.1 364.5H244.2ZM260.2 374.3C259.1 375.3 255.5 376.9 252.5 376.9C248.1 376.9 244.3 373.3 244.3 368.1H262.2C262.2 359.6 259.5 353.2 251.2 353.2C244.1 353.2 239.5 358.7 239.5 366.5C239.5 375.1 243.6 380.8 252.5 380.8C256.6 380.8 258.6 379.8 260.4 379.1V374.3H260.2ZM237 353.9H232.5L226.4 375.7H226.3L220 353.9H214.5L208.1 375.7H208L202.3 353.9H197.2L205.3 380.3H210.7L217 358.2H217.1L223.6 380.3H229L237 353.9ZM179.9 379.4C182 380.6 184.4 380.8 186.8 380.8C191.3 380.8 195.5 378.2 195.5 372.6C195.5 364.2 184.6 365.7 184.6 360.4C184.6 358.2 186.7 357.2 189 357.2C190.1 357.2 192.8 357.9 193.9 358.5L194.3 354.3C192.5 353.7 190.7 353.3 188.6 353.3C183.6 353.3 179.8 355.9 179.8 361.4C179.8 368.8 190.7 368.4 190.7 373.2C190.7 376 188.1 377 186 377C184.7 377 182.1 376.4 180.3 375.1L179.9 379.4ZM169.4 380.3H174V341.5H169.4V380.3ZM145.2 364.5C145.5 361.1 147.4 357.2 151.9 357.2C155.9 357.2 158.1 360.5 158.1 364.5H145.2ZM161.2 374.3C160.1 375.3 156.5 376.9 153.5 376.9C149.1 376.9 145.3 373.3 145.3 368.1H163.2C163.2 359.6 160.5 353.2 152.2 353.2C145.1 353.2 140.5 358.7 140.5 366.5C140.5 375.1 144.6 380.8 153.5 380.8C157.6 380.8 159.6 379.8 161.4 379.1V374.3H161.2ZM130.1 370C130.1 373.9 127.7 376.9 123.1 376.9C121 376.9 118.4 375.4 118.4 372.6C118.4 368 124.6 367.5 127.1 367.5C128.1 367.5 129.1 367.6 130 367.6V370H130.1ZM116.6 359.9C118.5 358.2 121.3 357.2 123.9 357.2C128.5 357.2 130.1 359.5 130.1 364.1C128.3 364 127 364 125.2 364C120.4 364 113.6 366.1 113.6 372.7C113.6 378.4 117.4 380.9 122.8 380.9C127 380.9 129.5 378.5 130.6 376.9H130.7V380.3H135C134.9 379.5 134.8 378.1 134.8 375.1V364.2C134.8 357 131.9 353.3 124.7 353.3C121.5 353.3 118.8 354.3 116.5 355.8L116.6 359.9ZM85.7 380.3H90.3V368.1C90.3 361.7 92.7 357.2 97.9 357.2C101.9 357.5 103.2 360.5 103.2 366.6V380.2H107.8V364.5C107.8 357.4 104.9 353.2 98.4 353.2C94.9 353.2 92 354.8 90.3 357.6H90.2V341.4H85.6V380.3H85.7ZM75.1 367C75.2 371.1 73.1 377 68.2 377C63 377 60.8 371.6 60.8 367.2C60.8 362.3 63.2 357.3 68.2 357.3C73.2 357.2 75.2 362.2 75.1 367ZM56.2 391.1H60.8V376.6H60.9C62.3 379 64.8 380.8 69.2 380.8C76.8 380.8 80 373.8 80 366.5C80 359.4 76.5 353.2 69.2 353.2C64.8 353.2 61.9 355.6 60.7 357.5H60.6V353.8H56.2V391.1ZM44.2 370C44.2 373.9 41.8 376.9 37.2 376.9C35.1 376.9 32.5 375.4 32.5 372.6C32.5 368 38.7 367.5 41.2 367.5C42.2 367.5 43.2 367.6 44.1 367.6V370H44.2ZM30.8 359.9C32.7 358.2 35.5 357.2 38.1 357.2C42.7 357.2 44.3 359.5 44.3 364.1C42.5 364 41.2 364 39.4 364C34.6 364 27.8 366.1 27.8 372.7C27.8 378.4 31.6 380.9 37 380.9C41.2 380.9 43.7 378.5 44.8 376.9H44.9V380.3H49.2C49.1 379.5 49 378.1 49 375.1V364.2C49 357 46.1 353.3 38.9 353.3C35.7 353.3 33 354.3 30.7 355.8L30.8 359.9ZM4.9 348.7H9.6C14.1 348.7 16.5 350.8 16.5 354C16.5 357.5 13.8 359.6 9.1 359.6H4.9V348.7ZM0 380.3H4.9V364.3H7.8C10.4 364.3 12 364.3 13.5 367.8L18.8 380.3H24.4L18 365.9C16.9 363.7 16.2 362.9 14.7 362.3V362.2C18.7 361.5 21.8 358.4 21.8 354C21.8 345.4 15.3 344.2 8.5 344.2H0V380.3Z"
+							fill="white"
+						/>
+						<g opacity="0.6">
+							<path
+								d="M218.2 292.6V305.9H193.2V262.4H186C172.4 262.4 168.5 283 157.7 305.9H130.1C144.2 274.5 147.3 261.3 157.7 255.8C146.2 250.1 140.5 241.3 140.5 229.6C140.5 213.6 152.7 198.8 173.4 198.8H195.3L218.2 292.6ZM193.2 218.3H180.5C171.1 218.3 166.2 222.4 166.2 230.4C166.2 239.5 173 242.7 179.9 242.7H193.2V218.3Z"
+								fill="white"
+							/>
+							<path
+								d="M196.3 205.4C196.3 205.4 191.3 198.3 201.4 189.5C215.9 176.9 206.4 154.7 241 115.9C270.4 82.9 252.3 45.2 254.6 47C276.9 64.2 268.3 105 268.3 105C268.3 105 274.3 83.3 275.1 75C275.8 67.6 283.2 84.6 279.6 115.2C278.1 127.6 273.6 140.9 273.6 140.9C273.6 140.9 283.8 122.1 285 117.5C285.5 115.5 287.8 137.9 280.3 157.7C277.4 165.5 274.9 171.7 274.9 171.7C274.9 171.7 279.9 164.9 284 158.3C289.3 149.7 286.5 170.3 283.7 177.6C279.6 188 275.4 194 275.4 194C275.4 194 277.5 195 281.7 191.5C283.7 189.9 283.9 194.8 281.6 201.7C280.1 206.2 273.4 213.3 273.4 213.3C273.4 213.3 279.5 212.2 282.3 209.4C283.3 208.4 283.5 212.4 280.9 218.4C277.3 226.9 265.5 237.3 265.5 237.3C265.5 237.3 270.5 237.8 275.7 233.6C277.8 231.9 275 240.2 271.4 244.2C266.6 249.5 260.7 252.1 260.7 252.1C260.7 252.1 271.1 252.4 262.3 262.4C256 269.5 247.3 269.8 247.3 269.8C247.3 269.8 245.6 283.2 232.6 280.6C232.6 280.6 225.3 295.3 212.3 291.9C212.3 291.9 209.8 291.4 209.9 289.5C210.2 286.1 215.1 282.6 215.1 276.6C215.1 272.6 212.9 269.2 212.9 269.2"
+								fill="white"
+							/>
+							<path
+								d="M218.2 292.6V305.9H193.2V262.4H186C172.4 262.4 168.5 283 157.7 305.9H130.1C144.2 274.5 147.3 261.3 157.7 255.8C146.2 250.1 140.5 241.3 140.5 229.6C140.5 213.6 152.7 198.8 173.4 198.8H195.3L218.2 292.6ZM193.2 218.3H180.5C171.1 218.3 166.2 222.4 166.2 230.4C166.2 239.5 173 242.7 179.9 242.7H193.2V218.3ZM196.3 205.4C196.3 205.4 191.3 198.3 201.4 189.5C216 176.8 206.4 154.7 241 115.9C270.4 82.9 252.3 45.2 254.6 47C276.9 64.2 268.3 105 268.3 105C268.3 105 274.3 83.3 275.1 75C275.8 67.6 283.2 84.6 279.6 115.2C278.1 127.6 276.7 133.7 273.4 141.4C273.4 141.4 283.7 122.2 284.9 117.6C285.4 115.6 287.7 138 280.2 157.8C277.8 164.2 273.9 173.2 273.9 173.2C273.9 173.2 279.8 165.9 283.9 158.4C288.7 149.5 286.4 170.4 283.6 177.7C279.5 188.1 275.3 194.1 275.3 194.1C275.3 194.1 277.4 195.1 281.6 191.6C283.6 190 283.8 194.9 281.5 201.8C280 206.3 273.3 213.4 273.3 213.4C273.3 213.4 279.4 212.3 282.2 209.5C283.2 208.5 283.4 212.5 280.8 218.5C277.2 227 265.4 237.4 265.4 237.4C265.4 237.4 270.4 237.9 275.6 233.7C277.7 232 274.9 240.3 271.3 244.3C266.5 249.6 260.6 252.2 260.6 252.2C260.6 252.2 271 252.5 262.2 262.5C255.9 269.6 247.2 269.9 247.2 269.9C247.2 269.9 245.5 283.3 232.5 280.7C232.5 280.7 225.2 295.4 212.2 292C212.2 292 209.7 291.5 209.8 289.6C210.1 286.2 215 282.7 215 276.7C215 272.7 212.1 269.5 212.1 269.5L196.3 205.4ZM260.3 127.7C260.3 127.7 265.7 117.6 268.5 103.7L260.3 127.7ZM262.2 160.5C270.2 149 274.3 139.6 274.3 139.6L262.2 160.5ZM264.4 186.2C270.6 178.4 273.9 173.3 273.9 173.3L264.4 186.2Z"
+								fill="white"
+							/>
+						</g>
+						<path
+							d="M217.695 306.9H192.765C192.166 306.9 191.767 306.5 191.767 305.9V263.413H185.584C176.011 263.413 171.523 274.609 165.241 290.205C163.246 295.203 160.953 300.802 158.36 306.4C158.161 306.8 157.861 307 157.462 307H129.939C129.64 307 129.241 306.8 129.141 306.5C129.041 306.2 128.942 305.9 129.041 305.5C132.432 298.003 135.124 291.604 137.518 285.906C144.498 269.311 148.288 260.513 155.368 255.815C144.698 250.016 139.213 241.119 139.213 229.522C139.213 221.125 142.404 213.227 148.188 207.429C154.471 201.13 163.047 197.731 173.019 197.731H194.161C194.859 195.132 196.654 192.133 200.343 188.834C206.028 183.935 207.823 177.537 210.416 168.74C214.105 156.043 219.091 138.548 239.734 115.355C255.49 97.6602 260.078 75.6665 253.396 49.9739C252.698 47.3746 252.598 46.6748 253.297 46.1749C253.596 45.975 254.094 45.875 254.593 46.2749C263.867 53.4729 269.252 65.9693 270.15 82.4645C270.349 85.6636 270.25 88.6627 270.15 91.4619C271.646 85.4637 273.042 78.7656 273.441 75.0666C273.54 74.0669 273.74 72.9673 274.637 72.5674C275.136 72.3674 275.635 72.4674 276.133 72.7673C278.925 74.7667 282.715 91.7619 279.923 115.455C279.225 121.553 278.527 126.152 277.629 130.151C280.222 125.052 282.815 119.654 283.313 117.454C283.513 116.855 283.912 116.655 284.111 116.555C284.51 116.455 284.909 116.655 285.108 116.855C286.305 118.254 287.402 139.948 280.521 158.143C279.624 160.542 278.427 163.441 277.43 166.041C279.125 163.641 280.92 160.742 282.416 157.943C282.914 156.943 284.211 154.644 285.707 155.344C288.798 156.743 285.607 173.638 283.812 178.137C280.92 185.535 278.028 190.733 276.432 193.133C277.33 192.933 278.527 192.333 280.222 190.933C281.319 190.034 282.117 190.434 282.416 190.633C284.311 191.933 283.214 197.631 281.718 202.13C280.72 205.129 277.629 209.128 275.535 211.627C277.33 211.028 279.424 210.128 280.82 208.828C281.419 208.228 282.117 208.328 282.516 208.628C284.211 209.728 282.516 215.426 281.02 218.925C278.227 225.324 271.147 232.721 267.557 236.12C269.352 235.721 271.845 234.921 274.338 232.921C275.435 232.022 276.133 232.721 276.333 232.821C277.829 234.421 273.939 242.119 271.446 244.918C268.455 248.217 265.164 250.416 262.97 251.716C264.166 252.116 265.263 252.916 265.862 254.215C266.859 256.415 265.662 259.314 262.371 263.013C256.986 269.111 249.906 270.411 247.513 270.61C247.114 272.41 245.817 277.009 242.128 279.708C239.535 281.607 236.344 282.207 232.554 281.607C231.258 283.907 226.571 291.205 218.892 292.904V305.6C218.693 306.5 218.294 306.9 217.695 306.9ZM193.762 304.901H216.698V293.604C215.003 293.804 213.208 293.604 211.313 293.104C210.216 292.904 208.022 291.904 208.222 289.605C208.421 288.005 209.418 286.406 210.416 284.806C211.812 282.607 213.407 280.108 213.407 276.809C213.407 273.21 210.814 270.311 210.814 270.211C210.416 269.811 210.515 269.211 210.914 268.811C211.313 268.411 211.911 268.511 212.31 268.911C212.41 269.011 215.402 272.41 215.402 276.809C215.402 280.708 213.507 283.507 212.011 285.906C211.014 287.406 210.216 288.705 210.116 289.805C210.017 290.805 211.612 291.105 211.612 291.105C223.778 294.204 230.759 280.408 230.759 280.308C230.959 279.908 231.457 279.708 231.856 279.808C235.346 280.508 238.338 280.008 240.632 278.408C244.621 275.509 245.418 269.911 245.418 269.911C245.518 269.411 245.917 269.011 246.416 269.011C246.515 269.011 254.792 268.611 260.676 262.013C263.269 259.014 264.366 256.714 263.768 255.315C263.169 253.915 260.676 253.515 259.779 253.415C259.28 253.415 258.881 253.016 258.881 252.616C258.781 252.116 259.081 251.716 259.479 251.516C259.579 251.516 265.263 248.917 269.851 243.918C272.045 241.519 273.74 237.72 274.338 235.621C269.252 239.12 264.765 238.72 264.565 238.62C264.166 238.62 263.867 238.32 263.668 237.92C263.568 237.52 263.668 237.12 263.967 236.82C264.067 236.72 275.635 226.523 279.125 218.326C280.521 215.126 281.02 212.727 281.119 211.328C277.928 213.727 273.042 214.627 272.743 214.627C272.344 214.727 271.845 214.527 271.646 214.127C271.446 213.727 271.546 213.327 271.745 212.927C271.845 212.827 278.227 205.929 279.723 201.63C281.319 196.932 281.518 193.833 281.219 192.733C276.732 196.332 274.338 195.332 274.039 195.232C273.74 195.132 273.54 194.832 273.54 194.632C273.441 194.332 273.54 194.032 273.64 193.833C273.64 193.733 277.829 187.734 281.817 177.637C284.011 171.939 285.308 160.742 284.709 157.843C284.51 158.143 284.311 158.443 283.912 159.143C280.022 166.34 274.438 173.438 273.939 174.038C273.64 174.538 270.25 179.537 264.466 186.935C264.166 187.334 263.468 187.434 263.069 187.134C262.671 186.835 262.571 186.135 262.87 185.735C268.554 178.537 271.845 173.638 272.244 173.039C272.643 172.139 276.133 163.641 278.427 157.743C283.712 143.747 284.111 128.451 283.712 121.753C280.621 128.851 273.84 141.548 273.54 142.247C271.845 145.646 268.056 152.944 262.272 161.242C261.973 161.742 261.374 161.842 260.876 161.542C260.377 161.242 260.277 160.642 260.576 160.142C266.261 151.845 270.05 144.647 271.646 141.248C274.837 133.75 276.233 127.951 277.729 115.355C280.322 93.0615 277.13 78.5656 275.236 75.1666V75.3666C274.438 83.6642 268.654 104.658 268.455 105.558C268.455 105.658 268.455 105.658 268.355 105.758C265.463 118.854 260.477 128.251 260.477 128.351C260.177 128.851 259.579 129.051 259.081 128.751C258.582 128.451 258.382 127.851 258.682 127.352C258.781 127.252 263.967 117.255 266.759 103.658C267.258 100.559 268.455 92.1617 267.956 82.6645C267.058 64.7696 260.876 54.8725 255.092 49.6739C256.288 54.3726 258.781 64.1698 258.083 76.2663C257.086 91.4619 251.402 105.058 241.031 116.755C220.687 139.648 215.701 156.843 212.111 169.34C209.618 178.137 207.623 185.035 201.44 190.334C192.166 198.431 196.355 204.629 196.554 204.829C196.853 205.329 196.753 205.929 196.355 206.229C195.956 206.529 195.258 206.429 194.958 206.029C194.958 206.029 193.263 203.53 193.562 199.731H172.82C151.978 199.731 141.008 214.727 141.008 229.522C141.008 240.919 146.593 249.417 157.662 254.915C157.961 255.115 158.26 255.415 158.26 255.815C158.26 256.215 158.061 256.515 157.762 256.714C150.282 260.713 146.593 269.411 139.213 286.706C136.919 292.004 134.426 298.003 131.335 304.901H156.665C159.158 299.502 161.252 294.404 163.246 289.405C169.629 273.81 174.615 261.413 185.485 261.413H192.665C193.263 261.413 193.662 261.813 193.662 262.413V304.901H193.762ZM192.765 243.818H179.501C170.327 243.818 164.842 238.82 164.842 230.522C164.842 224.524 167.535 217.426 180.1 217.426H192.765C193.363 217.426 193.762 217.826 193.762 218.425V242.818C193.762 243.318 193.263 243.818 192.765 243.818ZM180.1 219.425C171.224 219.425 166.836 223.024 166.836 230.522C166.836 240.719 175.712 241.819 179.501 241.819H191.767V219.425H180.1Z"
+							fill="white"
+						/>
+						<path
+							d="M235 232.6C234.8 232.6 234.5 232.5 234.3 232.3C233.9 231.9 233.9 231.3 234.3 230.9C242.6 222 246.3 208.8 246.1 205.5C241.9 208.2 238 208.6 237.9 208.6C237.5 208.6 237.1 208.4 237 208.1C236.8 207.8 236.8 207.4 237 207C241.2 200.6 242.8 191.2 242.9 187.8C239.6 190.1 235.6 192.3 235.4 192.4C235 192.6 234.5 192.6 234.2 192.2C233.9 191.9 233.8 191.4 234.1 191C235.5 188.6 238.5 182.3 236.9 179.2C235.2 181.1 231.3 185.1 227.8 185.3C227.3 185.3 226.8 184.9 226.7 184.4C226.7 183.8 227.1 183.4 227.6 183.3C230.7 183.1 234.8 178.6 236.1 176.9C236.3 176.7 236.5 176.5 236.8 176.5C237.1 176.5 237.4 176.6 237.6 176.8C240.7 179.7 239 185.3 237.4 188.9C239.1 187.9 241.2 186.6 242.9 185.3C243.4 185 243.9 185 244.3 185.4C245.7 186.8 244 198.3 239.8 206.1C241.3 205.7 243.4 204.9 245.5 203.4C246.4 202.8 247.2 203.1 247.6 203.6C249.7 206.1 244.8 222.6 235.8 232.2C235.5 232.5 235.3 232.6 235 232.6Z"
+							fill="white"
+						/>
+					</g>
+					<defs>
+						<clipPath id="clip0">
+							<rect width="391" height="391" fill="white" />
+						</clipPath>
+					</defs>
+				</>
+			);
+		case LOGO_KEYS.SKF:
+			return (
+				<>
+					<defs>
+						<path id="a-1" d="M.222.237H31.91v12.948H.222z" />
+					</defs>
+					<g fill="none" fillRule="evenodd">
+						<path d="M-9 0h60v60H-9z" />
+						<g transform="translate(5.946 46.8)">
+							<mask id="b-1" fill="#fff">
+								<use xlinkHref="#a" />
+							</mask>
+							<path
+								d="M23.117 12.973V.464h8.792v2.114h-6.2v2.97h5.36v2.114h-5.36v5.31h-2.592zm-10.312 0V.464h2.448v6.64l2.877-3.198h3.02l-3.18 3.32 3.412 5.747h-2.644l-2.36-4.088-1.125 1.188v2.9h-2.448zM.222 8.92l2.52-.245c.143.821.464 1.433.911 1.817.465.384 1.09.577 1.877.577.822 0 1.447-.175 1.876-.524.411-.333.626-.734.626-1.206 0-.28-.09-.542-.268-.751-.18-.21-.483-.385-.93-.542-.286-.105-.983-.28-2.055-.541-1.376-.332-2.341-.752-2.895-1.24-.786-.7-1.18-1.538-1.18-2.534 0-.629.18-1.223.555-1.782.375-.559.893-.978 1.59-1.275C3.546.377 4.386.237 5.351.237c1.608 0 2.806.35 3.628 1.048.804.682 1.233 1.625 1.269 2.778l-2.592.087c-.107-.646-.34-1.117-.697-1.397-.375-.28-.911-.42-1.626-.42-.75 0-1.34.158-1.751.455a.924.924 0 00-.411.786c0 .297.125.541.375.75.34.28 1.126.56 2.377.84 1.25.297 2.18.593 2.788.908a3.78 3.78 0 011.411 1.275c.34.525.5 1.188.5 1.992 0 .716-.196 1.38-.607 2.01-.41.628-.983 1.082-1.733 1.397-.75.296-1.662.454-2.788.454-1.608 0-2.86-.367-3.717-1.1C.9 11.382.383 10.316.222 8.92z"
+								fill="#FFF"
+								mask="url(#b)"
+							/>
+						</g>
+						<path
+							d="M17.716 30.691c-.461.221-.99.542-1.537 1.151-.49.545-.712 1.017-1.033 1.643a15.53 15.53 0 01-2.024-1.17c.469-.946 1.13-1.926 1.973-2.745.76-.74 1.682-1.357 2.62-1.665v2.786zm6.664 6.636l.368-.06c7.078-1.14 12.795-6.261 14.654-13.276H44C41.847 34.28 32.812 42 22 42 11.18 42 2.142 34.268 0 23.967h4.608c1.852 6.985 7.648 12.09 14.64 13.278l.371.063V21.667H5.791l-3.333-4.815H19.62V0h4.761v16.852h17.163l-3.332 4.815H24.38v15.66zm1.945-9.16c.529.956 1.255 1.728 2.158 2.498.948.81 1.804 1.357 2.458 1.673-.69.5-1.367.89-2.133 1.272-.472-.256-1.007-.659-1.421-1.012a8.101 8.101 0 01-1.06-1.111l-.002-3.32zm7.645 1.375a16.29 16.29 0 01-1.614 1.678c-.746-.37-2.141-1.243-3.524-2.642-1.369-1.385-2.475-3.108-2.507-4.99h2.437c.237 1.361.994 2.611 1.985 3.647 1 1.045 2.321 1.981 3.223 2.307zm-.623-5.954c.219.466.536 1 1.138 1.554.538.495 1.005.72 1.624 1.045a15.822 15.822 0 01-1.156 2.048 10.065 10.065 0 01-2.714-1.997c-.731-.769-1.341-1.701-1.646-2.65h2.754zM30.52 2.804c.4-.418.869-.704 1.644-.897a18.11 18.11 0 011.894 1.331c-.207.074-1.47.23-2.85 2.402-.74 1.164-1.183 1.972-1.919 2.935-.762.998-1.77 1.838-2.998 2.062V8.092c.972-.227 1.45-.966 1.944-1.865.608-1.105 1.43-2.533 2.285-3.423zm7.596 4.623c.323.432.798 1.15 1.172 1.8a3.42 3.42 0 00-1.127.64c-1.344 1.133-1.863 2.407-2.775 3.643-.342.463-.978 1.103-1.433 1.376h-3.203c1.194-.389 1.986-1.419 2.935-2.701.939-1.268 1.89-2.854 2.636-3.554.597-.559.946-.863 1.795-1.204zm-2.69-3.061a21.792 21.792 0 011.56 1.63c-.7.199-1.801 1.075-2.284 1.692-1.496 1.91-3.165 4.483-4.097 5.311-.924.821-2.307 1.735-4.314 1.887V12.41c3.947-.481 4.899-4.07 6.96-6.502.592-.7 1.607-1.514 2.174-1.542zM5.771 10.607c-.413-.405-.696-.879-.886-1.663a18.33 18.33 0 011.316-1.916c.073.21.227 1.486 2.374 2.884 1.15.748 1.95 1.196 2.902 1.94.986.771 1.816 1.79 2.038 3.034L11 14.885c-.225-.983-.955-1.466-1.844-1.967-1.092-.614-2.504-1.447-3.384-2.31zm4.57-7.684a19.772 19.772 0 011.78-1.186c.057.185.236.658.633 1.14 1.12 1.36 2.379 1.885 3.6 2.808.458.346 1.091.988 1.36 1.45v3.24c-.384-1.209-1.402-2.01-2.67-2.97-1.253-.95-2.82-1.911-3.512-2.667-.552-.603-.854-.956-1.19-1.815zM7.318 5.645c.493-.534 1.065-1.101 1.612-1.578.196.707 1.062 1.82 1.672 2.31 1.887 1.512 4.431 3.201 5.25 4.144.812.935 1.715 2.334 1.865 4.365h-2.448c-.475-3.993-4.023-4.957-6.427-7.041-.691-.6-1.496-1.626-1.524-2.2zm5.872 17.943c-.945.534-1.709 1.27-2.47 2.182-.8.96-1.341 1.826-1.654 2.486A14.968 14.968 0 017.808 26.1c.254-.478.652-1.019 1-1.438a8.1 8.1 0 011.099-1.072l3.282-.001zm-1.36 7.733a16.276 16.276 0 01-1.658-1.633c.365-.754 1.229-2.166 2.611-3.565 1.37-1.384 3.073-2.503 4.934-2.535v2.465c-1.347.24-2.582 1.005-3.606 2.008-1.033 1.011-1.959 2.348-2.28 3.26z"
+							fill="#FFF"
+						/>
+					</g>
+				</>
+			);
+		case LOGO_KEYS.SKM:
+			return (
+				<g fill="none" fillRule="evenodd">
+					<path d="M0-7h60v60H0z" />
+					<path
+						d="M60 43.473C60 44.869 58.887 46 57.514 46H2.487C1.113 46 0 44.869 0 43.473V2.527C0 1.132 1.113 0 2.487 0h55.027C58.887 0 60 1.132 60 2.527v40.946"
+						fill="#fff"
+					/>
+					<path
+						d="M25.909 11.835s1.131-1.268 2.332-3.372c.91-1.595 3.122-3.05 3.194-3.051.257-.008 2.5 1.627 2.655 1.795.05.055-1.357.751-2.398 1.864-.784.838-1.255 1.545-2.62 3.749-.545.879-1.224 1.726-1.643 2.125-.418.399-1.52 1.29-1.52 1.29v-4.4"
+						fill="#CC1E1C"
+					/>
+					<path
+						d="M25.909 18.941v-2.305s1.624-1.16 2.312-1.904c.694-.752 1.51-2.005 2.024-2.856.514-.852 1.423-2.22 1.889-2.823.71-.919 2.447-2.288 2.569-2.288.211 0 2.054 2.088 2.115 2.254.024.066-1.496 1.12-2.055 1.82-.523.657-1.453 2.054-1.662 2.44-.363.668-1.073 1.87-1.48 2.388-.373.473-1.21 1.387-1.769 1.854-.367.307-1.662 1.42-1.662 1.42H25.91M30 18.941s1.473-1.144 2.039-1.67c.741-.689 1.451-1.715 2.023-2.619.512-.81 1.158-1.777 1.668-2.297.541-.552 2.008-1.516 2.162-1.532.155-.015 1.635 2.33 1.653 2.42.015.077-.958.797-1.622 1.64-.561.711-1.328 1.792-1.698 2.297-.372.505-.523.63-.834.965-.387.413-.897.78-.897.78L30 18.942M9.756 18.941s-1.119-1.123-2.975-2.313c-1.406-.903-2.688-3.098-2.69-3.17-.007-.255 1.435-2.481 1.583-2.633.049-.05.663 1.345 1.644 2.378.74.778 1.363 1.245 3.306 2.6.776.54 1.522 1.214 1.875 1.63.351.415 1.137 1.508 1.137 1.508h-3.88M17.727 18.941h-2.065s-1.04-1.813-1.706-2.58c-.673-.776-1.796-1.687-2.559-2.26-.763-.574-1.99-1.589-2.529-2.108-.823-.793-2.05-2.732-2.05-2.867 0-.236 1.87-2.294 2.02-2.361.059-.027 1.003 1.67 1.631 2.294.588.583 1.84 1.62 2.185 1.854.599.405 1.676 1.198 2.14 1.653.423.416 1.242 1.35 1.66 1.973.276.41 1.273 1.855 1.273 1.855v2.547"
+						fill="#CC1E1C"
+					/>
+					<path
+						d="M17.727 14.882s-1.153-1.461-1.683-2.023c-.694-.735-1.729-1.44-2.64-2.007-.816-.509-1.79-1.15-2.315-1.655-.556-.537-1.528-1.993-1.544-2.146-.015-.153 2.347-1.622 2.44-1.64.076-.015.802.95 1.651 1.61.718.556 1.806 1.318 2.315 1.685.51.368.637.518.973.828.417.383.788.889.788.889l.015 4.459M2.727 27.059h39.546v-6.765H2.727z"
+						fill="#CC1E1C"
+					/>
+					<path
+						fill="#CC1E1C"
+						d="M19.091 43.294h6.818V4.059h-6.818z"
+					/>
+					<path
+						d="M36.852 42.623L31.9 35.424l4.952-7.23v14.43zm12.93-15.564l-2.128 7.762-2.143-7.762H29.237l-4.691 8.928 4.69 7.307h14.232V35.15l2.275 8.144h3.822l2.423-8.144v8.144h6.65V27.06h-8.855zM18.626 27.727s-9.081-1.726-12.282.37c-2.351 1.54-2.253 3.804-2.253 5.16 0 1.913.99 3.69 3.1 4.078 3.88.714 5.308-.119 5.308.924 0 .93-.637.791-2.927.655-2.052-.122-5.118-.686-5.118-.686l-.005 4.28s5.04 1.133 9.253.679c3.081-.332 5.389-2.075 5.389-5.189 0-2.137-.433-4.601-3.693-4.712-3.96-.134-4.459.062-4.459-.816 0-.76.623-.648 3.68-.579 2.071.047 4.007.394 4.007.394v-4.558"
+						fill="#CC1E1C"
+					/>
+				</g>
+			);
+		default:
+			return;
+	}
+};
